### 2021-05-13 #3277 @whedon
[PRE REVIEW]: udocker: a user oriented tool for unprivileged Linux containers
**Submitting author:** @jorge-lip (<a href="http://orcid.org/0000-0002-9142-2596">Jorge Gomes</a>)
**Repository:** <a href="https://github.com/indigo-dc/udocker" target ="_blank">https://github.com/indigo-dc/udocker</a>
**Version:** v1.1.7
