### 2021-05-13 #3277 @danielskatz
👋 @jorge-lip - The first two ORCIDs in your paper.md are too short. They are likely missing the initial set of four digits.

In addition, you could work on the possibly missing DOIs that whedon suggests, but note that some may be incorrect. Please feel free to make changes to your .bib file.

Then use the command `@whedon check references` to check the references again, and the command `@whedon generate pdf` to make a new PDF. Whedon commands need to be the first entry in a new comment.
