+++ paper/paper.bib
@@ -1,26 +1,3 @@
-@article{GOMES2017,
-  author    = {Jorge Gomes and
-               Isabel Campos Plasencia and
-               Emanuele Bagnaschi and
-               M{\'{a}}rio David and
-               Lu{\'{\i}}s Alves and
-               Jo{\~{a}}o Martins and
-               Jo{\~{a}}o Murta Pina and
-               {\'{A}}lvaro L{\'{o}}pez Garc{\'{\i}}a and
-               Pablo Orviz Fern{\'{a}}ndez},
-  title     = {Enabling rootless Linux Containers in multi-user environments: the
-               udocker tool},
-  journal   = {CoRR},
-  volume    = {abs/1711.01758},
-  year      = {2017},
-  url       = {http://arxiv.org/abs/1711.01758},
-  archivePrefix = {arXiv},
-  eprint    = {1711.01758},
-  timestamp = {Wed, 11 Dec 2019 15:12:01 +0100},
-  biburl    = {https://dblp.org/rec/journals/corr/abs-1711-01758.bib},
-  bibsource = {dblp computer science bibliography, https://dblp.org}
-}
-
 @article{GOMES2018,
 title = {Enabling rootless Linux Containers in multi-user environments: The udocker tool},
 journal = {Computer Physics Communications},
@@ -28,7 +5,7 @@ volume = {232},
 pages = {84-97},
 year = {2018},
 issn = {0010-4655},
-doi = {https://doi.org/10.1016/j.cpc.2018.05.021},
+doi = {10.1016/j.cpc.2018.05.021},
 url = {https://www.sciencedirect.com/science/article/pii/S0010465518302042},
 author = {Jorge Gomes and Emanuele Bagnaschi and Isabel Campos and Mario David and Luís Alves and João Martins and João Pina and Alvaro López-García and Pablo Orviz},
 keywords = {Linux containers, HPC on cloud, Virtualization, Phenomenology, QCD, Biophysics},
@@ -58,7 +35,7 @@ abstract = {Containers are increasingly used as means to distribute and run Linu
   address = "United States",
 }
 
-﻿@article{GRUNING2018,
+@article{GRUNING2018,
   author={Gr{\"u}ning, Bj{\"o}rn
   and Dale, Ryan
   and Sj{\"o}din, Andreas
@@ -90,6 +67,7 @@ abstract = {Containers are increasingly used as means to distribute and run Linu
   Author={Romain Reuillon, Mathieu Leclaire, Sebastien Rey-Coyrehourcq},
   Url={http://www.openmole.org/files/FGCS2013.pdf},
   Journal={Future Generation Computer Systems}
+  doi={10.1016/j.future.2013.05.003},
 }
 
 
@@ -100,7 +78,7 @@ abstract = {Containers are increasingly used as means to distribute and run Linu
   pages = {50-59},
   year = {2018},
   issn = {0167-739X},
-  doi = {https://doi.org/10.1016/j.future.2018.01.022},
+  doi = {10.1016/j.future.2018.01.022},
   url = {https://www.sciencedirect.com/science/article/pii/S0167739X17316485},
   author = {Alfonso Pérez and Germán Moltó and Miguel Caballer and Amanda Calatrava},
   keywords = {Cloud computing, Serverless, Docker, Elasticity, AWS lambda},
@@ -174,7 +152,7 @@ and Zok, T.},
 
 @article{BAGNASCHI2019,
[…]
