+++ paper/paper.bib
@@ -66,7 +66,7 @@ abstract = {Containers are increasingly used as means to distribute and run Linu
   Year=2013,
   Author={Romain Reuillon, Mathieu Leclaire, Sebastien Rey-Coyrehourcq},
   Url={http://www.openmole.org/files/FGCS2013.pdf},
-  Journal={Future Generation Computer Systems}
+  Journal={Future Generation Computer Systems},
   doi={10.1016/j.future.2013.05.003},
 }
 
