### 2021-05-14 ca49ea3
add SLURM containers guide
