+++ paper/paper.bib
@@ -550,6 +550,13 @@ and Zok, T.},
   url = "http://pluton.des.udc.es/guide/user-guide.pdf"
 }
 
+@electronic{SLURM,
+  year = {2021},
+  author = {SchedMD},
+  title = {Containers Guide},
+  url = "https://slurm.schedmd.com/containers.html"
+}
+
 @article{KURTZER2017,
     doi = {10.1371/journal.pone.0177459},
     author = {Kurtzer, Gregory M. AND Sochat, Vanessa AND Bauer, Michael W.},
