### 2021-05-14 88e358b
add SLURM containers guide
