### 2021-05-14 #3277 @vsoch
huh, udocker has been around for a bit, I thought it was already published? https://www.sciencedirect.com/science/article/pii/S0010465518302042
