### 2021-05-14 #3277 @jorge-lip
The mentioned publication was focused on version 1.1.0 and 1.1.1 for Python 2 since then udocker evolved a lot. Has more capabilities and was much rewritten for Python 3. Currently the latest Python 3 candidate release is 2.7.1. The initial paper introduced udocker but was also about benchmarking three scientific applications and providing performance comparisons for them with several containerization tools. This paper is only about udocker and also provides consolidated references for further scientific applications of udocker.
