### 2021-05-14 #3277 @vsoch
yes! I love containers and would love to give udocker another spin. I also hugely support that JoSS reviews "non traditional" scientific software papers - this is about containers and environments, and this is on par with my recent submission https://github.com/openjournals/joss-reviews/issues/3217. Sign me up! I can try to make some time in the next few weeks.
