### 2021-05-17 #3277 @mviereck
Thank you for the invitation, I am interested to review.
I am currently looking at udocker, run some first tests and have a look at the paper.

I am a bit unsure yet if I can fulfill the requirements (lack of English language skills to understand everything well). Though, I have some experience with containers. Maybe I should just do the review and point out when/where I am unsure in single check points.
