### 2021-05-19 #3295 @danielskatz
We aim for reviews to be completed within about 2-4 weeks, ideally. Please let me know if either of you require some more time. We can also use Whedon (our bot) to set automatic reminders if you know you'll be away for a known period of time.
