### 2021-05-19 indigo-dc/udocker#311 @mviereck
[joss review] installation instructions & dependencies
Hello,

I am @mviereck, one reviewer of your submisson at https://github.com/openjournals/joss-reviews/issues/3295.

Yet I look at installation instructions and dependency declaration.
Partially I found it a bit unclear and thought to write it down as the impressions are fresh from a non experienced new user.

An unpolished list from first impressions:

----------
> 1. DEPENDENCIES

> Python dependencies are described in the file requirements.txt.

I am lucky to have all dependencies fulfilled as it seems. At least I get no error. Users with rather low experience who get errors might have problems to check or fulfill the python dependencies. How to check, what to do? Why is the list in a separate file at all?

> Other host libraries and tools required: [...]

It might help to write `tar` and `find` instead of tar and find (and others).

Proposal: Provide a package list for some major distributions, maybe also with `apt install` and `dnf install` examples. For comparision, a [dependency table](https://github.com/mviereck/x11docker/wiki/dependencies#recommended-base) I've set up once for another project.

---------------
> 2.1. PYTHON2: INSTALL DIRECTLY FROM GITHUB

So far, the instructions are clear and straightforward. But:

> downloading and installing the udockertools will be performed automatically.

Maybe give a short explanation here already that those are binaries stored in `~/.udocker` and are build from source in the same repo. From first impression it is not clear what the udockertools are, from where they come and where they are stored. Those information can be obtained further down, though, but in parts that are not core installation instructions.

2.1 to 2.4 are independent full installations.
2.5 is not.
2.6? Not sure

3. to 6. are not exactly installation instructions.

I recommend to overthink the structure.

-------------
Missing: Is there an `udocker uninstall`? How to cleanly remove udocker and maybe also images and containers?
