### 2021-05-20 indigo-dc/udocker#312 @vsoch
[joss-review] paper review
The paper does a good job of introducing the software, and explaining the use case. Since some readers (like myself) are probably familiar with udocker, I'd want to know beyond "several new features" how this new release is different, and more importantly, why it warrants a new paper. For the latter, we should have that discussion here or in the main review (not in the paper).

https://github.com/openjournals/joss-reviews/issues/3295
