### 2021-05-20 indigo-dc/udocker#313 @vsoch
[joss-review] repository
These are a set of suggestions for the repository part of the review!

## Link to documentation

I'd like to see this alongisde the repository description, front and center:

![image](https://user-images.githubusercontent.com/814322/118914141-f1e1df00-b8e7-11eb-8083-acbfccdc043e.png)

nice job with tags btw!

## Install instructions

You say "the end user itself" and I think you can just say "the end user" (the itself sounds weird). For your install docs, this is a suggestion, but in the long term you might consider a GitHub pages site (that renders from the docs folder here) that has a bit of branding and looks professional. I think (even though it doesn't make a lot of sense) these small details, or how professional and inviting a project looks, makes a huge difference. You could probably very easy add some jekyll frontend matter headers to your
current markdown files in doc (and rename to docs) and grab a template and get most of the way there, or readthedocs is a nice strategy too for versioned docs! You already have beautifully written docs -> https://github.com/indigo-dc/udocker/blob/devel3/doc/user_manual.md but they would be much more pleasant to read in a non GitHub README / markdown format.

## Getting Started

Right now the README has what I'd call a "quick start" - and this was very good because I was able to pull and create a container. But as an informed user - I really want to know more (even in the quick start). So I'd like to see each command followed by a description of what is happening, what configs are being read and where, and where layers/images are being stored. For example, shelling into a container and seeing:

```
$ udocker run -v /var -v /proc -v /sys -v /tmp  myfed  /bin/bash
Warning: non-existing user will be created

 ******************************************************************************
 *                                                                            *
 *               STARTING e91963ab-918a-3a05-8723-50709b7ce900                *
 *                                                                            *
 ******************************************************************************
 executing: bash
e91963ab#whoami
root
e91963ab#
```
Is a little disconcerting. Why am I root?  Why does it say a non existing user will be created? I shouldn't have any doubts about what is going on here, especially in my quick start when I'm making a decision if this software is something I can trust and want to try. And then for installing software:

```bash
$ udocker run  --user=root myfed  yum install -y firefox pulseaudio gnash-plugin
```
How does this work if I don't have escalated privileges? Even a quick explanation would be useful. This is also a bit scary:

> udocker must not be run by privileged users.

Why not? What happens?

## Examples

You have so many examples in the paper for real world use of udocker! It would be good to see these represented in the repository, and more importantly, links to the code that these projects used so the reader can see examples of using udocker, preserving in version control, etc. Some folks like to learn with tutorials, others want to see how it is used in the wild.

## Python2 vs Python3

I would move Python3 to be the top / first / primary in all the places that you mention it (e.g., Install and README). While it's good to have Python2 support, I'd hope most people are not using it anymore :p  And why is there a devel and then devel3?

https://github.com/openjournals/joss-reviews/issues/3295
