### 2021-05-20 indigo-dc/udocker#311 @vsoch
I second this - it would be good to have a getting started guide that gives some help to installing dependencies, beyond the name in text. https://github.com/openjournals/joss-reviews/issues/3295
