### 2021-05-20 indigo-dc/udocker#315 @vsoch
[joss-review] state of the field
If I'm reading this, it's not clear why I would want to use udocker over something like Singularity or podman (the first which can be an engine, and which are more complete container technologies in terms of building from recipes).

https://github.com/openjournals/joss-reviews/issues/3295
