### 2021-05-20 indigo-dc/udocker#313 @mviereck
> Python2 vs Python3

Id like to add a question: If there is work in progress to move to python3, wouldn't it make sense to finish this before publishing a new paper?
The JOSS guidelines wish to document stable software that doesn't change too much after review.
Currently I focus on the stable python2 version that is going to be outdated.
