+++ .sqa/docker-compose.yml
@@ -1,4 +1,4 @@
-version: "3.6"
+version: "3.7"
 
 services:
   udocker-testing:
@@ -6,5 +6,5 @@ services:
     container_name: "udocker-testing"
     volumes:
       - type: bind
-        source: ./udocker
+        source: ./
         target: /udocker
