### 2021-05-22 0532ec3
fix pylint warns, remove test assert
