+++ tests/unit/test_fakechroot.py
@@ -149,7 +149,7 @@ class FakechrootEngineTestCase(TestCase):
         ufake = FakechrootEngine(self.local, self.xmode)
         ufake.opt["vol"] = ("/tmp", "/bin",)
         status = ufake._get_volume_bindings()
-        self.assertEqual(status, ('', '/tmp!/tmp:/usr/bin!/bin'))
+        self.assertEqual(status, ('', '/tmp!/tmp:/bin!/bin'))
 
     @patch('udocker.engine.fakechroot.os.path.exists')
     @patch('udocker.engine.fakechroot.FileUtil.cont2host')
