### 2021-05-25 b4d4f52
Update build status in README.md to jenkins at eosc-synergy
Signed-off-by: Samuel Bernardo <samuel@lip.pt>
