### 2021-05-25 e8cac1b
Update build status in README.md to jenkins at eosc-synergy
Signed-off-by: Samuel Bernardo <samuel@lip.pt>
