### 2021-05-25 3a05aa4
Update build status in README.md to jenkins at eosc-synergy
Signed-off-by: Samuel Bernardo <samuel@lip.pt>
