+++ doc/user_manual.md
@@ -10,29 +10,32 @@ or other externally managed batch or interactive systems.
 
 udocker does not require any type of privileges nor the deployment of 
 services by system administrators. It can be downloaded and executed 
-entirely by the end user. 
+entirely by the end user. The limited root functionality provided by
+some of the udocker execution modes is either simulated or provided
+via user namespaces.
 
 udocker is a wrapper around several tools and technologies to mimic a 
 subset of the Docker capabilities including pulling images and running 
-then with minimal functionality. It is mainly meant to execute user 
-applications packaged in Docker containers. 
+then with minimal functionality. 
 
-We recommend the use of Docker whenever possible, but when it is 
-unavailable udocker can be the right tool to run your applications.
+udocker is mainly meant to execute user applications packaged in Docker 
+containers. We recommend the use of Docker whenever possible, but when 
+it is unavailable udocker can be the right tool to run your applications.
 
 ## 1. INTRODUCTION
 
 ### 1.1. How does it work
-udocker is a simple tool written in Python, it has a minimal set of 
-dependencies so that can be executed in a wide range of Linux systems. 
-udocker does not make use of Docker nor requires its installation.
+udocker is written in Python, from version 1.2.0 onwards udocker supports
+both Python 3 and Python 2. udocker has a minimal set of dependencies so 
+that can be executed in a wide range of Linux systems. udocker does not 
+make use of Docker nor requires its installation.
 
 udocker "executes" the containers by simply providing a chroot like 
 environment to the extracted container. udocker is meant to integrate
 several technologies and approaches hence providing an integrated environment
 that offers several execution options. This version provides execution engines
-based on PRoot, Fakechroot, runC and Singularity to facilitate the execution
-of Docker containers without privileges.
+based on PRoot, Fakechroot, runc, crun and Singularity to facilitate the
+execution of Docker containers without privileges.
 
 The basic usage flow starts by downloading the image from an image repository
 in the usual way; create the container out of that image (flatenning the image
@@ -74,24 +77,23 @@ Other limitations:
 * debugging and tracing in the PRoot engine will not work;
 * the Fakechroot engine does not support execution of statically linked executables;
 * udocker is mainly oriented at providing a run-time environment for containers execution in user space.
+* udocker does not offer robust isolation features such as the ones offered by docker.
 
 ### 1.3. Security
-Because of the limitations described in section 1.2 udocker does not offer 
-isolation features such as the ones offered by Docker. If the containers
-content is not trusted then they should not be executed within udocker as
-they will run inside the user environment. 
-
-Due to the lack of isolation features udocker must not be run by privileged 
+udocker does not offer robust isolation features such as the ones offered
+by docker. Therefore if the containers content is not trusted then these
+containers should not be executed with udocker as they will run inside the
+user environment. For this reason udocker should not be run by privileged
 users.
+
+udocker does not require privileges and runs under the identity of the user 
+invoking it.
 
 The containers data will be unpacked and stored in the user home directory or 
 other location of choice. Therefore the containers data will be subjected to
 the same filesystem protections as other files owned by the user. If the
 containers have sensitive information the files and directories should be
 adequately protected by the user.
-
-udocker does not require privileges and runs under the identity of the user 
-invoking it.
 
 Users can download the udocker tarball, install in the home directory and 
 execute it from their own accounts without requiring system administration 
@@ -111,10 +113,13 @@ no root privileges are involved. This feature enables tools that do not
 require privileges but that check the user id to work properly. This enables
 for instance software installation with rpm and yum inside the container.
 
-Similarly to Docker, the login credentials for private repositories are stored 
[…]
