+++ doc/installation_manual.md
@@ -1,21 +1,23 @@
-# udocker INSTALLATION MANUAL
+# udocker INSTALLATION and CONFIGURATION MANUAL
 
-In most circumstances the end user can download and execute the tool without
+In most cases the end user can download and execute udocker without
 system administrator intervention. udocker itself is written in Python, but 
 also uses external binaries and libraries to provide a chroot like
 environment where containers are executed in user space. These tools do not
-require any privileges.
+require any privileges and constitute the *udockertools* component that is
+downloaded and installed by udocker itself.
 
 ## 1. DEPENDENCIES
 
-Python dependencies are described in the file requirements.txt.
+udocker requires:
+ * Python 3 or alternatively Python >= 2.6
+ * pycurl or alteratively the curl command
+ * python hashlib or alternatively the openssl command
+ * tar
+ * find
+ * ldconfig (only used by the Fn execution modes)
 
-udocker requires either pycurl or the curl executable command,
-to download both the binaries and/or pull containers from repositories.
-
-tar is needed when using `udocker install` to unpackage binaries and libraries.
-
-## 2. USER INSTALLATION
+## 2. INSTALLATION
 
 ### 2.1. INSTALL FROM A RELEASED VERSION
 
@@ -40,7 +42,9 @@ alternatively use `curl` instead of `wget` as follows:
   export PATH=`pwd`/udocker:$PATH
 ```
 
-Complete the installation by invoking udocker install to download and install udockertools:
+Complete the installation by invoking `udocker install` to download and install 
+the `udockertools` tarball containing the tools and libraries required to execute 
+container:
 
 ```
   udocker install
@@ -51,42 +55,65 @@ you can copy it to your home directory `$HOME` or to `$UDOCKER`
 
 ---
 
-This installation method contains statically compiled binaries and is built 
-to be used across different hosts and OS distributions.
-
 ### 2.2. INSTALL FROM THE DEVELOPMENT BRANCH
 
-To install from the github development branch `devel3`:
+To install the latest stable code from the github `master` branch:
+
+```
+  git clone --depth=1 https://github.com/indigo-dc/udocker.git
+  
+  (cd udocker; ln -s maincmd.py udocker)  
+
+  export PATH=`pwd`/udocker:$PATH
+
+```
+
+or alternatively install the latest development code from the github `devel3` branch:
 
 ```
   git clone -b devel3 --depth=1 https://github.com/indigo-dc/udocker.git
   
+  (cd udocker; ln -s maincmd.py udocker)  
+
   export PATH=`pwd`/udocker:$PATH
+
 ```
 
-As before complete the installation by invoking udocker install to download and install udockertools:
+Complete the installation by invoking `udocker install` to download and install 
+the `udockertools` tarball containing the tools and libraries required to execute 
+containers.
 
 ```
   udocker install
 ```
 
 
-### 2.3. OBTAINING THE URL OF THE TARBALL
+### 2.3. OBTAINING THE URL OF THE udockertools TARBALL
 
-The udocker installation tarball mentioned in section 2.2 can be obtained using the 
-following method. First download udocker. Second use udocker itself to display the
-installation tarball URL by invoking the `version` command. The tarball location may 
-contain several URLs pointing to mirrors.
+When installation is performed without network connectivity the installation 
+of udockertools invoking `udocker install` will fail in the download step.
+The solution is to fetch the the tarball containing the `udockertools` in
+advance and then install it from the tarball file.
+
+The udockertools tarballs are available at https://github.com/jorge-lip/udocker-builds
+
+To download a file use:
 
 ```
-  udocker version
[…]
