+++ docs/installation_manual.md
@@ -1,4 +1,4 @@
-# udocker INSTALLATION MANUAL
+# Installation manual
 
 In most circumstances the end user can download and execute the tool without
 system administrator intervention. udocker itself is written in Python, but 
@@ -6,7 +6,7 @@ also uses external binaries and libraries to provide a chroot like
 environment where containers are executed in user space. These tools do not
 require any privileges.
 
-## 1. DEPENDENCIES
+## 1. Dependencies
 
 Python dependencies are described in the file requirements.txt.
 
@@ -15,195 +15,182 @@ to download both the binaries and/or pull containers from repositories.
 
 tar is needed when using `udocker install` to unpackage binaries and libraries.
 
-## 2. USER INSTALLATION
+## 2. User installation
 
-### 2.1. INSTALL FROM A RELEASED VERSION
+### 2.1. Install from a released version
 
----
 Download a release tarball from https://github.com/indigo-dc/udocker/releases: 
 
-```
-  wget https://github.com/indigo-dc/udocker/releases/download/devel3_1.2.7/udocker-1.2.7.tar.gz
-
-  tar zxvf udocker-1.2.7.tar.gz
-
-  export PATH=`pwd`/udocker:$PATH
+```bash
+wget https://github.com/indigo-dc/udocker/releases/download/1.3.0/udocker-1.3.0.tar.gz
+tar zxvf udocker-1.3.0.tar.gz
+export PATH=`pwd`/udocker:$PATH
 ```
 
-alternatively use `curl` instead of `wget` as follows:
+Alternatively use `curl` instead of `wget` as follows:
 
-```
-  curl -L https://github.com/indigo-dc/udocker/releases/download/devel3_1.2.7/udocker-1.2.7.tar.gz > udocker-1.2.7.tar.gz
-
-  tar zxvf udocker-1.2.7.tar.gz
-
-  export PATH=`pwd`/udocker:$PATH
+```bash
+curl -L https://github.com/indigo-dc/udocker/releases/download/1.3.0/udocker-1.3.0.tar.gz > udocker-1.3.0.tar.gz
+tar zxvf udocker-1.3.0.tar.gz
+export PATH=`pwd`/udocker:$PATH
 ```
 
-Complete the installation by invoking udocker install to download and install udockertools:
+Complete the installation by invoking udocker install to download and
+install udockertools:
 
+```bash
+udocker install
 ```
-  udocker install
-```
 
-A default configuration file can be found in directory `udocker/etc/udocker.conf`,
-you can copy it to your home directory `$HOME` or to `$UDOCKER` 
-
----
+A default configuration file can be found in directory
+`udocker/etc/udocker.conf`, you can copy it to your home directory `$HOME`
+or to `$UDOCKER`.
 
 This installation method contains statically compiled binaries and is built 
 to be used across different hosts and OS distributions.
 
-### 2.2. INSTALL FROM THE DEVELOPMENT BRANCH
+### 2.2. Obtaining the URL of the tarball
 
-To install from the github development branch `devel3`:
+The udocker installation tarball mentioned in section 2.1 can be obtained using
+the following method. First download udocker. Second use udocker itself to
+display the installation tarball URL by invoking the `version` command. The
+tarball location may contain several URLs pointing to mirrors.
 
-```
-  git clone -b devel3 --depth=1 https://github.com/indigo-dc/udocker.git
-  
-  export PATH=`pwd`/udocker:$PATH
-```
-
-As before complete the installation by invoking udocker install to download and install udockertools:
-
-```
-  udocker install
-```
-
-
-### 2.3. OBTAINING THE URL OF THE TARBALL
-
-The udocker installation tarball mentioned in section 2.2 can be obtained using the 
[…]

[…]
