+++ AUTHORS.md
@@ -1,17 +1,20 @@
 # Credits
 
 * Docker https://www.docker.com/
-* PRoot: http://proot.me
-* runC: https://runc.io
+* PRoot https://proot-me.github.io/
+* Fakechroot https://github.com/dex4er/fakechroot/wiki
+* runC https://runc.io/
 * crun https://github.com/containers/crun
-* Fakechroot: https://github.com/dex4er/fakechroot
-* Singularity: http://singularity.lbl.gov
+* Singularity https://www.sylabs.io/
 * Open Container Initiative https://www.opencontainers.org/
 * INDIGO DataCloud https://www.indigo-datacloud.eu
 * DEEP-Hybrid-DataCloud https://deep-hybrid-datacloud.eu
-* LIP https://www.lip.pt
-* INCD https://www.incd.pt
 * EOSC-hub https://eosc-hub.eu
+* EGI-ACE https://www.egi.eu/projects/egi-ace/
+* EOSC-Synergy https://www.eosc-synergy.eu/
+* LIP [https://www.lip.pt](https://www.lip.pt/?section=home&page=homepage&lang=en)
+* INCD [https://www.incd.pt](https://www.incd.pt/?lang=en)
+
 
 # Development Lead
 
@@ -20,6 +23,6 @@
 # Contributors
 
 * Jorge Gomes <jorge@lip.pt>
-* Isabel Campos
 * Mario David <mariojmdavid@gmail.com>
+* Isabel Campos
 * many more
