+++ paper/paper.md
@@ -1,0 +1,183 @@
+
+---
+title: 'udocker: a user oriented tool for unprivileged Linux containers'
+tags:
+  - Python
+  - Linux containers
+  - Distributed computing
+  - Advanced computing
+authors:
+  - name: Jorge Gomes^[corresponding author]
+    orcid: 0000-0002-9142-2596
+    affiliation: "1, 2" # (Multiple affiliations must be quoted)
+  - name: Mario David
+    orcid: 0000-0003-1802-5356
+    affiliation: "1, 2"
+  - name: João Martins
+    orcid: 0000-0003-4308-266X
+    affiliation: "1, 2"
+  - name: João Pina
+    orcid: 0000-0001-8959-5044
+    affiliation: "1, 2" 
+  - name: Samuel Bernardo
+    orcid: 0000-0002-6175-4012
+    affiliation: "1, 2"
+  - name: Isabel Campos
+    orcid: 0000-0002-9350-0383
+    affiliation: 3
+  - name: Pablo Orviz
+    orcid: 0000-0002-2473-6405
+    affiliation: 3
+  - name: Alvaro López-García
+    orcid: 0000-0002-0013-4602
+    affiliation: 3
+    
+affiliations:
+ - name: LIP, Laboratório de Instrumentação e Física Experimental de Partículas, Lisboa, Portugal
+   index: 1
+ - name: INCD, Infraestrutura Nacional de Computação Distribuída, Lisboa, Portugal
+   index: 2
+ - name: IFCA, Consejo Superior de Investigaciones Científicas-CSIC, Santander, Spain
+   index: 3
+date: 12 May 2021
+bibliography: paper.bib
+---
+
+# Summary
+
+Containers are increasingly used to package, distribute and run scientific software. 
+udocker is a tool to enable execution of Linux containers in advanced computing
+environments. Distinctively from other tools, udocker is meant for easy deployment
+and provides multiple execution engines to cope with different host environments. 
+udocker can execute containers with or without using Linux namespaces. udocker is 
+being used by a wide range of projects and research communities to facilitate the 
+execution of Linux containers across heterogeneous computing environments.
+
+# Statement of need
+
+Researchers have at their disposal a wide range of computing resources ranging
+from laptops to high performance computing clusters and cloud services. Enabling 
+execution of scientific codes across such resources often requires significant 
+effort to adapt to the underlying system configurations. This can be particularly
+difficult for codes with complex software dependencies and can become a continuous 
+effort due to system changes and software updates. Furthermore ensuring the 
+reproducibility across heterogeneous computing resources can be challenging when 
+the software needs to be adapted to the specificity of each resource. In this 
+context Linux containers have gained interest as means to enable encapsulation of 
+research software for easier execution across these environments.
+
+udocker is designed to address the requirement of executing scientific 
+applications easily across a wide range of computing systems and digital 
+infrastructures where the user may not have administration privileges, and
+where run times and functionalities to support Linux containers may not be
+available. 
+In addition, udocker also simplifies the researcher interaction with the tools 
+required to execute containers by providing an integrated solution to execute 
+Linux containers leveraging different approaches suitable for unprivileged
+users. Finally by executing containers without privileges udocker decreases the 
+risks of privilege escalation. The udocker development started in 2016 and 
+the original udocker paper [@GOMES2018] documented versions 1.1.0 and 1.1.1.
+Since then several new features have been released.
+
+# Concept
+udocker provides a self contained solution with minimal dependencies to enable
+execution across systems without need of compilation. udocker itself was 
[…]
