+++ paper/paper.bib
@@ -1,0 +1,614 @@
+@article{GOMES2018,
+title = {Enabling rootless Linux Containers in multi-user environments: The udocker tool},
+journal = {Computer Physics Communications},
+volume = {232},
+pages = {84-97},
+year = {2018},
+issn = {0010-4655},
+doi = {10.1016/j.cpc.2018.05.021},
+url = {https://www.sciencedirect.com/science/article/pii/S0010465518302042},
+author = {Jorge Gomes and Emanuele Bagnaschi and Isabel Campos and Mario David and Luís Alves and João Martins and João Pina and Alvaro López-García and Pablo Orviz},
+keywords = {Linux containers, HPC on cloud, Virtualization, Phenomenology, QCD, Biophysics},
+abstract = {Containers are increasingly used as means to distribute and run Linux services and applications. In this paper we describe the architectural design and implementation of udocker, a tool which enables the user to execute Linux containers in user mode. We also present a few practical applications, using a range of scientific codes characterized by different requirements: from single core execution to MPI parallel execution and execution on GPGPUs.}
+}
+
+@misc{GOMES_2017,
+  author="Jorge Gomes and Isabel Campos",
+  title="Researchers Advance User-Level Container Solution for HPC",
+  journal="HPCwire",
+  year="2017",
+  month="Dec. 18,",
+  url= "https://www.hpcwire.com/2017/12/18/researchers-advance-user-level-container-solution-hpc/"
+}  
+
+@book{CWL2016,
+  title = "Common Workflow Language, v1.0",
+  author = "Brad Chapman and John Chilton and Michael Heuer and Andrey Kartashov and Dan Leehr and Herv{\'e} M{\'e}nager and Maya Nedeljkovich and Matt Scales and Stian Soiland-Reyes and Luka Stojanovic",
+  editor = "Peter Amstutz and Crusoe, {Michael R.} and Neboj{\v s}a Tijani{\'c}",
+  note = "Specification, product of the Common Workflow Language working group. http://www.commonwl.org/v1.0/",
+  year = "2016",
+  month = jul,
+  day = "8",
+  doi = "10.6084/m9.figshare.3115156.v2",
+  language = "English",
+  publisher = "figshare ",
+  address = "United States",
+}
+
+@article{GRUNING2018,
+  author={Gr{\"u}ning, Bj{\"o}rn
+  and Dale, Ryan
+  and Sj{\"o}din, Andreas
+  and Chapman, Brad A.
+  and Rowe, Jillian
+  and Tomkins-Tinch, Christopher H.
+  and Valieris, Renan
+  and K{\"o}ster, Johannes
+  and Team, The Bioconda},
+  title={Bioconda: sustainable and comprehensive software distribution for the life sciences},
+  journal={Nature Methods},
+  year={2018},
+  month={Jul},
+  day={01},
+  volume={15},
+  number={7},
+  pages={475-476},
+  issn={1548-7105},
+  doi={10.1038/s41592-018-0046-7},
+  url={https://doi.org/10.1038/s41592-018-0046-7},
+}
+
+@article {REUILLON2013,
+  Title={OpenMOLE, a workflow engine specifically tailored for the distributed exploration of simulation models},
+  Volume=29,
+  Number=8,
+  Pages={1981 - 1990},
+  Year=2013,
[…]
