### 2021-05-27 81f7668
add links to docs/ md fix logo location
