+++ doc/user_manual.md
@@ -1,0 +1,1 @@
+../docs/user_manual.md
