+++ SUMMARY.md
@@ -1,6 +1,6 @@
 # Summary
 
 * [About udocker](README.md)
-* [User manual](doc/user_manual.md)
-* [Installation manual](doc/installation_manual.md)
+* [User manual](docs/user_manual.md)
+* [Installation manual](docs/installation_manual.md)
 * [Contributing](CONTRIBUTING.md)
