+++ utils/build_freng_rpm.sh
@@ -1,242 +1,0 @@
-#!/bin/bash
-
-# ##################################################################
-#
-# Build udocker-freng rpm package
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-# http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-#
-# ##################################################################
-
-sanity_check() 
-{
-    if [ ! -f "$REPO_DIR/udocker.py" ] ; then
-        echo "$REPO_DIR/udocker.py not found aborting"
-        exit 1
-    fi
-}
-
-setup_env()
-{
-    mkdir -p ~/rpmbuild/{BUILD,RPMS,SOURCES,SPECS,SRPMS}
-    if [ ! -e ~/.rpmmacros ]; then
-        echo '%_topdir %(echo $HOME)/rpmbuild' > ~/.rpmmacros
-    fi
-}
-
-udocker_version()
-{
-    $REPO_DIR/udocker.py version | grep "version:" | cut -f2- '-d ' | cut -f1 '-d-'
-}
-
-udocker_tarball_url()
-{
-    $REPO_DIR/udocker.py version | grep "tarball:" | cut -f2- '-d ' | cut -f1 '-d '
-}
-
-patch_fakechroot_source()
-{
-    echo "patch_fakechroot_source"
-
-    pushd "$TMP_DIR/${BASE_DIR}-${VERSION}/fakechroot"
-
-    if [ -e "Fakechroot.patch" ] ; then
-        echo "patch fakechroot source already applied: $PWD/Fakechroot.patch"
-        return
-    fi
-
-    cp ${utils_dir}/fakechroot_source_glibc.patch Fakechroot.patch
-    patch -p1 < Fakechroot.patch
-    popd
-}
-
-patch_patchelf_source1()
-{
-    echo "patch_patchelf_source1"
-
-    pushd "$TMP_DIR/${BASE_DIR}-${VERSION}/patchelf/src"
-
-    if [ -e "Patchelf_make.patch" ] ; then
-        echo "patch patchelf make already applied: $PWD/Patchelf_make.patch"
-        return
-    fi
-
-    cp ${utils_dir}/patchelf_make_dynamic.patch Patchelf_make.patch
-    patch < Patchelf_make.patch
-    popd
-}
-
-patch_patchelf_source2()
-{
-    echo "patch_patchelf_source2"
-
-    pushd "$TMP_DIR/${BASE_DIR}-${VERSION}/patchelf/src"
-
-    if [ -e "Patchelf_code.patch" ] ; then
-        echo "patch patchelf code already applied: $PWD/Patchelf_code.patch"
-        return
-    fi
-
-    cp ${utils_dir}/patchelf_code.patch Patchelf_code.patch
-    patch < Patchelf_code.patch
-    popd
-}
-
-create_source_tarball()
-{
-    /bin/rm $SOURCE_TARBALL 2> /dev/null
-    pushd $TMP_DIR
[…]

[…]
