### 2021-05-28 #3295 @mviereck
@jorge-lip Could you please give some feedback, here or in the tickets? I've seen you are back contributing to udocker again.

@mariojmdavid You are an active contributer of udocker. Maybe you could have a look here, too?

If you don't have the time yet to work at the review, please tell us.

Maybe it makes sense to finish the port to python3 first (and afterwards wait some time for possible bug reports by users) before continuing the review (compare https://github.com/indigo-dc/udocker/issues/313). A beta version does not fit the requirements of joss, and the python2 version is going to be outdated.

I assume it would be no problem to pause the review until the work is done.
