### 2021-05-28 #3295 @mariojmdavid
we have been busy preparing a new release which we plan over the next few days, this is the version in devel3 branch that we consider mature for release and supports python3. We have been also reviewing the documentation and its location, as well as ensuring the proper run of the Jenkins CI pipeline with success
