### 2021-05-28 indigo-dc/udocker#313 @mariojmdavid
clarifying some things
devel is the development branch supporting the 1.1.* releases, it is only python 2 with the monolithic script
devel3 is the new developments, where the code and unit tests where refactored (modulerised), support for python3,
these last few days we have been preparing the release of the version in devel3 which is basically ready, as well as reviewing the documentation
