### 2021-05-28 #3295 @danielskatz
We should decide what's being reviewed here - it's fine for changes to be made in response to review comments, but if there's going to be a new version, perhaps we should not continue this review and start a new review when that version is ready.  Otherwise, it's unfair to the reviewers.  Alternatively, the submitted version (plus changes made in response to the review) can continue to be reviewed.
