### 2021-05-28 #3295 @jorge-lip
Hi,
udocker supports both the Python 2 only version as 1.1.x series and the Python 3 version as 1.2.x series. We will continue maintaining both in parallel for some more time to allow a smooth migration. Both versions reside in different branches and have been maintained in parallel, they are equal in terms of features and user interface, the documentation has very small differences. The differences are mainly in the code structure.

For the sake of clarity we would like to move the Python 3 version to the master branch as this also makes more sense. I hope moving these branches addresses the concerns raised regarding Python 2 and 3 in the best way for everybody. The changes that are ongoing are the adjustments  needed for doing this move, and we are also adding the corrections and improvements suggested by the reviewers to both versions.

Regarding the question on tests we always had SQA tests. udocker must pass these tests before each release and this applies to both versions. We also use Jenkins and have a pipeline that executes the unit tests, integration tests and bandit for security.
