### 2021-05-28 #3295 @vsoch
> They are equal in terms of features and user interface, the documentation has very small differences. The differences are mainly in the code structure.

If the two versions are equal in terms of all of these things and it's primarily the structure of the code that has changed, I'd like to ask again why this warrants another publication? An update from Python 2.x to 3.x with function and feature equivalent I am not convinced is sufficient enough for a new publication.
