### 2021-05-28 #3295 @jorge-lip
As I explained there is a lot of changes in udocker from the version that was published initially back in 2018 to the one we have today, there have been several releases since then. These changes have been applied to both versions Python 2 and Python 3. The changes are in the changelog.
