### 2021-05-29 53d86e0
add --entrypoint to help
