+++ utils/build_tarball.sh
@@ -2,7 +2,7 @@
 
 # ##################################################################
 #
-# Build binaries and create udocker tarball
+# Build binaries and create the udocker-englib tarball
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
@@ -18,14 +18,21 @@
 #
 # ##################################################################
 
-UDOCKER_VERSION_OVERRIDE="1.1.7"
-TARBALL_VERSION_OVERRIDE="1.1.7"
+# The tarball containing the build binaries for udocker is
+# maintained separately and has its own version.
+# Each release of udocker requires a udocker-englib tarball 
+# that is equal or greather than a given base version.
+# i.e. udocker 1.3.0 requires a udocker-englib tarball >= 1.2.7
+
+# This script produces the following udocker-englib tarball
+TARBALL_VERSION_OVERRIDE="1.2.7"
 
 sanity_check() 
 {
     echo "sanity_check"
-    if [ ! -f "$REPO_DIR/udocker.py" ] ; then
-        echo "$REPO_DIR/udocker.py not found aborting"
+    maincmd="$REPO_DIR/udocker/maincmd.py"
+    if [ ! -f "$maincmd" ] ; then
+        echo "$maincmd not found aborting"
         exit 1
     fi
 
@@ -34,21 +41,12 @@ sanity_check()
     fi
 }
 
-udocker_version()
-{
-    if [ -n "$UDOCKER_VERSION_OVERRIDE" ]; then
-        echo "$UDOCKER_VERSION_OVERRIDE"
-    else
-        $REPO_DIR/udocker.py version | grep "version:" | cut -f2- '-d ' | cut -f1 '-d-'
-    fi
-}
-
 tarball_version()
 {
     if [ -n "$TARBALL_VERSION_OVERRIDE" ]; then
 	echo "$TARBALL_VERSION_OVERRIDE"
     else
-        $REPO_DIR/udocker.py version | grep "tarball_release:" | cut -f2- '-d ' | cut -f1 '-d-'
+        $REPO_DIR/maincmd.py version | grep "tarball_release:" | cut -f2- '-d ' | cut -f1 '-d-'
     fi
 }
 
@@ -81,7 +79,7 @@ prepare_proot_source()
     #git clone --branch udocker-1.1.1 --depth=1 https://github.com/jorge-lip/PRoot.git
     #git clone --branch udocker-2 --depth=1 https://github.com/jorge-lip/proot-udocker.git
 
-    git clone --branch udocker-2 https://github.com/jorge-lip/proot-udocker.git
+    git clone --branch udocker-3 https://github.com/jorge-lip/proot-udocker.git
 
     #/bin/rm -Rf $BUILD_DIR/proot-udocker/.git
     #/bin/rm -Rf $BUILD_DIR/proot-udocker/static/care*
@@ -127,7 +125,7 @@ prepare_fakechroot_glibc_source()
     fi
 
     #git clone --depth=1 --branch 2.18 https://github.com/dex4er/fakechroot.git
-    git clone --branch udocker-1 --depth=1 \
[…]
