### 2021-05-30 indigo-dc/udocker#311 @mviereck
I had a closer look at the udockertools.
I found that those are binaries from third party repositories with their own licenses and the source code stored elsewhere. The binaries are stored as tarball in the udocker releases too, if I see right.
This need to be clarified in the documentation:
 - a list of all third party tools and their origin repositories
 - all licenses
 - source location of all tools
