### 2021-05-31 9e97ef9
extract docs at installation
