+++ utils/build_tarball.sh
@@ -206,6 +206,7 @@ prepare_package()
         /bin/mkdir -p "${PACKAGE_DIR}"
         /bin/mkdir -p "${PACKAGE_DIR}/udocker_dir/bin"
         /bin/mkdir -p "${PACKAGE_DIR}/udocker_dir/lib"
+        /bin/mkdir -p "${PACKAGE_DIR}/udocker_dir/doc"
     fi
 }
 
@@ -229,18 +230,21 @@ addto_package_simplejson()
 addto_package_other()
 {
     echo "addto_package_other"
-    /bin/cp -f "${REPO_DIR}/LICENSE"       "${PACKAGE_DIR}/"
-    /bin/cp -f "${REPO_DIR}/README.md"     "${PACKAGE_DIR}/"
-    /bin/cp -f "${REPO_DIR}/CHANGELOG.md"  "${PACKAGE_DIR}/"
-    /bin/cp -R "${REPO_DIR}/doc"           "${PACKAGE_DIR}/udocker_dir/"
+    /bin/cp -f "${REPO_DIR}/LICENSE"                       "${PACKAGE_DIR}/udocker_dir/doc/LICENSE.udocker"
+    /bin/cp -f "${REPO_DIR}/README.md"                     "${PACKAGE_DIR}/udocker_dir/doc/"
+    /bin/cp -f "${REPO_DIR}/CHANGELOG.md"                  "${PACKAGE_DIR}/udocker_dir/doc/"
+    /bin/cp -R "${REPO_DIR}/doc/udocker.1"                 "${PACKAGE_DIR}/udocker_dir/doc/"
+    /bin/cp -R "${REPO_DIR}/docs/installation_manual.md"   "${PACKAGE_DIR}/udocker_dir/doc/"
+    /bin/cp -R "${REPO_DIR}/docs/reference_card.md"        "${PACKAGE_DIR}/udocker_dir/doc/"
+    /bin/cp -R "${REPO_DIR}/docs/user_manual.md"           "${PACKAGE_DIR}/udocker_dir/doc/"
 
-    /bin/cp -f "${REPO_DIR}/ansible_install.yaml" "${PACKAGE_DIR}/"
-    /bin/cp -f "${REPO_DIR}/setup.py"             "${PACKAGE_DIR}/"
+    /bin/cp -f "${REPO_DIR}/ansible_install.yaml"          "${PACKAGE_DIR}/"
+    /bin/cp -f "${REPO_DIR}/setup.py"                      "${PACKAGE_DIR}/"
 
-    #/bin/cp -f "${S_PROOT_DIR}/proot-x86"    "${PACKAGE_DIR}/udocker_dir/bin/"
-    #/bin/cp -f "${S_PROOT_DIR}/proot-x86_64" "${PACKAGE_DIR}/udocker_dir/bin/"
-    /bin/cp -f "${S_PROOT_DIR}/proot-arm"    "${PACKAGE_DIR}/udocker_dir/bin/"
-    /bin/cp -f "${S_PROOT_DIR}/proot-arm64"  "${PACKAGE_DIR}/udocker_dir/bin/"
+    #/bin/cp -f "${S_PROOT_DIR}/proot-x86"                 "${PACKAGE_DIR}/udocker_dir/bin/"
+    #/bin/cp -f "${S_PROOT_DIR}/proot-x86_64"              "${PACKAGE_DIR}/udocker_dir/bin/"
+    /bin/cp -f "${S_PROOT_DIR}/proot-arm"                  "${PACKAGE_DIR}/udocker_dir/bin/"
+    /bin/cp -f "${S_PROOT_DIR}/proot-arm64"                "${PACKAGE_DIR}/udocker_dir/bin/"
 }
 
 # #############################################################################
[…]
