### 2021-05-31 67be733
extract doc upon installation
