+++ README.md
@@ -6,6 +6,10 @@
 
 [![logo](https://raw.githubusercontent.com/indigo-dc/udocker/master/doc/logo-small.png)]()
 
+This branch contains the original udocker for Python 2. This version is
+still maintained but new features will not be added. The Python 3 version
+is now available from the [master branch](https://github.com/indigo-dc/udocker) and [devel3 branch](https://github.com/indigo-dc/udocker/tree/devel3).
+
 udocker is a basic user tool to execute simple docker containers in user
 space without requiring root privileges. Enables download and execution
 of docker containers by non-privileged users in Linux systems where
@@ -16,16 +20,17 @@ managed batch or interactive systems.
 
 udocker does not require any type of privileges nor the deployment of
 services by system administrators. It can be downloaded and executed
-entirely by the end user.
+entirely by the end user. The limited root functionality provided by
+some of the udocker execution modes is either simulated or provided 
+via user namespaces.
 
 udocker is a wrapper around several tools to mimic a subset of the
 docker capabilities including pulling images and running containers
 with minimal functionality.
 
 ## How does it work
-udocker is a simple tool written in Python, it has a minimal set
-of dependencies so that can be executed in a wide range of Linux
-systems.
+udocker is written in Python, it has a minimal set of dependencies so
+that can be executed in a wide range of Linux systems.
 
 udocker does not make use of docker nor requires its presence.
 
@@ -34,7 +39,7 @@ environment over the extracted container. The current implementation
 supports different methods to mimic chroot thus enabling execution of
 containers under a chroot like environment without requiring privileges.
 udocker transparently supports several methods to execute the containers
-using tools and libraries such as:
+based on tools and libraries such as:
 
 * PRoot
 * Fakechroot
@@ -42,27 +47,40 @@ using tools and libraries such as:
 * crun
 * Singularity
 
+With the exception of Singularity the tools and libraries to support 
+execution are downloaded and deployed by udocker during the installation 
+process. This installation is performed in the user home directory
+and does not require privileges. The udocker related files such as
+libraries, executables, container images and extracted directory
+trees are placed by default under `$HOME/.udocker`.
+
 ## Advantages
-* Provides a docker like command line interface
-* Supports a subset of docker commands:
-  search, pull, import, export, load, save, create and run
-* Understands docker container metadata
-* Allows loading of docker and OCI containers
+
 * Can be deployed by the end-user
 * Does not require privileges for installation
 * Does not require privileges for execution
-* Does not require compilation, just transfer the Python script and run
+* Does not require compilation, just transfer the Python script
 * Encapsulates several tools and execution methods
-* Includes the required tools already statically compiled to work across systems
-* Tested with GPGPU and MPI applications
+* Includes the required tools already statically compiled to work
+  across systems
+* Provides a docker like command line interface
+* Supports a subset of docker commands:
+  search, pull, import, export, load, save, create, login, logout and run
+* Understands docker container metadata
+* Allows loading of docker and OCI containers
+* Supports GPGPU and MPI applications
 * Runs both on new and older Linux distributions including:
-  CentOS 6, CentOS 7, CentOS 8, Ubuntu 14, Ubuntu 16, Ubuntu 18, Ubuntu 20, Fedora, etc
+  CentOS 6, CentOS 7, CentOS 8, Ubuntu 14, Ubuntu 16, Ubuntu 18, Ubuntu 20,
+  Fedora, etc
 
 ## Installation and Users manuals
 
 See the **[Installation manual](doc/installation_manual.md)**
 See the **[Users manual](doc/user_manual.md)**.
 See the **[All documentation for Python 2](SUMMARY.md)**
+
+For the new Python 3 version:
+
[…]
