+++ utils/build_tarball.sh
@@ -2,7 +2,7 @@
 
 # ##################################################################
 #
-# Build binaries and create udocker tarball
+# Build binaries and create the udocker-englib tarball
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
@@ -18,37 +18,35 @@
 #
 # ##################################################################
 
-UDOCKER_VERSION_OVERRIDE="1.1.7"
-TARBALL_VERSION_OVERRIDE="1.1.7"
+# The tarball containing the build binaries for udocker is
+# maintained separately and has its own version. Each release of
+# udocker requires a tarball that is equal or greather than a given
+# base version.
+# i.e. udocker 1.3.0 requires a tarball >= 1.2.8
+
+# This script produces the following udocker-englib tarball
+
+DEVEL3=$(realpath "$0" | grep "devel3")
+
+TARBALL_VERSION_P3="1.2.8"
+TARBALL_VERSION_P2="1.1.8"
 
 sanity_check() 
 {
     echo "sanity_check"
-    if [ ! -f "$REPO_DIR/udocker.py" ] ; then
-        echo "$REPO_DIR/udocker.py not found aborting"
+    local udocker_script
+    if [ -n "$DEVEL3" ]; then
+        udocker_script="$REPO_DIR/udocker/maincmd.py"
+    else
+	udocker_script="$REPO_DIR/udocker.py"
+    fi
+    if [ ! -f "$udocker_script" ] ; then
+        echo "$udocker_script not found aborting"
         exit 1
     fi
 
     if [ -e "$BUILD_DIR" ] ; then
         echo "$BUILD_DIR already exists"
-    fi
-}
-
-udocker_version()
-{
-    if [ -n "$UDOCKER_VERSION_OVERRIDE" ]; then
-        echo "$UDOCKER_VERSION_OVERRIDE"
-    else
-        $REPO_DIR/udocker.py version | grep "version:" | cut -f2- '-d ' | cut -f1 '-d-'
-    fi
-}
-
-tarball_version()
-{
-    if [ -n "$TARBALL_VERSION_OVERRIDE" ]; then
-	echo "$TARBALL_VERSION_OVERRIDE"
-    else
-        $REPO_DIR/udocker.py version | grep "tarball_release:" | cut -f2- '-d ' | cut -f1 '-d-'
     fi
 }
 
@@ -78,10 +76,9 @@ prepare_proot_source()
     fi
 
     #git clone --branch v5.1.0 --depth=1 https://github.com/proot-me/PRoot 
-    #git clone --branch udocker-1.1.1 --depth=1 https://github.com/jorge-lip/PRoot.git
     #git clone --branch udocker-2 --depth=1 https://github.com/jorge-lip/proot-udocker.git
 
-    git clone --branch udocker-2 https://github.com/jorge-lip/proot-udocker.git
+    git clone --branch udocker-3 https://github.com/jorge-lip/proot-udocker.git
 
     #/bin/rm -Rf $BUILD_DIR/proot-udocker/.git
     #/bin/rm -Rf $BUILD_DIR/proot-udocker/static/care*
[…]
