### 2021-05-31 f6911b1
add follow location to GetURL()
