### 2021-05-31 d80bcdf
improve installation
