### 2021-06-03 indigo-dc/udocker#311 @mviereck
Just want to note, that the third party tools of udocker are not clearly stated and are missing source, license and origin. This needs to be documented well. In the current state this is an acceptance blocker. Compare https://github.com/indigo-dc/udocker/issues/311
