### 2021-06-05 6e710ca
fix unit tests, remove unused udockettools test
