+++ tests/unit/test_udockertools.py
@@ -1,304 +1,0 @@
-#!/usr/bin/env python
-"""
-udocker unit tests: UdockerTools
-"""
-
-import sys
-from unittest import TestCase, main
-from udocker.tools import UdockerTools
-from udocker.utils.curl import CurlHeader
-from udocker.config import Config
-from udocker.container.localrepo import LocalRepository
-try:
-    from unittest.mock import patch
-except ImportError:
-    from mock import patch
-
-STDOUT = sys.stdout
-STDERR = sys.stderr
-UDOCKER_TOPDIR = "test_topdir"
-
-if sys.version_info[0] >= 3:
-    BUILTINS = "builtins"
-else:
-    BUILTINS = "__builtin__"
-
-
-class UdockerToolsTestCase(TestCase):
-    """Test UdockerTools() download and setup of tools needed by udocker."""
-
-    def setUp(self):
-        Config().getconf()
-        self.local = LocalRepository()
-
-    def tearDown(self):
-        pass
-
-    @patch('udocker.tools.GetURL')
-    def test_01_init(self, mock_geturl):
-        """Test01 UdockerTools() constructor."""
-        Config().conf['tmpdir'] = "/tmp"
-        Config().conf['tarball'] = "/tmp/xxx"
-        Config().conf['installinfo'] = "/tmp/xxx"
-        Config().conf['tarball_release'] = "0.0.0"
-        self.local.bindir = "/bindir"
-        utools = UdockerTools(self.local)
-        self.assertEqual(utools.localrepo, self.local)
-        self.assertTrue(mock_geturl.called)
-
-    # def test_02__instructions(self):
-    #     """Test02 UdockerTools()._instructions."""
-
-    # def test_03__version2int(self):
-    #     """Test03 UdockerTools()._version2int."""
-
-    # def test_04__version_isok(self):
-    #     """Test04 UdockerTools()._version_isok."""
-
-    # @patch.object(UdockerTools, '_version_isequal')
-    # def test_05_is_available(self, mock_version_isequal):
-    #     """Test05 UdockerTools.is_available()."""
-    #     mock_version_isequal.return_value = False
-    #     utools = UdockerTools(self.local)
-    #     status = utools.is_available()
-    #     self.assertFalse(status)
-
-    #     mock_version_isequal.return_value = True
-    #     utools = UdockerTools(self.local)
-    #     status = utools.is_available()
-    #     self.assertTrue(status)
-
-    @patch('udocker.tools.FileUtil.remove')
-    @patch('udocker.tools.os.listdir')
-    def test_06_purge(self, mock_listdir, mock_remove):
-        """Test06 UdockerTools.purge()."""
-        mock_listdir.return_value = []
-        utools = UdockerTools(self.local)
-        utools.purge()
-        self.assertFalse(mock_remove.called)
-
-        mock_listdir.return_value = ["F1", "F2"]
-        utools = UdockerTools(self.local)
-        utools.purge()
-        self.assertTrue(mock_remove.called)
-
-    # @patch('udocker.tools.GetURL.get')
-    # @patch('udocker.tools.FileUtil')
-    # def test_07__download(self, mock_futil, mock_gurl):
-    #     """Test07 UdockerTools._download()."""
-    #     hdr = CurlHeader()
-    #     hdr.data["X-ND-HTTPSTATUS"] = "200"
-    #     hdr.data["X-ND-CURLSTATUS"] = 0
-    #     mock_futil.return_value.mktmp.return_value = "tmptarball"
-    #     mock_gurl.get.return_value = (hdr, "")
-    #     utools = UdockerTools(self.local)
-    #     utools.curl = mock_gurl
[…]
