+++ docs/installation_manual.md
@@ -4,20 +4,21 @@ In most cases the end user can download and execute udocker without
 system administrator intervention. udocker itself is written in Python, but 
 also uses external binaries and libraries to provide a chroot like
 environment where containers are executed in user space. These tools do not
-require any privileges and constitute the `udockertools` component that is
-downloaded and installed by udocker itself.
+require any privileges and constitute the udocker tools and libraries for 
+engines that is downloaded and installed by udocker itself.
 
 ## 1. Dependencies
 
+The udocker dependencies are minimal and should be supported by most Linux installations.
 udocker requires:
- * Python 3 or alternatively Python >= 2.6
- * pycurl or alteratively the curl command
- * python hashlib or alternatively the openssl command
- * tar
- * find
- * chmod
- * chgrp
- * ldconfig (only used by the Fn execution modes)
+* Python 3 or alternatively Python >= 2.6
+* pycurl or alteratively the curl command
+* python hashlib or alternatively the openssl command
+* tar
+* find
+* chmod
+* chgrp
+* ldconfig (only used by the Fn execution modes)
 
 ## 2. Installation
 
@@ -41,20 +42,14 @@ export PATH=`pwd`/udocker:$PATH
 ```
 
 Complete the installation by invoking `udocker install` to download and install 
-the `udockertools` tarball containing the tools and libraries required to execute 
-container:
+the tarball containing the tools and libraries required to execute containers:
 
 ```
 udocker install
 ```
 
-A default configuration file can be found in directory `udocker/etc/udocker.conf`,
-you can copy it to your home directory `$HOME` or to `$UDOCKER`.
 
-Both installed files, as well as the containers to be downloaded or created  
-with udocker, will be installed by default under `$HOME/.udocker`.
-
-### 2.2. Install from the development branch
+### 2.2. Install from the GitHub repositories
 
 To install the latest stable code from the github `master` branch:
 
@@ -80,65 +75,66 @@ containers.
 udocker install
 ```
 
-### 2.3. Obtaining the URL of the udockertools tarball
+### 2.3. Installing without outbound network access
 
 When installation is performed without network connectivity the installation 
-of `udockertools` invoking `udocker install` will fail in the download step.
-The solution is to fetch the the tarball containing the `udockertools` in
-advance and then install it from the tarball file.
+of the udocker tools and libraries invoking `udocker install` will fail in the 
+download step. The solution is to fetch the the tarball in advance and then 
+install it directly from the tarball file.
 
 The `udockertools` tarballs are available at https://github.com/jorge-lip/udocker-builds.
 
-To download a file use:
+To download a tarball file use:
 
 ```bash
-curl -L https://github.com/jorge-lip/udocker-builds/raw/master/tarballs/udockertools-1.2.7.tar.gz > udockertools-1.2.7.tar.gz
+curl -L https://github.com/jorge-lip/udocker-builds/raw/master/tarballs/udocker-englib-1.2.8.tar.gz > udocker-englib-1.2.8.tar.gz
 ```
 
-Finally transfer udocker and udockertools to the target destination host and
-perform the `udocker install` step using the tarball file.
+Finally transfer both udocker and tools and libraries tarball to the target 
+destination host and perform the `udocker install` step using the tarball file.
 
 ```bash
-export UDOCKER_TARBALL=udockertools-filename
+export UDOCKER_TARBALL=udocker-tools-filename
 udocker install
 ```
 
 The environment variable `UDOCKER_TARBALL` can also point to an URL to fetch 
 the `udockertools` from a specific or alternate location.
 
-### 2.4. Force the re-installation of udockertools
+### 2.4. Force the re-installation of the tools and libraries
 
[…]

[…]
