+++ docs/user_manual.md
@@ -39,7 +39,7 @@ based on PRoot, Fakechroot, runc, crun and Singularity to facilitate the
 execution of Docker containers without privileges.
 
 The basic usage flow starts by downloading the image from an image repository
-in the usual way; create the container out of that image (flatenning the image
+in the usual way; create the container out of that image (flattening the image
 on the filesystem), and finally run the container with the name we gave it in
 the creation process:
 
@@ -108,7 +108,7 @@ currently implemented by:
 
 * PRoot engine via the kernel ptrace system call;
 * Fakechroot engine via shared library preload;
-* runC engine using rootless namespaces;
+* runc engine using rootless namespaces;
 * Singularity if available in the host system.
 
 udocker via PRoot offers the emulation of the root user. This emulation
@@ -823,15 +823,15 @@ udocker setup [--execmode=XY] [--force] [--nvidia] [--purge] CONTAINER-ID|CONTAI
 ```
 
 With `--execmode` chooses an execution mode to define how a given container 
-will be executed, namelly enables selection of an execution engine and 
+will be executed, namely enables selection of an execution engine and 
 its related execution modes. Without options, setup will print the current 
 execution mode for the given container. 
 The option `--nvidia` enables access to GPGPUs by adding the necessary host 
 libraries to the container.
 The option `--force` can be used both with `--execmode` and with `--nvidia` to 
 force the setup of the container to the specified mode.
-The option `--purge` removes mountpoints, auxiliary files and directories 
-created by udockr inside the container directory tree to support its execution.
+The option `--purge` removes mount points, auxiliary files and directories 
+created by udocker inside the container directory tree to support its execution.
 It should only be invoked when there is no execution taking place as it may
 affect processes running in the container tree.
 
@@ -843,7 +843,7 @@ Options:
   force the change of an execution mode when it fails namely if it is
   transferred to a remote host while in one of the Fn modes. Can be
   used with --nvidia.
-* `--purge` remove mountpoints, auxiliary files and directories created
+* `--purge` remove mount points, auxiliary files and directories created
   by udocker to support the container execution.
 
 |Mode| Engine      | Description                               | Changes container
@@ -854,9 +854,9 @@ Options:
 | F2 | Fakechroot  | F1 plus modified loader                   | F1 + ld.so
 | F3 | Fakechroot  | fix ELF headers in binaries               | F2 + ELF headers
 | F4 | Fakechroot  | F3 plus enables new executables and libs  | same as F3
-| R1 | runC        | rootless user mode namespaces             | resolv, passwd
-| R2 | runC        | R1 plus P1 for software installation      | resolv, passwd, proot
-| R3 | runC        | R1 plus P2 for software installation      | resolv, passwd, proot
+| R1 | runc        | rootless user mode namespaces             | resolv, passwd
+| R2 | runc        | R1 plus P1 for software installation      | resolv, passwd, proot
+| R3 | runc        | R1 plus P2 for software installation      | resolv, passwd, proot
 | S1 | Singularity | uses singularity if available in the host | passwd
 
 The default execution mode is P1 using PRoot and starting in root
@@ -906,7 +906,7 @@ execution chain is:
 
 When using the Rn modes, udocker will search for a runc executable in the
 host system, only if it does not find one it will default to use the runc
-provided with the udockertools. This behavior can be change through
+provided with the udocker tools. This behavior can be change through
 environment variables and configuration settings.
 Fakechroot requires libraries compiled for each guest operating system,
 udocker provides these libraries for several distributions including
@@ -921,7 +921,7 @@ host, and then convert again from Pn to the desired Fn mode.
 Singularity must be available in the host system for execution mode S1.
[…]
