+++ doc/user_manual.md
@@ -1,1382 +1,0 @@
-# User manual
-
-A basic user tool to execute simple Docker containers in user space 
-without requiring root privileges. udocker enables basic download and execution 
-of Docker containers by non-privileged users in Linux systems where Docker 
-is not available. It can be used to access and execute the content of 
-docker containers in Linux batch systems and interactive clusters that 
-are managed by other entities such as grid infrastructures, HPC clusters
-or other externally managed batch or interactive systems.
-
-udocker does not require any type of privileges nor the deployment of 
-services by system administrators. It can be downloaded and executed 
-entirely by the end user. The limited root functionality provided by
-some of the udocker execution modes is either simulated or provided
-via user namespaces.
-
-udocker is a wrapper around several tools and technologies to mimic a 
-subset of the Docker capabilities including pulling images and running 
-then with minimal functionality. 
-
-udocker is mainly meant to execute user applications packaged in Docker 
-containers. We recommend the use of Docker whenever possible, but when 
-it is unavailable udocker can be the right tool to run your applications.
-
-## 1. Introduction
-
-### 1.1. How does it work
-
-udocker is written in Python, from version 1.2.0 onwards udocker supports
-both Python 3 and Python 2. udocker has a minimal set of dependencies so 
-that can be executed in a wide range of Linux systems. udocker does not 
-make use of Docker nor requires its installation.
-
-udocker "executes" the containers by simply providing a chroot like 
-environment to the extracted container. udocker is meant to integrate
-several technologies and approaches hence providing an integrated environment
-that offers several execution options. This version provides execution engines
-based on PRoot, Fakechroot, runc, crun and Singularity to facilitate the
-execution of Docker containers without privileges.
-
-The basic usage flow starts by downloading the image from an image repository
-in the usual way; create the container out of that image (flattening the image
-on the filesystem), and finally run the container with the name we gave it in
-the creation process:
-
-* `udocker pull` busybox
-* `udocker create` --name=verybusy busybox
-* `udocker run` verybusy
-
-This sequence allows the created container to be executed many times. If simultaneous
-executions are envisage just make sure that input/output files are not overwritten by
-giving them different names during execution as the container will be shared among
-executions.
-
-Containers can also be pulled, created and executed in a single step. However in this
-case a new container is created for every run invocation thus occupying more storage
-space. To pull, create and execute in a single step invoke run with an image name
-instead of container name:
-
-* `udocker run` busybox
-
-### 1.2. Limitations
-
-Since root privileges are not involved, any operation that really requires privileges
-is not possible. The following are examples of operations that are not possible:
-
-* accessing host protected devices and files;
-* listening on TCP/IP privileged ports (range below 1024);
-* mount file-systems;
-* the su command will not work;
-* change the system time;
-* changing routing tables, firewall rules, or network interfaces.
-
-Other limitations:
-
-* the current implementation is limited to the pulling of Docker images and its execution;
-* the actual containers should be built using Docker and dockerfiles;
-* udocker does not provide all the Docker features, and is not intended as a Docker replacement;
-* debugging and tracing in the PRoot engine will not work;
[…]
