+++ doc/installation_manual.md
@@ -1,461 +1,0 @@
-# Installation and configuration manual
-
-In most cases the end user can download and execute udocker without
-system administrator intervention. udocker itself is written in Python, but 
-also uses external binaries and libraries to provide a chroot like
-environment where containers are executed in user space. These tools do not
-require any privileges and constitute the udocker tools and libraries for 
-engines that is downloaded and installed by udocker itself.
-
-## 1. Dependencies
-
-The udocker dependencies are minimal and should be supported by most Linux installations.
-udocker requires:
-* Python 3 or alternatively Python >= 2.6
-* pycurl or alternatively the curl command
-* python hashlib or alternatively the openssl command
-* tar
-* find
-* chmod
-* chgrp
-* ldconfig (only used by the Fn execution modes)
-
-## 2. Installation
-
-### 2.1. Install from a released version
-
-Download a release tarball from https://github.com/indigo-dc/udocker/releases: 
-
-```bash
-wget https://github.com/indigo-dc/udocker/releases/download/1.3.0/udocker-1.3.0.tar.gz
-tar zxvf udocker-1.3.0.tar.gz
-export PATH=`pwd`/udocker:$PATH
-```
-
-Alternatively use `curl` instead of `wget` as follows:
-
-```bash
-curl -L https://github.com/indigo-dc/udocker/releases/download/1.3.0/udocker-1.3.0.tar.gz \
-  > udocker-1.3.0.tar.gz
-tar zxvf udocker-1.3.0.tar.gz
-export PATH=`pwd`/udocker:$PATH
-```
-
-Complete the installation by invoking `udocker install` to download and install 
-the tarball containing the tools and libraries required to execute containers:
-
-```
-udocker install
-```
-
-
-### 2.2. Install from the GitHub repositories
-
-To install the latest stable code from the github `master` branch:
-
-```bash
-git clone --depth=1 https://github.com/indigo-dc/udocker.git
-(cd udocker; ln -s maincmd.py udocker)  
-export PATH=`pwd`/udocker:$PATH
-```
-
-Alternatively, install the latest development code from the github `devel3` branch:
-
-```bash
-git clone -b devel3 --depth=1 https://github.com/indigo-dc/udocker.git
-(cd udocker; ln -s maincmd.py udocker)  
-export PATH=`pwd`/udocker:$PATH
-```
-
-Complete the installation by invoking `udocker install` to download and install 
-the `udockertools` tarball containing the tools and libraries required to execute 
-containers.
-
-```bash
-udocker install
-```
-
-### 2.3. Installing without outbound network access
-
-When installation is performed without network connectivity the installation 
-of the udocker tools and libraries invoking `udocker install` will fail in the 
-download step. The solution is to fetch the the tarball in advance and then 
-install it directly from the tarball file.
-
-The `udockertools` tarballs are available at https://github.com/jorge-lip/udocker-builds.
-
-To download a tarball file use:
-
-```bash
-curl -L https://github.com/jorge-lip/udocker-builds/raw/master/tarballs/udocker-englib-1.2.8.tar.gz > udocker-englib-1.2.8.tar.gz
-```
-
-Finally transfer both udocker and tools and libraries tarball to the target 
-destination host and perform the `udocker install` step using the tarball file.
-
-```bash
-export UDOCKER_TARBALL=udocker-tools-filename
-udocker install
-```
-
-The environment variable `UDOCKER_TARBALL` can also point to an URL to fetch 
-the `udockertools` from a specific or alternate location.
-
-### 2.4. Force the re-installation of the tools and libraries
-
-To force download and re-installation of the udocker tools and libraries. Invoke `udocker install` 
-with the flag `--force`:
-
-```bash
[…]
