### 2021-06-08 indigo-dc/udocker#313 @jorge-lip
According to comments, we moved the focus to the Python 3 version which is now in the master branch,
and we also released version 1.3.0. As recommended Tthe documentation has been improved including
the README, user manual and installation manual. The rendering of the documentation is now performed
with a jekyll frontend.
