### 2021-06-08 indigo-dc/udocker#315 @jorge-lip
udocker enables execution without using namespaces or any other of the kernel functionalities required by Singularity or
Podman. Also does not require privileges to execute. Many sites refuse to support Singularity due to the past security
vulnerabilities. We have many users from the life sciences community where security and privacy are a concern.

With the udocker default execution mode you can even execute applications compiled for the latest kernels and libc
in systems as old as CentOS 6, where supposedly you would get a kernel too old message when using other solutions.
The fakechroot based mode of udocker is also in many cases faster than other containerization solutions.

The tool also does not require compilation or system installation making it ideal to execute in remote clusters and
grid infrastructures as you can submit udocker with the batch job.

Notice that computer clusters are usually conservative they run with OSes like CentOS 7 using batch systems like
Slurm. In such environment you likely don't have the user namespace enabled, meaning that containerized solutions
like Singularity need to be installed by the sysadmins with privileges. So if Singularity is not installed you will not be
able to run containers.

Being able to use a single solution across systems is also very important when you are using multiple computing
clusters with different characteristics and configurations. It provides a common basis for the execution of the
applications across these heterogeneous environments. The EGI infrastructure www.egi.eu and several of its user
communities are using  udocker to support execution of containers in batch systems in computing clusters across
Europe.

udocker is also unique in supporting multiple execution modes (using different tools) to give the best chances of
running applications across different environments. We revised the documentation, hopefully the advantages and
characteristics are now more clear.
