### 2021-06-08 indigo-dc/udocker#312 @jorge-lip
We lust released 1.3.0 and we added anew section to the paper with the most relevant changes since 1.1.1
when it was initially published.
