### 2021-06-08 indigo-dc/udocker#315 @vsoch
Could you comment about Shifter and Sarus (and possibly add this to the paper?)
