### 2021-06-08 indigo-dc/udocker#311 @jorge-lip
We change the dependencies section and placed in the beginning of the installation manual. The dependencies file had al dependencies necessary both for usage but also for development and testing within our Jenkins. The actual dependencies for normal usage are smaller and we made that more clear.

The list of tools was already in the manuals and in the acknowledgments. The licenses are shipped in the udockertools tarball, but we were not extracting them upon installation. Now since the just release version 1.3.0 we are extracting the licenses of the tools and libraries to the installation directory ($HOME/.udocker/doc).

The source code for the tools is in separate repositories. For runc and crun we just take the software from the upstream source code repositories and we make minor changes to produce static binaries. For Proot and Fakechroot we have cloned repositories with all the required changes. Some we have pushed already upstream. All the required steps for building the tarball are in the build script in utils/build_tarball.sh

The repositories containing the source code are now described in the documentation
https://github.com/indigo-dc/udocker/blob/master/docs/installation_manual.md#6-source-code-for-the-tools-and-libraries)
