### 2021-06-08 #3295 @jorge-lip
The changes since v1.1.1 are now highlighted in text. The external tools used by udocker are highlighted in the documentation and in the acknowledgements, the licenses are included in the tools tarball and since the just released v1.3.0  they are extracted to the installation directory under (.udocker/doc). The installation documentation was revised and now makes clear what source code repositories are used to build the tools. The script to build the tools was already provided in the udocker repository in utils/build_tarball.sh
