### 2021-06-08 indigo-dc/udocker#312 @vsoch
Thank you! I'm reading now and I have a few suggestions:

> Support for OCI containers format was added.

OCI has several standards for containers - could you clarify in the paper what "OCI containers format" is referring to?

> The udocker implementation of the Docker Hub API was reim-77plemented and support for the Docker Hub v2 API was added.

It's confusing that this line follows the line about OCI - the "Docker Hub V2 API is the OCI distribution spec." If there is some variance between those two, what is it?

> The verification of containers78images including checksumming was also implemented

When and how does this happen?

> The search functionality was reimplemented

How? Is it better/faster/something else?

> Support for v1 and v2 repositories and listing of tags was added.

I don't think the average reader will know what v1/v2 repositories are.

> The command line was significantly improved fixing many of the original lim-82itations.

Like?

Ill stop copy pasting line by line - this feels like you copy pasted lines from a changelog, and it reads very robotically and is not friendly to the reader, who might not know some of these technical references, and if they do, there isn't enough detail to understand what was really done (e.g., saying "was fixed or was implemented" doesn't really tell me much.

Since the entire purpose of this new publication is for these significant updates, I'd like to see them possibly more fleshed out so I can read the paper and it's sufficient to understand the changes.
