### 2021-06-09 indigo-dc/udocker#312 @mviereck
I second this.
The basic need and advantage of udocker as a container solution for unprivileged users is clearly stated.
Though, this goes for the udocker version of the first paper, too.
There has been some development and improvement since then; though, I still don't see a major difference.

Maybe it helps to change the perspective:
If someone writes a scientific paper and wants to cite udocker: at which point he would have the _urgent  need_  to cite the new paper instead of the previous one?
