### 2021-06-09 indigo-dc/udocker#312 @vsoch
That’s a good idea @mviereck !
