### 2021-06-10 indigo-dc/udocker#312 @mviereck
For refererence, a link to the udocker paper of 2018:
https://www.sciencedirect.com/science/article/pii/S0010465518302042
