### 2021-06-12 indigo-dc/udocker#311 @mviereck
The structure of the installation instructions looks better overall.
However, I am still missing some core points:
 - Please explicitly state that `udocker install` installs third party tools.
 - Give also a link to the origin of the third party tools, not only to forked repositories.
   - > we make minor changes to produce static binaries.

      Please note this, too.
 - Please explicitly list every licenses at each tool, also with a link to see the full license.

Minor detail: The links should be clickable, not only copy&paste. (I refer to the link given on the main site: https://indigo-dc.github.io/udocker/installation_manual.html)

> The list of external tools to support the actual execution was already in the manuals and in the acknowledgments.

You could have pointed me to this. I only found general notes mixing installed udocker-tools like proot with non-installed tools like singularity.
