### 2021-06-13 #3295 @mviereck
There is already a [paper for udocker](https://www.sciencedirect.com/science/article/pii/S0010465518302042) as @vsoch already pointed out in the preview.
udocker has seen some development since then, but the major difference is still not obvious.
Discussion at: https://github.com/indigo-dc/udocker/issues/312

@danielskatz It is normal that software evolves further after a paper has published. How is this handled within joss? I did not found a guidance for this.
I found a rule that software should have reached a stable state before publishing a paper and should not expect major or even breaking changes.
But how is regular development handled? What if someone cites a tool but uses the recent version instead of the one published in the paper?
I assume those points are already discussed somewhere. This might help here.
