### 2021-06-13 #3295 @danielskatz
The general feeling on this in JOSS is that each JOSS paper should be a clear contribution, in the same way that a research project can lead to multiple papers, some of which can be incremental progress in an area that was previously published. It's up the reviewers to decide if there is sufficient new material/progress.  This is described in https://joss.readthedocs.io/en/latest/submitting.html#substantial-scholarly-effort

Another view of this is that it might be reasonable to have a new paper for a new major release, but typically isn't for a minor release, assuming major releases don't happen too frequently.
