+++ paper/paper.bib
@@ -1,622 +1,0 @@
-@article{GOMES2017,
-  author    = {Jorge Gomes and
-               Isabel Campos Plasencia and
-               Emanuele Bagnaschi and
-               M{\'{a}}rio David and
-               Lu{\'{\i}}s Alves and
-               Jo{\~{a}}o Martins and
-               Jo{\~{a}}o Murta Pina and
-               {\'{A}}lvaro L{\'{o}}pez Garc{\'{\i}}a and
-               Pablo Orviz Fern{\'{a}}ndez},
-  title     = {Enabling rootless Linux Containers in multi-user environments: the
-               udocker tool},
-  journal   = {CoRR},
-  volume    = {abs/1711.01758},
-  year      = {2017},
-  url       = {http://arxiv.org/abs/1711.01758},
-  archivePrefix = {arXiv},
-  eprint    = {1711.01758},
-  timestamp = {Wed, 11 Dec 2019 15:12:01 +0100},
-  biburl    = {https://dblp.org/rec/journals/corr/abs-1711-01758.bib},
-  bibsource = {dblp computer science bibliography, https://dblp.org}
-}
-
-@article{GOMES2018,
-title = {Enabling rootless Linux Containers in multi-user environments: The udocker tool},
-journal = {Computer Physics Communications},
-volume = {232},
-pages = {84-97},
-year = {2018},
-issn = {0010-4655},
-doi = {https://doi.org/10.1016/j.cpc.2018.05.021},
-url = {https://www.sciencedirect.com/science/article/pii/S0010465518302042},
-author = {Jorge Gomes and Emanuele Bagnaschi and Isabel Campos and Mario David and Luís Alves and João Martins and João Pina and Alvaro López-García and Pablo Orviz},
-keywords = {Linux containers, HPC on cloud, Virtualization, Phenomenology, QCD, Biophysics},
-abstract = {Containers are increasingly used as means to distribute and run Linux services and applications. In this paper we describe the architectural design and implementation of udocker, a tool which enables the user to execute Linux containers in user mode. We also present a few practical applications, using a range of scientific codes characterized by different requirements: from single core execution to MPI parallel execution and execution on GPGPUs.}
-}
-
-@misc{GOMES_2017,
-  author="Jorge Gomes and Isabel Campos",
-  title="Researchers Advance User-Level Container Solution for HPC",
-  journal="HPCwire",
-  year="2017",
-  month="Dec. 18,",
-  url= "https://www.hpcwire.com/2017/12/18/researchers-advance-user-level-container-solution-hpc/"
-}  
-
-@book{CWL2016,
-  title = "Common Workflow Language, v1.0",
-  author = "Brad Chapman and John Chilton and Michael Heuer and Andrey Kartashov and Dan Leehr and Herv{\'e} M{\'e}nager and Maya Nedeljkovich and Matt Scales and Stian Soiland-Reyes and Luka Stojanovic",
-  editor = "Peter Amstutz and Crusoe, {Michael R.} and Neboj{\v s}a Tijani{\'c}",
-  note = "Specification, product of the Common Workflow Language working group. http://www.commonwl.org/v1.0/",
-  year = "2016",
-  month = jul,
-  day = "8",
-  doi = "10.6084/m9.figshare.3115156.v2",
-  language = "English",
[…]
