+++ README.md
@@ -28,8 +28,8 @@ with minimal functionality.
 The full documentation is available at:
 
 * [udocker documentation](https://indigo-dc.github.io/udocker/)
-  * [User manual](https://indigo-dc.github.io/udocker/user_manual.html)
   * [Installation manual](https://indigo-dc.github.io/udocker/installation_manual.html)
+  * [User manual](https://indigo-dc.github.io/udocker/user_manual.html)
   * [Reference card](https://indigo-dc.github.io/udocker/reference_card.html)
 
 ## How does it work
