+++ docs/installation_manual.md
@@ -121,11 +121,6 @@ udocker creates files and subdirectories under `$HOME/.udocker` these are:
 * `containers`: created containers.
 * `keystore.github`: authentication to access repositories (created on demand).
 
-These tools are installed under `$HOME/.udocker` and currently include:
-
-* **P** engine: A modified version of PRoot compiled statically for x86_64 and ARM
-* **F** engine: Modified fakechroot libraries for many Linux distributions on x86_64 and patchelf for x86_64
-* **R** engine: crun and runc compiled statically for x86_64
 
 A default configuration file can be found in directory `udocker/etc/udocker.conf`,
 you can copy it to your home directory `$HOME` or to `$UDOCKER`.
@@ -210,33 +205,61 @@ http_insecure = True
 verbose_level = 5
 ```
 
-## 6. Source code for the tools and libraries
+## 6. External tools and libraries
 
-udocker uses several external tools and libraries to execute the containers. 
+### 6.1. Source code repositories
+
+udocker uses several external tools and libraries to execute the created containers. 
 The source code for the udocker tools and libraries is taken from several repositories.
 The **F** modes need heavily modified Fakechroot libraries and also a modified Patchelf 
 both improved to work with udocker. The **P** modes need a modified PRoot that includes 
 fixes and enhancements to work with udocker. The **R** modes use the original runc and 
 crun software with small changes for static compilation. The following table highlights 
 the repositories used by udocker containing the modified source code and the original 
-repositories together with the software licenses.
+repositories.
 
 
-| Mode  | Engine           | Repository used by udocker                                 | Original repository                      | License 
-|-------|:-----------------|:-----------------------------------------------------------|:-----------------------------------------|:----------------------------------------------------------------------------
-| **P** | PRoot            | <https://github.com/jorge-lip/proot-udocker>               | <https://github.com/proot-me/proot>      | [GPL v2](https://github.com/jorge-lip/proot-udocker/COPYING)
-| **F** | Fakechroot glibc | <https://github.com/jorge-lip/libfakechroot-glibc-udocker> | <https://github.com/dex4er/fakechroot>   | [LGPL v2.1](https://github.com/jorge-lip/libfakechroot-glibc-udocker/LICENSE)
-| **F** | Fakechroot musl  | <https://github.com/jorge-lip/libfakechroot-musl-udocker>  | <https://github.com/dex4er/fakechroot>   | [LGPL v2.1](https://github.com/jorge-lip/libfakechroot-musl-udocker/LICENSE)
-| **F** | Patchelf         | <https://github.com/jorge-lip/patchelf-udocker>            | <https://github.com/NixOS/patchelf>      | [GPL v3](https://github.com/jorge-lip/patchelf-udocker/blob/master/COPYING)
-| **R** | runc             |                                                            | <https://github.com/opencontainers/runc> | [Apache v2.0](https://github.com/opencontainers/runc/blob/master/LICENSE)
-| **R** | crun             |                                                            | <https://github.com/containers/crun>     | [GPL v2](https://github.com/containers/crun/blob/master/COPYING)
+| Mode  | Engine           | Repository used by udocker                                 | Original repository                      
+|-------|:-----------------|:-----------------------------------------------------------|:-----------------------------------------
+| **P** | PRoot            | <https://github.com/jorge-lip/proot-udocker>               | <https://github.com/proot-me/proot>
+| **F** | Fakechroot glibc | <https://github.com/jorge-lip/libfakechroot-glibc-udocker> | <https://github.com/dex4er/fakechroot>
+| **F** | Fakechroot musl  | <https://github.com/jorge-lip/libfakechroot-musl-udocker>  | <https://github.com/dex4er/fakechroot>
[…]
