### 2021-06-15 8e1f756
host specific notes
