### 2021-06-15 0ca74f4
improve section on new developments
