+++ paper/paper.md
@@ -69,57 +69,64 @@ research software for easier execution across these environments.
 udocker is designed to address the requirement of executing scientific 
 applications easily across a wide range of computing systems and digital 
 infrastructures where the user may not have administration privileges, and
-where run times and functionalities to support Linux containers may not be
+where tools and functionalities to support Linux containers may not be
 available. 
 In addition, udocker also simplifies the researcher interaction with the tools 
 required to execute containers by providing an integrated solution to execute 
 Linux containers leveraging different approaches suitable for unprivileged
 users. Finally by executing containers without privileges udocker decreases the 
 risks of privilege escalation. The udocker development started in 2016 and 
-the original udocker paper [@GOMES2018] documented versions 1.1.0 and 1.1.1.
+the original udocker paper [@GOMES2018] documented the initial versions up
+to 1.1.1.
 
 # Concept
 
 udocker provides a self contained solution with minimal dependencies to enable
-execution across systems without need of compilation. udocker itself was 
-initially implemented in Python 2 and later ported to Python 3.
+execution across systems without need of source code compilation. udocker itself 
+was initially implemented in Python 2 and later ported to Python 3.
 
-udocker implements pulling, importing and loading of docker or OCI containers to 
+udocker implements pulling, importing and loading of *docker* or OCI containers to 
 a local repository in the user home directory. The layers composing a container 
 image can then be sequentially extracted to create a flattened directory tree.
 Furthermore udocker also provides the logic to interface with the several 
 execution engines that enable the execution of code extracted from the container 
 images, thus hiding as much as possible the execution engines specificity.
-The execution engines are based on existing software that in most cases has been
+The execution engines are based on existing software that where required has been
 further developed, integrated and packaged to be used with udocker. The following
 engines are currently provided:
 
 * **F** engine: uses the Linux shared library PRELOAD mechanism to intercept
   shared library calls and translate pathnames to provide an unprivileged chroot 
-  like functionality. It is implemented by an extensively enhanced Fakechroot 
-  shared library with versions for the `glibc` [@FAKECHROOT-GLIBC-UDOCKER] 
-  and `musl` [@FAKECHROOT-MUSL-UDOCKER] C standard libraries. 
+  like functionality. It is implemented by an extensively enhanced *Fakechroot*
+  shared library with versions for the *glibc* [@FAKECHROOT-GLIBC-UDOCKER] 
+  and *musl* [@FAKECHROOT-MUSL-UDOCKER] *C* standard libraries. 
   This approach requires the modification of pathnames in the ELF headers of 
   shared libraries and executables. These changes are performed by udocker using 
-  a modified `Patchelf` [@PATCHELF-UDOCKER]. This is the 
+  a modified *Patchelf* [@PATCHELF-UDOCKER]. This is the 
   execution engine that generally provides the highest performance. 
 * **P** engine: uses the Linux PTRACE mechanism to implement a chroot like 
   environment by intercepting system calls and translating pathnames. It is 
-  implemented by a modified PRoot [@PROOT-UDOCKER]. This 
+  implemented by a modified *PRoot* [@PROOT-UDOCKER]. This 
   engine provides the highest interoperability across Linux distributions both 
   older and newer, and constitutes the default execution engine for udocker. 
-* **R** engine: uses either `runc` [@RUNC] or `crun` [@CRUN] to execute the 
+* **R** engine: uses either *runc* [@RUNC] or *crun* [@CRUN] to execute the 
   containers without privileges using Linux user namespaces. Both tools are 
   provided with udocker for wider interoperability.
-* **S** engine: uses Singularity [@KURTZER2017] to execute the containers using 
-  user namespaces or other Singularity execution method when available. 
+* **S** engine: uses *Singularity* [@KURTZER2017] to execute the containers using 
[…]
