+++ paper/paper.bib
@@ -612,3 +612,32 @@ and Zok, T.},
   title={patchelf fork for udocker},
   url="https://github.com/jorge-lip/patchelf-udocker/tree/udocker-1"
 }
+
+@eletronic{JEPL,
+  year={2021},
+  author={Samuel Bernardo, Pablo Orviz},
+  title={A library to implement SQA checks in Jenkins environments},
+  url="https://github.com/indigo-dc/jenkins-pipeline-library"
+}
+
+@eletronic{COLAB,
+  year={2021},
+  author={Google},
+  title={Google Colaboratory},
+  url="https://colab.research.google.com/"
+}
+
+@eletronic{TERMUX,
+  year={2021},
+  author={Termux Developers},
+  title={Termux},
+  url="https://termux.com/"
+}
+
+@eletronic{PYPI,
+  year={2021},
+  author={Python Package Index},
+  title={PyPI},
+  url="https://pypi.org/"
+}
+
