### 2021-06-15 b2553fc
section on developments since 1.1.1
