+++ paper/paper.md
@@ -91,9 +91,9 @@ image can then be sequentially extracted to create a flattened directory tree.
 Furthermore udocker also provides the logic to interface with the several 
 execution engines that enable the execution of code extracted from the container 
 images, thus hiding as much as possible the execution engines specificity.
-The execution engines are based on existing software that where required has been
-further developed, integrated and packaged to be used with udocker. The following
-engines are currently provided:
+The execution engines are based on existing open source software that in several
+cases has been significantly improved, integrated and packaged to be used with 
+udocker. The following engines are currently provided:
 
 * **F** engine: uses the Linux shared library PRELOAD mechanism to intercept
   shared library calls and translate pathnames to provide an unprivileged chroot 
@@ -151,27 +151,28 @@ promoting the adoption of containers in these environments.
 Being a proof of concept the initial versions were not designed for
 production use. Later in the project it become evident that udocker
 had gain adoption beyond its original purpose and scope and that is was 
-already being actively used in production environments. After version 
-1.1.1  and the first udocker paper publication, the effort was directed
+already being actively used in production environments. After the 
+first udocker publication [@GOMES2018] that was produced using versions
+1.1.0 and 1.1.1, the development effort was directed
 to enhance udocker for production use by improving the design, robustness 
 and functionality. Two code branches became supported in parallel, versions 
 1.1.x retaining the original design, and a number of version 1.2.x pre-releases
-aimed at introducing enhancements. The version 1.3.0 marks the point
-of the first production release of the enhanced version.  
+aimed at introducing enhancements. The version 1.3.0 released in June of 
+2021 is the first production release of this enhanced version.  
 
 Since version 1.1.1 the udocker code was reorganized, largely rewritten 
 and improved. Starting with version 1.2.0, udocker supports both Python 3 and 
 Python 2 and was completely restructured moving from being a single large 
 monolithic Python script to become a modular Python application, making 
 maintenance and contributions easier. The new structure has 40 Python
-modules and supports both Python 2.6 or higher and Python 3. Container 
+modules and supports both Python 2.6 or higher and Python 3. Since container 
 technologies are in constant evolution, this new code structure was 
 essential to accommodate any future improvements such as new container 
 formats, APIs and execution engines as they become mainstream. 
 
-The improvements included a more robust command line interface. Several of
-the problems that affected the initial versions in terms of command line
-parsing and validation of arguments among others were addressed.
+The improvements added after 1.1.1 include a more robust command line 
+interface addressing several of the problems that affected the initial 
+versions in terms of command line parsing and validation of arguments.
 The parsing of the configuration files was reimplemented to simplify
 the design and prevent injection of code via the configuration files.
 Configuration is now possible at three levels, system configuration via 
@@ -194,8 +195,8 @@ or *crun* executable for use in **R** execution modes.
 *Singularity* executable for use in **S** execution modes.
 * `UDOCKER_FAKECHROOT_EXPAND_SYMLINKS`: to control the expansion of
 symbolic links in paths pointing to volumes when using the **F** modes. 
-Allows to disable the new path translation algorithms that are much more
-accurate but slower. 
+This variable allows to disable the new path translation algorithm that 
+is now accurate but slower. 
 * `PROOT_TMP_DIR`: is now supported and correctly passed to the *PRoot*
 execution engine.
 
@@ -211,20 +212,20 @@ Currently the external tools and libraries compiled and provided with
 udocker support *x86_64*, *aarch64*, *arm* 32bit and *i386* for use with the 
 **P** modes. The binaries for the remaining execution modes are currently 
 only provided for *x86_64* hosts, this will likely change in the future as 
-other architectures become more popular. 
+these and other architectures become more widely used. 
 
 The **F** mode is particularly specific to udocker. It relies on the interception 
[…]
