### 2021-06-15 0eafb20
updated section on developments since 1.1.1
