+++ paper/paper.bib
@@ -429,6 +429,16 @@ and Zok, T.},
   doi={10.1088/1742-6596/898/8/082039}
 }
 
+@inproceedings{BENEDICIC2019,
+  title={Sarus: Highly Scalable Docker Containers for HPC Systems},
+  author={Benedicic L., Cruz F.A., Madonna A., Mariotti K.},
+  booktitle={ISC High Performance 2019. Lecture Notes in Computer Science},
+  volume={11887},
+  publisher={Springer, Cham.},
+  year={2019},
+  doi={https://doi.org/10.1007/978-3-030-34356-9_5}
+}
+
 @article{CHILLARON2017,
   title={Combining grid computing and docker containers for the study and parametrization of CT image reconstruction methods},
   author={Chillar{\'o}n, M{\'o}nica and Vidal, Vicente and Segrelles, Dami{\'a}n and Blanquer, Ignacio and Verd{\'u}, Gumersindo},
