### 2021-06-16 #3295 @jorge-lip
A more extensive section on the developments since the last publication is now available.
The repositories and licenses for the tools have been added to the installation manual on section 6 as requested.
https://indigo-dc.github.io/udocker/installation_manual.html
