### 2021-06-16 indigo-dc/udocker#315 @jorge-lip
A comparison with Shifter and Sarus was added to the paper. They have different focus/constraints, while udocker main goal is to provide a runtime to execute containers without requiring privileges or installation by the systems administrator, these other tools require privileges and installation by the system administrator. The functionalities udocker are consequently limited to what can be achieved without privileges. On the other hand udocker does not pose the privilege escalation risks.
