### 2021-06-16 indigo-dc/udocker#311 @jorge-lip
The tools used are now listed in tables in section 6 together with the repositories and licenses.
