### 2021-06-16 indigo-dc/udocker#315 @vsoch
That should work - thanks for the addition @jorge-lip !
