### 2021-06-16 indigo-dc/udocker#311 @vsoch
I'm good to resolve this issue - let's wait for @mviereck to comment.
