### 2021-06-16 indigo-dc/udocker#313 @vsoch
I'm also good to close the issue here - @mviereck will wait for your comments!
