### 2021-06-16 087e57b
improve install instructions
