### 2021-06-16 1e8c799
improve install instructions
