+++ SECURITY.md
@@ -1,0 +1,18 @@
+# Security Policy
+
+## Supported Versions
+
+Support is provided for both:
+
+ * the latest 1.1.x tagged version; 
+ * the latest 1.3.x released version.
+
+## Reporting a Vulnerability
+
+Report security issues with detailed information by emailing to udocker@lip.pt 
+
+The udocker maintenance team will acknowledge your email within 72 hours, 
+and will send a more detailed response within the following 48 hours indicating 
+the next steps in handling the vulnerability. You will be kept informed of the 
+progress towards both a complete solution or mitigation and public announcement 
+of the issue. 
