### 2021-06-16 4fa0b89
add section on quality assurance
