+++ docs/user_manual.md
@@ -106,9 +106,9 @@ intervention.
 udocker provides a chroot like environment for container execution. This is
 currently implemented by:
 
-* PRoot engine via the kernel ptrace system call;
-* Fakechroot engine via shared library preload;
-* runc engine using rootless namespaces;
+* PRoot via the kernel ptrace system call;
+* Fakechroot via shared library preload;
+* runc and crun using rootless namespaces;
 * Singularity if available in the host system.
 
 udocker via PRoot offers the emulation of the root user. This emulation
@@ -155,7 +155,7 @@ preferred command name changed from udocker.py to udocker. A symbolic link
 between `udocker` and `maincmd.py` is provided when installing with the
 distribution tarball.
 
-```bash
+```
 udocker [GLOBAL-PARAMETERS] COMMAND [COMMAND-OPTIONS] [COMMAND-ARGUMENTS]
 ```
 
@@ -192,7 +192,7 @@ udocker COMMAND --help
 
 ### 3.3. install
 
-```bash
+```
 udocker install [OPTIONS]
 ```
 
@@ -215,7 +215,7 @@ udocker install --force --purge
 
 ### 3.4. search
 
-```bash
+```
 udocker search [-a] STRING
 udocker search --list-tags REPO/IMAGE
 ```
@@ -244,7 +244,7 @@ udocker search --list-tags centos
 
 ### 3.5. pull
 
-```bash
+```
 udocker pull [OPTIONS] REPO/IMAGE:TAG
 ```
 
@@ -277,7 +277,7 @@ udocker pull --httpproxy=socks5h://user:pass@host:port busybox
 
 ### 3.6. images
 
-```bash
+```
 udocker images [OPTIONS]
 ```
 
@@ -297,7 +297,7 @@ udocker images -l
 
 ### 3.7. create
 
-```bash
+```
 udocker create [OPTIONS] REPO/IMAGE:TAG
 ```
 
@@ -320,7 +320,7 @@ udocker create --name=mycontainer indigodatacloud/disvis:latest
 
 ### 3.8. ps
 
-```bash
+```
 udocker ps [options]
 ```
 
@@ -333,8 +333,8 @@ The command displays:
 * whether the container tree is writable (is in a R/W location)
 * the easier to remember name(s)
 * the name of the container image from which it was extracted
-* with option -m adds the execution mode
-* with option -s adds the container current size in MB
+* with option `-m` adds the execution mode
+* with option `-s` adds the container current size in MB
 
 Options:
 
@@ -364,7 +364,7 @@ Options:
 
 Examples:
 
-```
+```bash
 udocker rmi -f indigodatacloud/ambertools\_app:latest
 ```
 
@@ -384,7 +384,7 @@ Options:
 
 Examples:
 
-```
+```bash
 udocker rm 7b2d4456-9ee7-3138-ad01-63d1342d8545
 udocker rm mycontainer
 ```
@@ -406,7 +406,7 @@ Options:
 
 Examples:
 
-```
+```bash
 udocker inspect ubuntu:latest
 udocker inspect d2578feb-acfc-37e0-8561-47335f85e46d
 udocker inspect -p d2578feb-acfc-37e0-8561-47335f85e46d
@@ -423,7 +423,7 @@ This is an alternative to the use of `create --name=`
 
 Examples:
 
-```
+```bash
 udocker name d2578feb-acfc-37e0-8561-47335f85e46d BLUE
 ```
 
@@ -438,7 +438,7 @@ Remove a name previously given to an extracted container with
 
 Examples:
 
-```
+```bash
 udocker rmname BLUE
 ```
 
@@ -467,7 +467,7 @@ Performs sanity checks to verify a image available in the local repository.
 
[…]
