### 2021-06-17 indigo-dc/udocker#311 @mviereck
Yes, this looks way better now, thank you!
One remaining point:
 -  In "2.1. Install from a released version", please adjust the sentence "Complete the installation by invoking udocker install to download and install the tarball containing the tools and libraries required to execute containers:".
    - A note at this point that `udocker install` installs third party tools.
    - A link down to chapter "6. External tools and libraries".
