### 2021-06-17 indigo-dc/udocker#313 @mviereck
My only point here was the question python2 vs. python3. This is clarified now, we focus on the python3 version that is the new default in the repository.

So far I did not check closer the other points mentioned by @vsoch.
If I find something related to "Getting started" or "Examples", I'll just open a new ticket.

I think we can close here.
