### 2021-06-17 indigo-dc/udocker#312 @mviereck
@jorge-lip Thank you for rewriting the chapter about the differences to version 1.1.
For now I am a bit overwhelmed by the length of the text. Will read it again.

Can you describe the difference in a nutshell, in a few sentences that nail down to the point? Less going into several technical details, but rather from a meta perspective?
