### 2021-06-17 3378d7f
add fix against removal of .udocker dir in tests
