### 2021-06-17 bde3888
change to branch master in config.yml remove deprecated functional and integration tests add information on running tests
