+++ tests/functional/test_functional.py
@@ -1,236 +1,0 @@
-#!/usr/bin/env python2
-"""
-========================
-udocker functional tests
-========================
-Unit tests for udocker, a wrapper to execute basic docker containers
-without using docker.
-
-Licensed under the Apache License, Version 2.0 (the "License");
-you may not use this file except in compliance with the License.
-You may obtain a copy of the License at
-
-http://www.apache.org/licenses/LICENSE-2.0
-
-Unless required by applicable law or agreed to in writing, software
-distributed under the License is distributed on an "AS IS" BASIS,
-WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-See the License for the specific language governing permissions and
-limitations under the License.
-"""
-
-import os
-import sys
-import unittest
-import mock
-
-sys.path.append(".")
-
-import udocker
-from udocker.config import Config
-
-
-def set_env():
-    """Set environment variables"""
-    if not os.getenv("HOME"):
-        os.environ["HOME"] = os.getcwd()
-
-
-def find_str(self, find_exp, where):
-    """Find string in test ouput messages"""
-    found = False
-    for item in where:
-        if find_exp in str(item):
-            self.assertTrue(True)
-            found = True
-            break
-    if not found:
-        self.assertTrue(False)
-
-
-def set_msglevels(mock_msg):
-    """Set mock Msg() class message levels"""
-    mock_msg.ERR = 0
-    mock_msg.MSG = 1
-    mock_msg.WAR = 2
-    mock_msg.INF = 3
-    mock_msg.VER = 4
-    mock_msg.DBG = 5
-    mock_msg.DEF = mock_msg.INF
-
-
-class MainTestCase(unittest.TestCase):
-    """Test most udocker capabilities"""
-
-    @classmethod
-    def setUpClass(cls):
-        """Setup test"""
-        set_env()
-
-    @mock.patch('sys.exit')
-    @mock.patch('udocker.container.localrepo.LocalRepository')
-    @mock.patch('udocker.tools.UdockerTools')
-    @mock.patch('udocker.config.Config.user_init')
-    @mock.patch('udocker.UdockerCLI.do_images')
-    @mock.patch('udocker.msg.Msg')
-    def test_init(self, mock_msg, mock_images,
-                  mock_user_init, mock_utools, mock_localrepo, mock_exit):
-        """Test udocker global command line options"""
-        set_msglevels(mock_msg)
-        t_argv = ['./udocker.py', "images"]
-        with mock.patch.object(sys, 'argv', t_argv):
-            udocker.Main().start()
-            self.assertTrue(mock_images.called)
-        t_argv = ['./udocker.py', "--config=/myconf", "images"]
-        with mock.patch.object(sys, 'argv', t_argv):
-            udocker.Main().start()
-            self.assertTrue(mock_user_init.called_with("/myconf"))
-        t_argv = ['./udocker.py', "--cofig=/myconf"]
-        with mock.patch.object(sys, 'argv', t_argv):
-            udocker.Main().start()
-            self.assertTrue(mock_exit.called)
-        Config().verbose_level = 0
-        t_argv = ['./udocker.py', "-D", "images"]
-        with mock.patch.object(sys, 'argv', t_argv):
-            udocker.Main().start()
-            self.assertEqual(Config().verbose_level, mock_msg.DBG)
-        Config().verbose_level = 0
-        t_argv = ['./udocker.py', "--debug"]
-        with mock.patch.object(sys, 'argv', t_argv):
-            udocker.Main().start()
-            self.assertEqual(Config().verbose_level, mock_msg.DBG)
-        Config().verbose_level = 0
-        t_argv = ['./udocker.py', "-q"]
-        with mock.patch.object(sys, 'argv', t_argv):
[…]
