### 2021-06-17 a478237
update text in quality assurance
