### 2021-06-17 aea3719
Small updates to testing instructions
Also, should bandit be included in the dev requirements.txt or have some other instruction to install?
