### 2021-06-17 #3295 @vsoch
I'm good! My concerns with respect to the checklist were:

 - no information about automated tests
 - missing a proper discussion about other container technologies / comparisons
 - installation instructions

And the authors have resolved these issues so my checklist is complete.
