### 2021-06-17 4b91d41
add reference to tools and libraries
