### 2021-06-17 1845da2
uninstall instructions
