### 2021-06-17 dfb472c
Fixes for the tests section
