### 2021-06-17 indigo-dc/udocker#311 @jorge-lip
Thanks, these fixes have been added to the installation manual.
