### 2021-06-17 indigo-dc/udocker#311 @mviereck
Thank you! Now I am entirely fine with installation instructions and dependency declaration.
