### 2021-06-17 indigo-dc/udocker#335 @mviereck
[joss review] possible license conflict?
udocker is distributed under Apache License 2.0.
The third party tools have more restrictive licenses, notable GPLv2 and GPLv3.

I am not familiar with details in deep of the licenses.
But AFAIK because udocker hardly depends on the third party tools, it can only be used in the most restrictive combination of all licenses, likely GPLv3.

I point to this because users might use udocker or redistribute e.g. in a commercial way that might be forbidden by the other licenses.

I am not sure how to resolve this, just want to point on a possible issue.

https://github.com/openjournals/joss-reviews/issues/3295
