### 2021-06-18 indigo-dc/udocker#335 @mariojmdavid
please suggest what we should do, but keep in mind
there where/are several people contributing to the SW and have done so under the current license
and there are currently 95 forks
