### 2021-06-18 indigo-dc/udocker#335 @jorge-lip
The Apache License 2.0 is for udocker only, the other tools retain their original licenses and we make the source code changes public. What is done by udocker is not different than a Linux distro does by distributing multiple software with multiple licenses. The problem with mixed licenses arises typically when a single indivisible object incorporates code with multiple licenses. That is not the case of udocker itself that is a Python script and is not linked against those other codes.
