### 2021-06-18 indigo-dc/udocker#335 @mviereck
> The Apache License 2.0 is for udocker only, the other tools retain their original licenses and we make the source code changes public. What is done by udocker is not different than a Linux distro does by distributing multiple software with multiple licenses. The problem with mixed licenses arises typically when a single indivisible object incorporates code with multiple licenses. That is not the case of udocker itself that is a Python script and is not linked against those other codes.

Yes, I agree.

> please suggest what we should do

As I've said, I am not sure myself.
Basically a user must be aware which licenses he must regard if going further than personal use, especially in case of redistribution. This is one reason why I've insisted on a list of licenses and origin of `udocker-tools`.

A quick solution would be a hint like:
_"Redistribution and commercial use must regard all licenses shipped with udocker`."_
This is formally correct and enough, but leaves the user to figure out the consequences himself.

It would help the user to give some hints on the consequences. Though, I can't tell the consequences myself, I am not an expert in interfering software licenses. (Is _anyone_ an expert in this?)
