### 2021-06-18 #3295 @mviereck
The installtion instructions and dependency declaration has been rewritten and clearly state the third party tools now including their source, license and origin. This is done well and no longer an acceptance blocker.

Now I continue to work on the checklist.
