+++ docs/installation_manual.md
@@ -265,8 +265,9 @@ containing the modified source code and the original repositories.
 
 ### 6.2. Software Licenses
 
-Redistribution, commercial use and code changes must regard all licenses 
-shipped with udocker. 
+Redistribution, commercial use and code changes must regard all licenses
+shipped with udocker. These include the udocker license and the individual
+licences of the external tools and libraries packaged for use with udocker.
 
 | Mode  | Engine           | License 
 |-------|:-----------------|:----------------------------------------------------------------------------
