### 2021-06-18 indigo-dc/udocker#335 @jorge-lip
Agree I have added the suggested statement to the README, to the introduction of the installation manual and also in the licenses section of the installation manual. The full statement is:

Redistribution, commercial use and code changes must regard all licenses shipped with udocker. These include the [udocker license](LICENSE) and the individual licences of the external tools and libraries packaged for use
with udocker. For further information see the [software licenses section](https://indigo-dc.github.io/udocker/installation_manual.html#62-software-licenses) of the installation manual.
