### 2021-06-18 10b6545
licenses notice
