+++ tests/unit/test_commonlocalfile.py
@@ -3,15 +3,9 @@
 udocker unit tests: CommonLocalFileApi
 """
 
-import sys
 from unittest import TestCase, main
+from unittest.mock import Mock, patch
 from udocker.commonlocalfile import CommonLocalFileApi
-try:
-    from unittest.mock import Mock, MagicMock, patch, mock_open
-except ImportError:
-    from mock import Mock, MagicMock, patch, mock_open
-
-sys.path.append('.')
 
 
 class CommonLocalFileApiTestCase(TestCase):
