### 2021-06-18 0a2711e
list failed tests upon exit
