+++ utils/udocker_test.sh
@@ -31,8 +31,9 @@ FAIL_STR="${RED}[FAIL]${NC}"
 THIS_SCRIPT_NAME=$( basename "$0" )
 
 # Variables for the tests
+declare -a FAILED_TESTS
 DEFAULT_UDIR=$HOME/.udocker
-TEST_UDIR=$HOME/.udockermy
+TEST_UDIR=$HOME/.udocker-test-h45y7k9X
 TAR_IMAGE="centos7.tar"
 TAR_CONT="centos7-cont.tar"
 TAR_IMAGE_URL="https://download.ncg.ingrid.pt/webdav/udocker_test/${TAR_IMAGE}"
@@ -66,6 +67,7 @@ function result
       print_ok; echo "    $STRING"
   else
       print_fail; echo "    $STRING"
+      FAILED_TESTS+=("$STRING")
   fi
   echo "------------------------------------------------------------>"
 }
@@ -77,6 +79,7 @@ function result_inv
       print_ok; echo "    $STRING"
   else
       print_fail; echo "    $STRING"
+      FAILED_TESTS+=("$STRING")
   fi
   echo "------------------------------------------------------------>"
 }
@@ -116,150 +119,183 @@ STRING="T007: udocker version"
 udocker version; return=$?
 result
 
-STRING="T008: udocker -V"
+STRING="T008: udocker -D version"
+udocker -D version; return=$?
+result
+
+STRING="T009: udocker --quiet version"
+udocker --quiet version; return=$?
+result
+
+STRING="T010: udocker -q version"
+udocker -q version; return=$?
+result
+
+STRING="T011: udocker --debug version"
+udocker --debug version; return=$?
+result
+
+STRING="T012: udocker -V"
 udocker -V >/dev/null 2>&1 ; return=$?
 result
 
-STRING="T009: udocker --version"
+STRING="T013: udocker --version"
 udocker --version >/dev/null 2>&1 ; return=$?
 result
 
-STRING="T010: udocker search -a"
+STRING="T014: udocker search -a"
 udocker search -a gromacs | grep ^gromacs; return=$?
 result
 
-STRING="T011: udocker pull ${DOCKER_IMG}"
+STRING="T015: udocker pull ${DOCKER_IMG}"
 udocker pull ${DOCKER_IMG}; return=$?
 result
 
-STRING="T012: udocker verify ${DOCKER_IMG}"
+STRING="T016: udocker --insecure pull ${DOCKER_IMG}"
+udocker --insecure pull ${DOCKER_IMG}; return=$?
+result
+
+STRING="T017: udocker verify ${DOCKER_IMG}"
 udocker verify ${DOCKER_IMG}; return=$?
 result
 ## TODO: Add test to check layers after pull
 
-STRING="T013: udocker images"
+STRING="T018: udocker images"
 udocker images; return=$?
 result
 
-STRING="T014: udocker inspect (image)"
+STRING="T019: udocker inspect (image)"
 udocker inspect ${DOCKER_IMG}; return=$?
 result
 
-STRING="T015: udocker create ${DOCKER_IMG}"
+STRING="T020: udocker create ${DOCKER_IMG}"
 CONT_ID=`udocker create ${DOCKER_IMG}`; return=$?
 echo "ContainerID = ${CONT_ID}"
 result
 
-STRING="T016: udocker create --name=${CONT} ${DOCKER_IMG}"
+STRING="T021: udocker create --name=${CONT} ${DOCKER_IMG}"
 CONT_ID_NAME=`udocker create --name=${CONT} ${DOCKER_IMG}`; return=$?
 result
 
-STRING="T017: udocker ps"
[…]
