### 2021-06-18 b8999ec
list failed tests upon exit
