+++ tests/unit/test_umain.py
@@ -24,13 +24,16 @@ class UMainTestCase(TestCase):
         udoc = UMain(argv)
         self.assertEqual(udoc.argv, argv)
 
+    @patch('udocker.cli.Msg')
     @patch('udocker.umain.UdockerCLI')
     @patch('udocker.umain.LocalRepository')
     @patch('udocker.umain.os.geteuid')
     def test_02__prepare_exec(self, mock_getuid,
-                              mock_local, mock_ucli):
+                              mock_local, mock_ucli, mock_msg):
         """Test02 UMain()._prepare_exec()."""
         argv = ["udocker", "-h"]
+        mock_msg.level = 0
+        mock_msg.VER = 4
         mock_getuid.return_value = 0
         with patch('sys.exit') as mock_exit:
             umain = UMain(argv)
@@ -38,6 +41,8 @@ class UMainTestCase(TestCase):
             self.assertTrue(mock_exit.called)
 
         argv = ["udocker", "-h", "--debug", "--insecure"]
+        mock_msg.level = 0
+        mock_msg.VER = 4
         mock_getuid.return_value = 100
         mock_local.return_value.is_repo.return_value = True
         mock_local.return_value.create_repo.return_value = None
