+++ tests/unit/test_umain.py
@@ -24,7 +24,7 @@ class UMainTestCase(TestCase):
         udoc = UMain(argv)
         self.assertEqual(udoc.argv, argv)
 
-    @patch('udocker.cli.Msg')
+    @patch('udocker.umain.Msg')
     @patch('udocker.umain.UdockerCLI')
     @patch('udocker.umain.LocalRepository')
     @patch('udocker.umain.os.geteuid')
@@ -53,29 +53,50 @@ class UMainTestCase(TestCase):
         self.assertTrue(mock_local.return_value.is_repo.called)
         self.assertTrue(mock_ucli.called)
 
-    # @patch('udocker.umain.Msg')
-    # @patch('udocker.umain.UdockerCLI')
-    # @patch.object(UMain, '_prepare_exec')
-    # def test_03_execute(self, mock_prep, mock_ucli, mock_msg):
-    #     """Test03 UMain().execute()."""
-    #     mock_msg.level = 0
-    #     argv = ["udocker", "-h"]
-    #     mock_prep.return_value = None
-    #     mock_ucli.return_value.do_help.return_value = 0
-    #     um = UMain(argv)
-    #     status = um.execute()
-    #     self.assertTrue(mock_prep.called)
-    #     self.assertTrue(mock_ucli.return_value.do_help.called)
-    #     self.assertEqual(status, 0)
+    @patch('udocker.umain.Msg')
+    @patch('udocker.umain.UdockerCLI')
+    def test_03_execute(self, mock_ucli, mock_msg):
+        """Test03 UMain().execute()."""
+        mock_msg.level = 0
+        argv = ["udocker", "-h"]
+        mock_ucli.return_value.do_help.return_value = 0
+        umain = UMain(argv)
+        status = umain.execute()
+        self.assertTrue(mock_ucli.return_value.do_help.called)
+        self.assertEqual(status, 0)
 
-    #     argv = ['udocker', '--version']
-    #     mock_prep.return_value = None
-    #     mock_ucli.return_value.do_help.return_value = 0
-    #     mock_ucli.return_value.do_version.return_value = 0
-    #     um = UMain(argv)
-    #     um.cli = mock_ucli
-    #     status = um.execute()
-    #     self.assertTrue(mock_ucli.do_version.called)
+        argv = ['udocker', '--version']
+        mock_ucli.return_value.do_version.return_value = 0
+        umain = UMain(argv)
+        status = umain.execute()
+        self.assertTrue(mock_ucli.return_value.do_version.called)
+        self.assertEqual(status, 0)
+
+        argv = ['udocker', 'install']
+        mock_ucli.return_value.do_install.return_value = 0
+        umain = UMain(argv)
+        status = umain.execute()
+        self.assertTrue(mock_ucli.return_value.do_install.called)
+        self.assertEqual(status, 0)
+
+        argv = ['udocker', 'showconf']
+        mock_ucli.return_value.do_showconf.return_value = 0
+        umain = UMain(argv)
+        status = umain.execute()
+        self.assertTrue(mock_ucli.return_value.do_showconf.called)
+        self.assertEqual(status, 0)
+
+        argv = ['udocker', 'rm']
+        mock_ucli.return_value.do_rm.return_value = 0
+        umain = UMain(argv)
+        status = umain.execute()
+        self.assertTrue(mock_ucli.return_value.do_rm.called)
+        self.assertEqual(status, 0)
+
+        argv = ['udocker', 'faking']
+        umain = UMain(argv)
+        status = umain.execute()
+        self.assertEqual(status, 1)
 
 
 if __name__ == '__main__':
