+++ tests/unit/test_executionmode.py
@@ -73,6 +73,7 @@ class ExecutionModeTestCase(TestCase):
         status = uexm.get_mode()
         self.assertEqual(status, "F3")
 
+    @patch('udocker.engine.execmode.Msg')
     @patch('udocker.engine.execmode.os.path')
     @patch('udocker.engine.execmode.FileUtil.putdata')
     @patch('udocker.engine.execmode.FileBind.setup')
@@ -85,11 +86,11 @@ class ExecutionModeTestCase(TestCase):
     @patch('udocker.engine.execmode.ElfPatcher')
     @patch.object(ExecutionMode, 'get_mode')
     def test_03_set_mode(self, mock_getmode, mock_elfp, mock_fbind,
-                         mock_futil, mock_getdata, mock_restore,
-                         mock_links, mock_getld,
-                         mock_fbset, mock_putdata,
-                         mock_path):
+                         mock_futil, mock_getdata, mock_restore, mock_links,
+                         mock_getld, mock_fbset, mock_putdata, mock_path,
+                         mock_msg):
         """Test03 ExecutionMode().set_mode."""
+        mock_msg.level = 0
         mock_getmode.side_effect = \
             ["", "P1", "R1", "R1", "F4", "P1", "F3", "P2", "P2", "F4", "F4"]
         mock_getdata.return_value = "F3"
