+++ tests/unit/test_dockerioapi.py
@@ -523,12 +523,13 @@ class DockerIoAPITestCase(TestCase):
         out = doia.get_v2_layers_all(imagerepo, fslayers)
         self.assertEqual(out, ['foolayername'])
 
+    @patch('udocker.docker.Msg')
     @patch('udocker.docker.GetURL.get_status_code')
     @patch.object(DockerIoAPI, 'get_v2_image_manifest')
     @patch.object(DockerIoAPI, 'get_v2_layers_all')
     @patch.object(DockerIoAPI, '_get_url')
     def test_28_get_v2(self, mock_dgu, mock_dgv2, mock_manif,
-                       mock_getstatus):
+                       mock_getstatus, mock_msg):
         """Test28 DockerIoAPI().get_v2"""
         imgrepo = "img1"
         hdr = type('test', (object,), {})()
@@ -538,6 +539,7 @@ class DockerIoAPITestCase(TestCase):
         hdr.data = hdr_data
         buff = strio()
         manifest = list()
+        mock_msg.level = 0
         mock_dgu.return_value = (hdr, buff)
         mock_manif.return_value = (hdr_data, manifest)
         mock_dgv2.return_value = []
@@ -603,6 +605,7 @@ class DockerIoAPITestCase(TestCase):
         out = doia._get_v1_id_from_images(imgarr, shortid)
         self.assertEqual(out, "1234567890")
 
+    @patch('udocker.docker.Msg')
     @patch('udocker.docker.GetURL.get_status_code')
     @patch.object(DockerIoAPI, 'get_v1_layers_all')
     @patch.object(DockerIoAPI, 'get_v1_image_ancestry')
@@ -612,7 +615,7 @@ class DockerIoAPITestCase(TestCase):
     @patch.object(DockerIoAPI, 'get_v1_repo')
     def test_31_get_v1(self, mock_dgv1repo, mock_v1imgtag,
                        mock_v1idtag, mock_v1idimg, mock_v1ances,
-                       mock_v1layer, mock_status):
+                       mock_v1layer, mock_status, mock_msg):
         """Test31 DockerIoAPI().get_v1"""
         imgarr = [{"id": "1234567890"}]
         imagerepo = "REPO"
@@ -623,6 +626,7 @@ class DockerIoAPITestCase(TestCase):
                     "X-ND-CURLSTATUS": 0,
                     "x-docker-endpoints": "https://registry-1.docker.io"}
         hdr.data = hdr_data
+        mock_msg.level = 0
         mock_dgv1repo.return_value = (hdr_data, imgarr)
         mock_status.return_value = 401
         doia = DockerIoAPI(self.local)
