### 2021-06-21 597a5d8
refactor unit test cli
