### 2021-06-23 58e08e6
remove commented code, update README, bump version
