### 2021-06-23 b2f5efb
add --entrypoint to run --help
