+++ docs/installation_manual.md
@@ -1,13 +1,13 @@
 # Installation and configuration manual
 
 In most cases the end user can download and execute udocker without
-system administrator intervention. udocker itself is written in Python, but 
+system administrator intervention. udocker itself is written in Python, but
 also uses external binaries and libraries to provide a chroot like
 environment where containers are executed in user space. These tools do not
-require any privileges and constitute the udocker tools and libraries for 
+require any privileges and constitute the udocker tools and libraries for
 engines that is downloaded and installed by udocker itself.
 
-Redistribution, commercial use and code changes must regard all licenses 
+Redistribution, commercial use and code changes must regard all licenses
 shipped with udocker. For more information see
 [section 6 External tools and libraries](#6-external-tools-and-libraries).
 
@@ -15,6 +15,7 @@ shipped with udocker. For more information see
 
 The udocker dependencies are minimal and should be supported by most Linux installations.
 udocker requires:
+
 * Python 3 or alternatively Python >= 2.6
 * pycurl or alternatively the curl command
 * python hashlib or alternatively the openssl command
@@ -28,30 +29,30 @@ udocker requires:
 
 ### 2.1. Install from a released version
 
-Download a release tarball from <https://github.com/indigo-dc/udocker/releases>: 
+Download a release tarball from <https://github.com/indigo-dc/udocker/releases>:
 
 ```bash
-wget https://github.com/indigo-dc/udocker/releases/download/v1.3.0/udocker-1.3.0.tar.gz
-tar zxvf udocker-1.3.0.tar.gz
+wget https://github.com/indigo-dc/udocker/releases/download/v1.3.1/udocker-1.3.1.tar.gz
+tar zxvf udocker-1.3.1.tar.gz
 export PATH=`pwd`/udocker:$PATH
 ```
 
 Alternatively use `curl` instead of `wget` as follows:
 
 ```bash
-curl -L https://github.com/indigo-dc/udocker/releases/download/v1.3.0/udocker-1.3.0.tar.gz \
-  > udocker-1.3.0.tar.gz
-tar zxvf udocker-1.3.0.tar.gz
+curl -L https://github.com/indigo-dc/udocker/releases/download/v1.3.1/udocker-1.3.1.tar.gz \
+  > udocker-1.3.1.tar.gz
+tar zxvf udocker-1.3.1.tar.gz
 export PATH=`pwd`/udocker:$PATH
 ```
 
 udocker executes containers using external tools and libraries that
-are enhanced and packaged for use with udocker. For more information see 
+are enhanced and packaged for use with udocker. For more information see
 [section 6 External tools and libraries](#6-external-tools-and-libraries).
 Therefore to complete the installation invoke `udocker install` to download
-and install the required tools and libraries. 
+and install the required tools and libraries.
 
-```
+```bash
 udocker install
 ```
 
@@ -74,10 +75,10 @@ export PATH=`pwd`/udocker/udocker:$PATH
 ```
 
 udocker executes containers using external tools and libraries that
-are enhanced and packaged for use with udocker. For more information see 
+are enhanced and packaged for use with udocker. For more information see
 [section 6 External tools and libraries](#6-external-tools-and-libraries).
 Therefore to complete the installation invoke `udocker install` to download
-and install the required tools and libraries. 
+and install the required tools and libraries.
 
 ```bash
 udocker install
@@ -86,32 +87,31 @@ udocker install
 ### 2.3. Install from PyPI using pip
 
 For installation with pip it is advisable to setup a Python3 virtual environment
-before proceeding, see 
+before proceeding, see
[…]
