+++ udocker/config.py
@@ -262,8 +262,8 @@ class Config(object):
 
     def getconf(self, user_cfile="u.conf"):
         """Return all configuration variables"""
-        self._file_override(user_cfile)   # Override with variables in conf file
-        self._env_override()              # Override with variables in environment
+        self._file_override(user_cfile) # Override with variables in conf file
+        self._env_override()          # Override with variables in environment
 
     def container(self, user_cfile="u.conf"):
         """
