### 2021-06-24 928c38b
improve quality assurance section
