### 2021-06-29 #3295 @mviereck
Sorry that I've been away that long. I was distracted by RL.
Most remaining points I just have to check without expecting a problem.

One major point was a missing description of the difference to the previous paper.
Now there is a long chapter on this. It seems to me that this fits in the way you've said:
> in the same way that a research project can lead to multiple papers, some of which can be incremental progress in an area that was previously published.

Overall the paper is quite long and detailed. A bit I am missing is to bring it to a short and clear point. I asked for this in https://github.com/indigo-dc/udocker/issues/312 but got no response.

A bit I am lost in getting the full content of the paper just because it is a bit hard to me to grok it in a non-native language.

To bring it to a point: The paper is rather written like a full documentation than like a core desciption that comes to a point. This can be weighted good or bad. A scientific paper on research should show a lot of detail. But I'd say a paper on software should tell less detail but should rather show the essential core.
