### 2021-06-30 indigo-dc/udocker#312 @mviereck
Could you please answer these questions?

> If someone writes a scientific paper and wants to cite udocker: at which point he would have the urgent need to cite the new paper instead of the previous one?

Similar, but not the same:

> Can you describe the difference in a nutshell, in a few sentences that nail down to the point? Less going into several technical details, but rather from a meta perspective?
