### 2021-06-30 #3295 @jorge-lip
@danielskatz Along these weeks we have put a lot of effort to address the recommendations and requests both in terms of the software itself and also in terms of improving the paper that indeed now starts to become very close to a full paper. At this point all the material and detailed information to decide if this work is suitable for publication is available. We are happy to perform further improvements if needed but first we need to know if the work is to be published. I wish to thank the reviewers for their positive suggestions and contributions to improve udocker.
