### 2021-06-30 #3295 @danielskatz
@jorge-lip - the expectation for JOSS papers is that they will be improved via collaboration between the reviewers and the authors to the point where they can be published, unless there is an issue of scope, which I don't think it the case here.

On the other hand, a typical paper is fairly short, with most of the description of the software in the documentation (README, etc.), not in the paper. I think we have gotten away from that goal, but perhaps this is ok for this software.
