### 2021-06-30 #3295 @jorge-lip
Many thanks for the clarifications, therefore assuming that the paper is in scope and since all software and documentation issues have been addressed, just let us know in this thread if there are additional points that you feel must still be corrected in the paper.
