### 2021-06-30 indigo-dc/udocker#312 @danielskatz
ref: https://github.com/openjournals/joss-reviews/issues/3295
