### 2021-06-30 #3295 @danielskatz
👋 @mviereck - To try to wrap this up, can you be specific about the changes you want to see to finish checking off the items in your review list?  Perhaps as bullets or as new issues in the source repo?

Of course, this is in addition to https://github.com/indigo-dc/udocker/issues/312, where it looks like @jorge-lip still needs to respond.
