### 2021-06-30 #3295 @mviereck
@danielskatz  Basically this one question brings it to the point:
>  If someone writes a scientific paper and wants to cite udocker: at which point he would have the urgent need to cite the new paper instead of the previous one?

As long as this cannot be answered, there is no obvious need for a new paper and I cannot check of "substantial scholary effort".

Other remaining points in the checklist I just  have to work through. If I find issues there, I'll open new tickets.
