### 2021-06-30 #3295 @jorge-lip
> If someone writes a scientific paper and wants to cite udocker: at which point he would have the urgent need to cite the new paper instead of the previous one?

From our point of view yes version 1.3.0 is largely different from 1.1.0. These differences and new developments are described in the paper in section "Developments since 1.1.1".
