### 2021-06-30 #3295 @danielskatz
>> If someone writes a scientific paper and wants to cite udocker: at which point he would have the urgent need to cite the new paper instead of the previous one?

> As long as this cannot be answered, there is no obvious need for a new paper and I cannot check of "substantial scholary effort".

Please note that the [discussion about this in the JOSS documentation](https://joss.readthedocs.io/en/latest/submitting.html#substantial-scholarly-effort) says:

> As a rule of thumb, JOSS’ minimum allowable contribution should represent not less than three months of work for an individual. Some factors that may be considered by editors and reviewers when judging effort include:
>
> * Age of software (is this a well-established software project) / length of commit history.
> * Number of commits.
> * Number of authors.
> * Total lines of code (LOC). Submissions under 1000 LOC will usually be flagged, those under 300 LOC will be desk rejected.
> * Whether the software has already been cited in academic papers.
> * Whether the software is sufficiently useful that it is likely to be cited by your peer group.
>
> In addition, JOSS requires that software should be feature-complete (i.e. no half-baked solutions) and designed for maintainable extension (not one-off modifications of existing tools). “Minor utility” packages, including “thin” API clients, and single-function packages are not acceptable.

Along with the comment from @jorge-lip in the comment above this one.  I think the work since the previous paper meets these factors, but if you don't, let's talk about which factors you think are the problem. (And note that this is not a case where all factors must be met - these are just samples of factors that could be considered.)
