### 2021-07-06 #3295 @danielskatz
👋 @mviereck - How are things going with the remaining items on your checklist?
