### 2021-07-07 #3295 @mviereck
I was still hoping for an answer in https://github.com/indigo-dc/udocker/issues/312 .
Yet I am working on a comparision to answer the questions myself.
