### 2021-07-07 6f8b7d5
implement new commands, start move to logging
