### 2021-07-07 indigo-dc/udocker#312 @jorge-lip
Since 1.1.1, udocker moved from a proof of concept tool suitable for production. The focus has been put on sustainability and interoperability. The code structure was redesigned moving from the Python 2 monolithic code to a new Python 3 modular implementation. New features and improvements were added to support a wider range of Linux distributions and environments. To this end the execution engines now support `crun` for namespace based execution with `cgroups v2` and a port of `Fakechroot` to support `MUSL libc`. The existing execution engines were also enhanced to address the shortcomings present in the proof of concept and improve interoperability for execution across a wider range of distributions, kernels and better supporting the functionalities required for containers execution.  Support for multiple architectures was added and binaries for ARM and x86 are now provided for `PRoot` based execution, further binaries can be easily added as needed. Support for accelerated computing was implemented by introducing the capability to setup containers to use NVIDIA GPUs according to the host environment and across the several execution engines. The handling of container images was improved both at the level of interaction with container registries and handling of formats. The interaction with registries is now more interoperable and robust and support for the OCI image specification v1 was introduced. Furthermore the search capabilities support docker registry v1 and v2 specifications and listing of tags. The command line interface and configuration was improved with new commands, flags and better parsing of both the command line and config files. Finally the installation of the tools and libraries was decoupled from udocker itself, both can now be released separately. The shared installation of udocker in distributed filesystems was addressed with changes introduced to handle the cases where udocker is executed from readonly filesystems. Finally, udocker is now also available from *PyPI*. A more complete description is provided in the paper.
