### 2021-07-09 8db394d
add command show metadata
