### 2021-07-12 indigo-dc/udocker#348 @mviereck
[joss review] list of contributors
Your paper lists the following contributors:
 - Jorge Gomes
 - Mario David
 - João Martins
 - João Pina
 - Samuel Bernardo
 - Isabel Campos
 - Pablo Orviz
 - Alvaro López-García

The [github contributors list](https://github.com/indigo-dc/udocker/graphs/contributors) shows that most contributions are done by Jorge Gomes and Mario David. Other contributions are minor (mostly a few lines).

What is the contribution of the others mentioned in the paper?
Why don't they appear as contributors on github?

https://github.com/openjournals/joss-reviews/issues/3295
