### 2021-07-13 1fe8a7c
implement show_metadata
