+++ udocker/tools.py
@@ -7,6 +7,7 @@ import tarfile
 import random
 import json
 import stat
+import logging
 
 from udocker import is_genstr
 from udocker.config import Config
@@ -14,6 +15,9 @@ from udocker.utils.curl import GetURL
 from udocker.utils.fileutil import FileUtil
 from udocker.msg import Msg
 from udocker import __version__
+
+LOG = logging.getLogger(__name__)
+MSG_LEVEL = 60
 
 def _str(data):
     """Safe str for Python 3 and Python 2"""
@@ -78,6 +82,11 @@ class UdockerTools(object):
                   "\nrequires udockertools tarball release :",
                   self._tarball_release, l=Msg.ERR)
 
+        msg = "LOG - udocker command line interface version: " + __version__ + \
+              "\nrequires udockertools tarball release :" + self._tarball_release
+        LOG.log(MSG_LEVEL, self._instructions.__doc__)
+        LOG.log(MSG_LEVEL, msg)
+
     def _version2int(self, version):
         """Convert version string to integer"""
         version_int = 0
@@ -110,14 +119,17 @@ class UdockerTools(object):
             f_path = self.localrepo.bindir + '/' + f_name
             FileUtil(f_path).register_prefix()
             FileUtil(f_path).remove(recursive=True)
+            LOG.debug("removed file: %s", f_path)
         for f_name in os.listdir(self.localrepo.libdir):
             f_path = self.localrepo.libdir + '/' + f_name
             FileUtil(f_path).register_prefix()
             FileUtil(f_path).remove(recursive=True)
+            LOG.debug("removed file: %s", f_path)
         for f_name in os.listdir(self.localrepo.docdir):
             f_path = self.localrepo.docdir + '/' + f_name
             FileUtil(f_path).register_prefix()
             FileUtil(f_path).remove(recursive=True)
+            LOG.debug("removed file: %s", f_path)
 
     def _download(self, url, fileout=""):
         """Download a file """
@@ -126,11 +138,11 @@ class UdockerTools(object):
         else:
             download_file = FileUtil("udockertools").mktmp()
 
-        if Msg.level <= Msg.DEF:
-            Msg().setlevel(Msg.NIL)
+        # if Msg.level <= Msg.DEF:
+        #     Msg().setlevel(Msg.NIL)
         (hdr, dummy) = self.curl.get(url, ofile=download_file, follow=True)
-        if Msg.level == Msg.NIL:
-            Msg().setlevel()
+        # if Msg.level == Msg.NIL:
+        #     Msg().setlevel()
         try:
             if "200" in hdr.data["X-ND-HTTPSTATUS"]:
                 return download_file
@@ -190,6 +202,7 @@ class UdockerTools(object):
                 if tar_in.name.startswith("udocker_dir/bin/"):
                     tar_in.name = os.path.basename(tar_in.name)
                     Msg().out("Info: extrating", tar_in.name, l=Msg.DBG)
+                    LOG.debug("LOG - Info: extrating %s" % tar_in.name)
                     tfile.extract(tar_in, self.localrepo.bindir)
             FileUtil(self.localrepo.bindir).rchmod(stat.S_IRUSR |
                                                    stat.S_IWUSR |
@@ -228,6 +241,7 @@ class UdockerTools(object):
     def get_installinfo(self):
         """Get json containing installation info"""
         Msg().out("Info: searching for messages:", l=Msg.VER)
+        LOG.log(MSG_LEVEL, "LOG - searching for messages:")
         for url in self._get_mirrors(self._installinfo):
             infofile = self._get_file(url)
             try:
@@ -235,27 +249,34 @@ class UdockerTools(object):
                     self._install_json = json.load(filep)
[…]
