+++ udocker/tools.py
@@ -13,7 +13,6 @@ from udocker import is_genstr
 from udocker.config import Config
 from udocker.utils.curl import GetURL
 from udocker.utils.fileutil import FileUtil
-from udocker.msg import Msg
 from udocker import __version__
 
 LOG = logging.getLogger(__name__)
@@ -77,13 +76,8 @@ class UdockerTools(object):
            by default under $HOME/.udocker
 
         """
-        Msg().out(self._instructions.__doc__, l=Msg.ERR)
-        Msg().out("udocker command line interface version:", __version__,
-                  "\nrequires udockertools tarball release :",
-                  self._tarball_release, l=Msg.ERR)
-
-        msg = "LOG - udocker command line interface version: " + __version__ + \
-              "\nrequires udockertools tarball release :" + self._tarball_release
+        msg = "udocker command line interface version: " + __version__ + \
+              "\nrequires udockertools tarball release: " + self._tarball_release
         LOG.log(MSG_LEVEL, self._instructions.__doc__)
         LOG.log(MSG_LEVEL, msg)
 
@@ -109,8 +103,8 @@ class UdockerTools(object):
 
     def is_available(self):
         """Are the tools already installed"""
-        version = \
-            FileUtil(self.localrepo.libdir + "/VERSION").getdata('r').strip()
+        version_file = self.localrepo.libdir + "/VERSION"
+        version = FileUtil(version_file).getdata('r').strip()
         return self._version_isok(version)
 
     def purge(self):
@@ -138,16 +132,13 @@ class UdockerTools(object):
         else:
             download_file = FileUtil("udockertools").mktmp()
 
-        # if Msg.level <= Msg.DEF:
-        #     Msg().setlevel(Msg.NIL)
         (hdr, dummy) = self.curl.get(url, ofile=download_file, follow=True)
-        # if Msg.level == Msg.NIL:
-        #     Msg().setlevel()
         try:
             if "200" in hdr.data["X-ND-HTTPSTATUS"]:
                 return download_file
         except (KeyError, TypeError, AttributeError):
             pass
+
         FileUtil(download_file).remove()
         return ""
 
@@ -160,8 +151,10 @@ class UdockerTools(object):
             filename = self._download(url, fileout)
         elif os.path.exists(url):
             filename = os.path.realpath(url)
+
         if filename and os.path.isfile(filename):
             return filename
+
         return ""
 
     def _verify_version(self, tarball_file):
@@ -179,10 +172,12 @@ class UdockerTools(object):
                 if tar_in.name.startswith("udocker_dir/lib/VERSION"):
                     tar_in.name = os.path.basename(tar_in.name)
                     tfile.extract(tar_in, path=tmpdir)
+
             tfile.close()
         except tarfile.TarError:
             FileUtil(tmpdir).remove(recursive=True)
             return (False, "")
+
         tarball_version = FileUtil(tmpdir + "/VERSION").getdata('r').strip()
         status = self._version_isok(tarball_version)
         FileUtil(tmpdir).remove(recursive=True)
@@ -201,84 +196,85 @@ class UdockerTools(object):
             for tar_in in tfile.getmembers():
                 if tar_in.name.startswith("udocker_dir/bin/"):
                     tar_in.name = os.path.basename(tar_in.name)
-                    Msg().out("Info: extrating", tar_in.name, l=Msg.DBG)
-                    LOG.debug("LOG - Info: extrating %s" % tar_in.name)
+                    LOG.debug("LOG - Info: extrating %s", tar_in.name)
                     tfile.extract(tar_in, self.localrepo.bindir)
+
             FileUtil(self.localrepo.bindir).rchmod(stat.S_IRUSR |
[…]
