### 2021-07-15 36f1777
implmenting new logging and messages
