+++ udocker/umain.py
@@ -11,7 +11,15 @@ from udocker.container.localrepo import LocalRepository
 from udocker.cli import UdockerCLI
 
 LOG = logging.getLogger(__name__)
-MSG_LEVEL = 60
+ch = logging.StreamHandler()
+ch.setFormatter(logging.Formatter("%(asctime)s:%(levelname)s: %(message)s"))
+LOG.addHandler(ch)
+
+MSG = logging.getLogger("Messages")
+ch = logging.StreamHandler(sys.stdout)
+ch.setFormatter(logging.Formatter("%(message)s"))
+MSG.setLevel(logging.INFO)
+MSG.addHandler(ch)
 
 
 class UMain(object):
@@ -35,8 +43,7 @@ class UMain(object):
         self.cmdp.parse(self.argv)
         allow_root = self.cmdp.get("--allow-root", "GEN_OPT")
         if not (os.geteuid() or allow_root):
-            Msg().err("Error: do not run as root !")
-            LOG.error("LOGGING - Error: do not run as root !")
+            LOG.error("do not run as root!")
             sys.exit(self.STATUS_ERROR)
 
         if self.cmdp.get("--config=", "GEN_OPT"):
@@ -63,13 +70,10 @@ class UMain(object):
         self.local = LocalRepository()
         if not self.local.is_repo():
             if topdir:
-                Msg().err("Error: invalid udocker repository:", topdir)
-                LOG.error("LOGGING - Error: invalid udocker repository: %s", topdir)
+                LOG.error("invalid udocker repository: %s", topdir)
                 sys.exit(self.STATUS_ERROR)
             else:
-                Msg().out("Info: creating repo: " + Config.conf['topdir'],
-                          l=Msg.INF)
-                LOG.info("LOGGING - Info: creating repo: %s", Config.conf['topdir'])
+                LOG.info("creating repo: %s", Config.conf['topdir'])
                 self.local.create_repo()
 
         self.cli = UdockerCLI(self.local)
@@ -111,8 +115,7 @@ class UMain(object):
         command = self.cmdp.get("", "CMD")
         if command in cmds:
             if self.cmdp.get("--help", "CMD_OPT"):
-                Msg().out(cmds[command].__doc__)
-                LOG.log(MSG_LEVEL, cmds[command].__doc__)
+                MSG.info(cmds[command].__doc__)
                 return self.STATUS_OK
             if command in ["version", "showconf", "avail"]:
                 return cmds[command](self.cmdp)
@@ -120,14 +123,11 @@ class UMain(object):
                 self.cli.do_install(None)
             exit_status = cmds[command](self.cmdp)  # executes command
             if self.cmdp.missing_options():
-                Msg().err("Error: syntax error at: %s" %
-                          " ".join(self.cmdp.missing_options()))
-                LOG.error("LOGGING - Error: syntax error at: %s",
+                LOG.error("Syntax error at: %s",
                           " ".join(self.cmdp.missing_options()))
                 return self.STATUS_ERROR
 
             return exit_status
 
-        Msg().err("Error: invalid command:", command, "\n")
-        LOG.error("LOG - Error: invalid command: %s\n", command)
+        LOG.error("Invalid command: %s\n", command)
         return self.STATUS_ERROR
