+++ udocker/tools.py
@@ -16,7 +16,15 @@ from udocker.utils.fileutil import FileUtil
 from udocker import __version__
 
 LOG = logging.getLogger(__name__)
-MSG_LEVEL = 60
+ch = logging.StreamHandler()
+ch.setFormatter(logging.Formatter("%(asctime)s:%(levelname)s: %(message)s"))
+LOG.addHandler(ch)
+
+MSG = logging.getLogger("Messages")
+ch = logging.StreamHandler(sys.stdout)
+ch.setFormatter(logging.Formatter("%(message)s"))
+MSG.setLevel(logging.INFO)
+MSG.addHandler(ch)
 
 def _str(data):
     """Safe str for Python 3 and Python 2"""
@@ -78,8 +86,8 @@ class UdockerTools(object):
         """
         msg = "udocker command line interface version: " + __version__ + \
               "\nrequires udockertools tarball release: " + self._tarball_release
-        LOG.log(MSG_LEVEL, self._instructions.__doc__)
-        LOG.log(MSG_LEVEL, msg)
+        MSG.info(self._instructions.__doc__)
+        MSG.info(msg)
 
     def _version2int(self, version):
         """Convert version string to integer"""
@@ -90,13 +98,16 @@ class UdockerTools(object):
                 version_int = version_int + (int(vitem) * factor)
             except (TypeError, ValueError):
                 pass
+
             factor = factor / 1000
+
         return int(version_int)
 
     def _version_isok(self, version):
         """Is version >= than the minimum required tarball release"""
         if not (version and self._tarball_release):
             return False
+
         tarball_version_int = self._version2int(version)
         required_version_int = self._version2int(self._tarball_release)
         return tarball_version_int >= required_version_int
@@ -114,11 +125,13 @@ class UdockerTools(object):
             FileUtil(f_path).register_prefix()
             FileUtil(f_path).remove(recursive=True)
             LOG.debug("removed file: %s", f_path)
+
         for f_name in os.listdir(self.localrepo.libdir):
             f_path = self.localrepo.libdir + '/' + f_name
             FileUtil(f_path).register_prefix()
             FileUtil(f_path).remove(recursive=True)
             LOG.debug("removed file: %s", f_path)
+
         for f_name in os.listdir(self.localrepo.docdir):
             f_path = self.localrepo.docdir + '/' + f_name
             FileUtil(f_path).register_prefix()
@@ -187,16 +200,16 @@ class UdockerTools(object):
         """Install the tarball"""
         if not (tarball_file and os.path.isfile(tarball_file)):
             return False
+
         FileUtil(self.localrepo.topdir).chmod()
         self.localrepo.create_repo()
-
         try:
             tfile = tarfile.open(tarball_file, "r:gz")
             FileUtil(self.localrepo.bindir).rchmod()
             for tar_in in tfile.getmembers():
                 if tar_in.name.startswith("udocker_dir/bin/"):
                     tar_in.name = os.path.basename(tar_in.name)
-                    LOG.debug("LOG - Info: extrating %s", tar_in.name)
+                    LOG.debug("extrating %s", tar_in.name)
                     tfile.extract(tar_in, self.localrepo.bindir)
 
             FileUtil(self.localrepo.bindir).rchmod(stat.S_IRUSR |
@@ -206,7 +219,7 @@ class UdockerTools(object):
             for tar_in in tfile.getmembers():
                 if tar_in.name.startswith("udocker_dir/lib/"):
                     tar_in.name = os.path.basename(tar_in.name)
-                    LOG.log(MSG_LEVEL, "Info: extrating %s", tar_in.name)
+                    LOG.debug("extrating %s", tar_in.name)
[…]
