+++ udocker/maincmd.py
@@ -21,15 +21,14 @@ limitations under the License.
 import os
 import sys
 import logging
-
-sys.path.append(os.path.dirname(os.path.abspath(sys.argv[0])) + '/../')
-
-# pylint: disable=wrong-import-position
-#from udocker.msg import Msg
 from udocker.umain import UMain
 from udocker.utils.fileutil import FileUtil
 
+sys.path.append(os.path.dirname(os.path.abspath(sys.argv[0])) + '/../')
 LOG = logging.getLogger(__name__)
+ch = logging.StreamHandler()
+ch.setFormatter(logging.Formatter("%(asctime)s:%(levelname)s: %(message)s"))
+LOG.addHandler(ch)
 
 
 def main():
@@ -39,8 +38,7 @@ def main():
     try:
         exit_status = UMain(sys.argv).execute()
     except KeyboardInterrupt:
-        #Msg().out("Info: keyboard interrupt", l=Msg.INF)
-        LOG.error("ERROR: keyboard interrupt")
+        LOG.error("keyboard interrupt")
         FileUtil().cleanup()
         exit_status = 1
     except SystemExit as error:

[…]
