### 2021-07-15 7e85241
wip on logging and messages
