+++ udocker/maincmd.py
@@ -21,16 +21,11 @@ limitations under the License.
 import os
 import sys
 import logging
+from udocker import LOG
 from udocker.umain import UMain
 from udocker.utils.fileutil import FileUtil
 
 sys.path.append(os.path.dirname(os.path.abspath(sys.argv[0])) + '/../')
-LOG = logging.getLogger(__name__)
-ch = logging.StreamHandler()
-logform = logging.Formatter("%(asctime)s:%(levelname)s: %(message)s")
-ch.setFormatter(logform)
-LOG.addHandler(ch)
-
 
 def main():
     """Program start and exception handling"""
