+++ udocker/umain.py
@@ -9,17 +9,13 @@ from udocker.cmdparser import CmdParser
 from udocker.config import Config
 from udocker.container.localrepo import LocalRepository
 from udocker.cli import UdockerCLI
+from udocker import LOG
 
-LOG = logging.getLogger(__name__)
-ch = logging.StreamHandler()
-ch.setFormatter(logging.Formatter("%(asctime)s:%(levelname)s: %(message)s"))
-LOG.addHandler(ch)
-
-MSG = logging.getLogger("Messages")
-msgout = logging.StreamHandler(sys.stdout)
-msgout.setFormatter(logging.Formatter("%(message)s"))
-MSG.setLevel(logging.INFO)
-MSG.addHandler(msgout)
+# MSG = logging.getLogger("Messages")
+# msgout = logging.StreamHandler(sys.stdout)
+# msgout.setFormatter(logging.Formatter("%(message)s"))
+# MSG.setLevel(logging.INFO)
+# MSG.addHandler(msgout)
 
 
 class UMain(object):
@@ -115,7 +111,7 @@ class UMain(object):
         command = self.cmdp.get("", "CMD")
         if command in cmds:
             if self.cmdp.get("--help", "CMD_OPT"):
-                MSG.info(cmds[command].__doc__)
+                # MSG.info(cmds[command].__doc__)
                 return self.STATUS_OK
             if command in ["version", "showconf", "avail"]:
                 return cmds[command](self.cmdp)
