### 2021-07-16 40c0d08
implment logging commonlocalfile
