### 2021-07-19 b8496d4
wip logging on docker
