+++ udocker/docker.py
@@ -7,6 +7,7 @@ import re
 import base64
 import json
 
+from udocker import LOG
 from udocker.config import Config
 from udocker.msg import Msg
 from udocker.commonlocalfile import CommonLocalFileApi
@@ -35,20 +36,24 @@ class DockerIoAPI(object):
 
     def set_proxy(self, http_proxy):
         """Select a socks http proxy for API access and file download"""
+        LOG.info("Setting proxy: %s", http_proxy)
         self.curl.set_proxy(http_proxy)
 
     def set_registry(self, registry_url):
         """Change docker registry url"""
+        LOG.info("Setting registry: %s", registry_url)
         self.registry_url = registry_url
 
     def set_index(self, index_url):
         """Change docker index url"""
+        LOG.info("Setting docker index: %s", index_url)
         self.index_url = index_url
 
     def is_repo_name(self, imagerepo):
         """Check if name matches authorized characters for a docker repo"""
         if imagerepo and re.match("^[a-zA-Z0-9][a-zA-Z0-9-_./:]+$", imagerepo):
             return True
+
         return False
 
     def _get_url(self, *args, **kwargs):
@@ -60,43 +65,55 @@ class DockerIoAPI(object):
         url = str(args[0])
         if "RETRY" not in kwargs:
             kwargs["RETRY"] = 3
+
         if "FOLLOW" not in kwargs:
             kwargs["FOLLOW"] = 3
+
         kwargs["RETRY"] -= 1
         (hdr, buf) = self.curl.get(*args, **kwargs)
-        Msg().out("Info: header: %s" % (hdr.data), l=Msg.DBG)
-        Msg().out("Info: buffer: %s" % (buf.getvalue()), l=Msg.DBG)
+        LOG.info("header: %s", (hdr.data))
+        LOG.info("buffer: %s", (buf.getvalue()))
         status_code = self.curl.get_status_code(hdr.data["X-ND-HTTPSTATUS"])
         if status_code == 200:
             return (hdr, buf)
+
         if not kwargs["RETRY"]:
             hdr.data["X-ND-CURLSTATUS"] = 13  # Permission denied
             return (hdr, buf)
+
         auth_kwargs = kwargs.copy()
         if "location" not in hdr.data:
             kwargs["FOLLOW"] = 3
+
         if "location" in hdr.data and hdr.data['location']:
             if not kwargs["FOLLOW"]:
                 hdr.data["X-ND-CURLSTATUS"] = 13
                 return (hdr, buf)
+
             kwargs["FOLLOW"] -= 1
             args = [hdr.data['location']]
             if "header" in auth_kwargs:
                 del auth_kwargs["header"]
+
         elif status_code == 401:
             if "www-authenticate" in hdr.data:
                 www_authenticate = hdr.data["www-authenticate"]
                 if not "realm" in www_authenticate:
                     return (hdr, buf)
+
                 if 'error="insufficient_scope"' in www_authenticate:
                     return (hdr, buf)
+
                 auth_header = ""
                 if "/v2/" in url:
                     auth_header = self._get_v2_auth(www_authenticate,
                                                     kwargs["RETRY"])
+
                 if "/v1/" in url:
                     auth_header = self._get_v1_auth(www_authenticate)
+
                 auth_kwargs.update({"header": [auth_header]})
+
         (hdr, buf) = self._get_url(*args, **auth_kwargs)
         return (hdr, buf)
 
@@ -111,30 +128,38 @@ class DockerIoAPI(object):
             layer_f_chksum = ChkSUM().hash(filename, match.group(1))
             if layer_f_chksum == match.group(2):
                 return True             # is cached skip download
+
             cache_mode = 0
+
         if self.curl.cache_support and cache_mode:
             if cache_mode == 1:
[…]
