### 2021-07-19 8a2c682
move to logging in docker
