+++ udocker/docker.py
@@ -9,7 +9,6 @@ import json
 
 from udocker import LOG
 from udocker.config import Config
-from udocker.msg import Msg
 from udocker.commonlocalfile import CommonLocalFileApi
 from udocker.utils.fileutil import FileUtil
 from udocker.utils.curl import GetURL
@@ -246,7 +245,7 @@ class DockerIoAPI(object):
     def get_v1_image_tag(self, endpoint, imagerepo, tag):
         """Get list of tags in a repo from Docker Hub"""
         url = endpoint + "/v1/repositories/" + imagerepo + "/tags/" + tag
-        Msg().out("Info: tags url", url, l=Msg.DBG)
+        LOG.debug("tags url: %s", url)
         (hdr, buf) = self._get_url(url)
         try:
             return (hdr.data, json.loads(buf.getvalue()))
@@ -256,7 +255,7 @@ class DockerIoAPI(object):
     def get_v1_image_ancestry(self, endpoint, image_id):
         """Get the ancestry which is an ordered list of layers"""
         url = endpoint + "/v1/images/" + image_id + "/ancestry"
-        Msg().out("Info: ancestry url", url, l=Msg.DBG)
+        LOG.debug("ancestry url: %s", url)
         (hdr, buf) = self._get_url(url)
         try:
             return (hdr.data, json.loads(buf.getvalue()))
@@ -266,21 +265,23 @@ class DockerIoAPI(object):
     def get_v1_image_json(self, endpoint, layer_id):
         """Get the JSON metadata for a specific layer"""
         url = endpoint + "/v1/images/" + layer_id + "/json"
-        Msg().out("Info: json url", url, l=Msg.DBG)
+        LOG.debug("json url: %s", url)
         filename = self.localrepo.layersdir + '/' + layer_id + ".json"
         if self._get_file(url, filename, 0):
             self.localrepo.add_image_layer(filename)
             return True
+
         return False
 
     def get_v1_image_layer(self, endpoint, layer_id):
         """Get a specific layer data file (layer files are tarballs)"""
         url = endpoint + "/v1/images/" + layer_id + "/layer"
-        Msg().out("Info: layer url", url, l=Msg.DBG)
+        LOG.debug("layer url: %s", url)
         filename = self.localrepo.layersdir + '/' + layer_id + ".layer"
         if self._get_file(url, filename, 3):
             self.localrepo.add_image_layer(filename)
             return True
+
         return False
 
     def get_v1_layers_all(self, endpoint, layer_list):
@@ -288,15 +289,18 @@ class DockerIoAPI(object):
         files = []
         if layer_list:
             for layer_id in reversed(layer_list):
-                Msg().out("Info: downloading layer", layer_id, l=Msg.INF)
+                LOG.info("downloading layer: %s", layer_id)
                 filesize = self.get_v1_image_json(endpoint, layer_id)
                 if not filesize:
                     return []
+
                 files.append(layer_id + ".json")
                 filesize = self.get_v1_image_layer(endpoint, layer_id)
                 if not filesize:
                     return []
+
                 files.append(layer_id + ".layer")
+
         return files
 
     def _get_v2_auth(self, www_authenticate, retry):
@@ -359,12 +363,14 @@ class DockerIoAPI(object):
                 return True
         except (KeyError, AttributeError, TypeError):
             pass
+
         return False
 
     def has_search_v2(self, url=None):
         """Check if registry has search capabilities in v2"""
         if url is None:
             url = self.registry_url
+
         (hdr, dummy) = self._get_url(url + "/v2/search/repositories")
         try:
             if ("200" in hdr.data["X-ND-HTTPSTATUS"] or
@@ -372,18 +378,20 @@ class DockerIoAPI(object):
                 return True
         except (KeyError, AttributeError, TypeError):
             pass
+
         return False
 
[…]
