### 2021-07-20 indigo-dc/udocker#312 @mviereck
Thanks.
This is more comprehensive than the current chapter in the paper that contains a lot of detail that does not belong there, like e.g. API documentation.
I suggest to replace it with this summary.

>  If someone writes a scientific paper and wants to cite udocker: at which point he would have the urgent need to cite the new paper instead of the previous one?

So far udocker got a lot of maintenance towards a better code base and broader support of different environments. The essential idea and concept is still the same and fits the description in the first paper. So there is basically no need for a new paper except to provide a recent version number.
