### 2021-07-20 #3295 @mviereck
@jorge-lip provided a more [comprehensive summary of the changes](https://github.com/indigo-dc/udocker/issues/312#issuecomment-875747957) done in udocker. It boils down to good code maintenance and a broader support of different environments.
The essential idea and concept is still the same and fits the description in the first paper. So there is basically no need for a new paper except to provide a recent version number.

Though, the changes are big enough to fit the definition of "incremental progress":
> The general feeling on this in JOSS is that each JOSS paper should be a clear contribution, in the same way that a research project can lead to multiple papers, some of which can be incremental progress in an area that was previously published.

However, my feeling is that JOSS needs a better handling of software versioning. A well  done paper describes the idea and concept of a software without going too much into details like API description or supported environments.
Good maintenance can be a lot of work, while the paper still fits the software after years. This is the case here.
A solution to update software versions to existing papers in JOSS is desirable, so one could cite a paper referring to recent versions.

The first paper hasn't been published at JOSS. So the review still has its use for the documentation of udocker that has severely improved meanwhile.

---------------

Currently I wait for an answer in ticket https://github.com/indigo-dc/udocker/issues/348 . The paper lists a lot of authors that do not appear as contributors on github.
