### 2021-07-20 #3295 @danielskatz
@mviereck - I think the changes/work are sufficient for a new version of the paper, so let's take this as ok from the point of view of the journal (though I agree it's a complicated issue where we don't have a good simply-stated policy), and focus on other issues, like the question about the authors. In addition to this, I see three more unchecked boxes in your review - what information/changes would be needed to check these?
