### 2021-07-20 #3295 @mviereck
> I think the changes/work are sufficient for a new version of the paper, so let's take this as ok from the point of view of the journal

I have checked this one off.

> In addition to this, I see three more unchecked boxes in your review - what information/changes would be needed to check these?

I am just checking point after point. Once I've checked one off, I go to the next. So I did not look closer yet to the remaining points.
