### 2021-07-20 9c79492
implement logging and msg in cli
