+++ udocker/cli.py
@@ -10,7 +10,6 @@ from getpass import getpass
 
 from udocker import __version__, LOG, MSG
 from udocker.config import Config
-from udocker.msg import Msg
 from udocker.docker import DockerIoAPI
 from udocker.localfile import LocalFileAPI
 from udocker.helper.keystore import KeyStore
@@ -79,7 +78,7 @@ class UdockerCLI(object):
 
         if not (imagerepo and tag and
                 self.dockerioapi.is_repo_name(imagespec)):
-            Msg().err("Error: must specify image:tag or repository/image:tag")
+            LOG.error("must specify image:tag or repository/image:tag")
             return(None, None)
 
         return imagerepo, tag
@@ -91,7 +90,7 @@ class UdockerCLI(object):
 
         if not (imagerepo and
                 self.dockerioapi.is_repo_name(imagerepo)):
-            Msg().err("Error: enter image or repository/image without tag")
+            LOG.error("enter image or repository/image without tag")
             return None
 
         return imagerepo
@@ -106,6 +105,7 @@ class UdockerCLI(object):
         if registry_url:
             if "://" not in registry_url:
                 registry_url = "https://" + registry_url
+
             self.dockerioapi.set_registry(registry_url)
             if not index_url:
                 self.dockerioapi.set_index("")
@@ -113,6 +113,7 @@ class UdockerCLI(object):
         if index_url:
             if "://" not in index_url:
                 index_url = "https://" + index_url
+
             self.dockerioapi.set_index(index_url)
             if not registry_url:
                 self.dockerioapi.set_registry("")
@@ -125,6 +126,7 @@ class UdockerCLI(object):
                     (hostname, dummy) = imagerepo.split('/')[:2]
             except (ValueError, IndexError, TypeError):
                 return False
+
             if "." in hostname:
                 try:
                     self.dockerioapi.set_registry( \
@@ -141,6 +143,7 @@ class UdockerCLI(object):
         """Split image repo into hostname, repo, tag"""
         if not imagerepo:
             return ("", "", "", "")
+
         transport = ""
         hostname = ""
         image = imagerepo
@@ -152,11 +155,14 @@ class UdockerCLI(object):
                 (hostname, image) = imagerepo.split('/', 1)
         except (ValueError, IndexError, TypeError):
             pass
+
         if hostname and '.' not in hostname:
             image = hostname + '/' + image
             hostname = ""
+
         if ':' in image:
             (image, tag) = image.split(':', 1)
+
         return (transport, hostname, image, tag)
 
     def do_mkrepo(self, cmdp):
@@ -167,13 +173,14 @@ class UdockerCLI(object):
         topdir = cmdp.get("P1")
         if cmdp.missing_options():
             return self.STATUS_ERROR
+
         if not os.path.exists(topdir):
             self.localrepo.setup(topdir)
             if not self.localrepo.create_repo():
-                Msg().err("Error: localrepo creation failure: ", topdir)
+                LOG.error("localrepo creation failure: %s", topdir)
                 return self.STATUS_ERROR
         else:
-            Msg().err("Error: localrepo directory already exists: ", topdir)
+            LOG.error("localrepo directory already exists: %s", topdir)
             return self.STATUS_ERROR
 
         return self.STATUS_OK
@@ -185,6 +192,7 @@ class UdockerCLI(object):
                 is_official = "[OK]"
             else:
                 is_official = "----"
+
             description = ""
             for dfield in ("description", "short_description"):
                 if dfield in repo and repo[dfield] is not None:
@@ -193,16 +201,21 @@ class UdockerCLI(object):
                             break
                         if char in string.printable:
                             description += char
+
                     break
+
             name = ""
[…]
