### 2021-07-21 52a474f
refactor to logging in localrepo
