+++ udocker/container/localrepo.py
@@ -7,9 +7,8 @@ import sys
 import stat
 import json
 
-from udocker import is_genstr
+from udocker import is_genstr, LOG
 from udocker.config import Config
-from udocker.msg import Msg
 from udocker.utils.fileutil import FileUtil
 from udocker.utils.chksum import ChkSUM
 from udocker.utils.uprocess import Uprocess
@@ -41,24 +40,27 @@ class LocalRepository(object):
         self.layersdir = Config.conf['layersdir']
         self.containersdir = Config.conf['containersdir']
         self.homedir = Config.conf['homedir']
-
         if not self.bindir:
             self.bindir = self.topdir + "/bin"
+
         if not self.libdir:
             self.libdir = self.topdir + "/lib"
+
         if not self.docdir:
             self.docdir = self.topdir + "/doc"
+
         if not self.reposdir:
             self.reposdir = self.topdir + "/repos"
+
         if not self.layersdir:
             self.layersdir = self.topdir + "/layers"
+
         if not self.containersdir:
             self.containersdir = self.topdir + "/containers"
 
         self.cur_repodir = ""
         self.cur_tagdir = ""
         self.cur_containerdir = ""
-
         FileUtil(self.reposdir).register_prefix()
         FileUtil(self.layersdir).register_prefix()
         FileUtil(self.containersdir).register_prefix()
@@ -74,23 +76,31 @@ class LocalRepository(object):
         try:
             if not os.path.exists(self.topdir):
                 os.makedirs(self.topdir)
+
             if not os.path.exists(self.reposdir):
                 os.makedirs(self.reposdir)
+
             if not os.path.exists(self.layersdir):
                 os.makedirs(self.layersdir)
+
             if not os.path.exists(self.containersdir):
                 os.makedirs(self.containersdir)
+
             if not os.path.exists(self.bindir):
                 os.makedirs(self.bindir)
+
             if not os.path.exists(self.libdir):
                 os.makedirs(self.libdir)
+
             if not os.path.exists(self.docdir):
                 os.makedirs(self.docdir)
+
             if not (Config.conf['keystore'].startswith("/") or \
                     os.path.exists(self.homedir)):
                 os.makedirs(self.homedir)
         except(IOError, OSError):
             return False
+
         return True
 
     def is_repo(self):
@@ -108,10 +118,12 @@ class LocalRepository(object):
         """
         if not is_genstr(obj):
             return False
+
         match = re.match(
             "^[a-z0-9]+-[a-z0-9]+-[a-z0-9]+-[a-z0-9]+-[a-z0-9]+$", obj)
         if match:
             return True
+
         return False
 
     def protect_container(self, container_id):
@@ -148,14 +160,17 @@ class LocalRepository(object):
         container_root = self.cd_container(container_id) + "/ROOT"
         if not os.path.exists(container_root):
             return 2
+
         if not os.path.isdir(container_root):
             return 3
+
         if os.access(container_root, os.W_OK):
             return 1
+
         return 0
 
     def get_size(self, container_id):
-        """See if a container root dir is writable by this user"""
+        """Get size of the container"""
         container_root = self.cd_container(container_id) + "/ROOT"
         try:
             size, dummy = Uprocess().get_output(["du", "-s", "-m", "-x",
@@ -174,6 +189,7 @@ class LocalRepository(object):
         containers_list = []
         if not os.path.isdir(self.containersdir):
             return []
+
         for fname in os.listdir(self.containersdir):
             container_dir = self.containersdir + '/' + fname
             if os.path.isdir(container_dir):
@@ -184,13 +200,16 @@ class LocalRepository(object):
                 else:
                     reponame = filep.read()
                     filep.close()
+
                 if dir_only:
[…]
