### 2021-07-21 3683715
implement logging structure
