+++ udocker/container/structure.py
@@ -2,11 +2,12 @@
 """Tools to manage the container structure"""
 
 import os
+import sys
 import subprocess
+import logging
 
-from udocker import is_genstr
+from udocker import is_genstr, LOG
 from udocker.config import Config
-from udocker.msg import Msg
 from udocker.utils.fileutil import FileUtil
 from udocker.utils.uprocess import Uprocess
 from udocker.helper.unique import Unique
@@ -34,13 +35,13 @@ class ContainerStructure(object):
         else:
             container_dir = self.localrepo.cd_container(self.container_id)
             if not container_dir:
-                Msg().err("Error: container id or name not found")
+                LOG.error("container id or name not found")
                 return(False, False)
 
             fjson = container_dir + "/container.json"
             container_json = self.localrepo.load_json(fjson)
             if not container_json:
-                Msg().err("Error: invalid container json metadata")
+                LOG.error("invalid container json metadata")
                 return(False, False)
 
         return(container_dir, container_json)
@@ -52,6 +53,7 @@ class ContainerStructure(object):
             confidx = "config"
         elif "container_config" in container_json:
             confidx = "container_config"
+
         if container_json[confidx]  and param in container_json[confidx]:
             if container_json[confidx][param] is None:
                 pass
@@ -69,14 +71,15 @@ class ContainerStructure(object):
                 return self._dict_to_list(container_json[confidx][param])
             else:
                 return container_json[confidx][param]
+
         return default
 
-    # DEBUG
     def _dict_to_str(self, in_dict):
         """Convert dict to str"""
         out_str = ""
         for (key, val) in in_dict.items():
             out_str += "%s:%s " % (str(key), str(val))
+
         return out_str
 
     def _dict_to_list(self, in_dict):
@@ -84,6 +87,7 @@ class ContainerStructure(object):
         out_list = []
         for (key, val) in in_dict.items():
             out_list.append("%s:%s" % (str(key), str(val)))
+
         return out_list
 
     def _chk_container_root(self, container_id=None):
@@ -92,8 +96,10 @@ class ContainerStructure(object):
             container_dir = self.localrepo.cd_container(container_id)
         else:
             container_dir = self.localrepo.cd_container(self.container_id)
+
         if not container_dir:
             return 0
+
         container_root = container_dir + "/ROOT"
         check_list = ["/lib", "/bin", "/etc", "/tmp", "/var", "/usr", "/sys",
                       "/dev", "/data", "/home", "/system", "/root", "/proc", ]
@@ -101,6 +107,7 @@ class ContainerStructure(object):
         for f_path in check_list:
             if os.path.exists(container_root + f_path):
                 found += 1
+
         return found
 
     def create_fromimage(self, imagerepo, tag):
@@ -115,12 +122,12 @@ class ContainerStructure(object):
         self.tag = tag
         image_dir = self.localrepo.cd_imagerepo(self.imagerepo, self.tag)
         if not image_dir:
-            Msg().err("Error: create container: imagerepo is invalid")
+            LOG.error("create container: imagerepo is invalid")
             return False
 
         (container_json, layer_files) = self.localrepo.get_image_attributes()
         if not container_json:
-            Msg().err("Error: create container: getting layers or json")
+            LOG.error("create container: getting layers or json")
             return False
 
         if not self.container_id:
@@ -129,17 +136,16 @@ class ContainerStructure(object):
         container_dir = self.localrepo.setup_container(
             self.imagerepo, self.tag, self.container_id)
         if not container_dir:
-            Msg().err("Error: create container: setting up container")
+            LOG.error("create container: setting up container")
             return False
 
         self.localrepo.save_json(
[…]
