### 2021-07-21 406fbf4
implement logging in base
