+++ udocker/engine/base.py
@@ -6,8 +6,7 @@ import sys
 import re
 import json
 
-from udocker import is_genstr
-from udocker.msg import Msg
+from udocker import is_genstr, LOG, MSG
 from udocker.config import Config
 from udocker.utils.fileutil import FileUtil
 from udocker.utils.uenv import Uenv
@@ -82,11 +81,13 @@ class ExecutionEngineCommon(object):
                 except (ValueError, TypeError):
                     host_port = None
                     cont_port = None
+
             if host_port:
                 if by_container:
                     indexed_portmap[cont_port] = host_port
                 else:
                     indexed_portmap[host_port] = cont_port
+
         return indexed_portmap
 
     def _check_exposed_ports(self):
@@ -107,12 +108,16 @@ class ExecutionEngineCommon(object):
                     if port_number in list(mapped_ports.keys()):
                         if mapped_ports[port_number] >= 1024:
                             continue
+
                     exposes_priv = True
+
         if exposes_priv and HostInfo.uid != 0:
-            Msg().err("Error: this container exposes privileged TCP/IP ports")
+            LOG.error("this container exposes privileged TCP/IP ports")
             return False
+
         if exposes_port:
-            Msg().out("Warning: this container exposes TCP/IP ports", l=Msg.WAR)
+            LOG.warning("this container exposes TCP/IP ports")
+
         return True
 
     def _set_cpu_affinity(self):
@@ -120,6 +125,7 @@ class ExecutionEngineCommon(object):
         # find set affinity executable
         if not self.opt["cpuset"]:
             return []
+
         exec_cmd = []
         for exec_cmd in Config.conf['cpu_affinity_exec_tools']:
             exec_name = \
@@ -129,7 +135,9 @@ class ExecutionEngineCommon(object):
                 for (index, arg) in enumerate(exec_cmd):
                     if arg == "%s":
                         exec_cmd[index] = self.opt["cpuset"]
+
                 return exec_cmd
+
         self.opt["cpuset"] = ""
         return []
 
@@ -137,9 +145,11 @@ class ExecutionEngineCommon(object):
         """Create mountpoint"""
         if dirs_only and not FileUtil(host_path).isdir():
             return True
+
         if self.mountp.create(host_path, cont_path):
             self.mountp.save(cont_path)
             return True
+
         return False
 
     def _check_volumes(self):
@@ -147,28 +157,34 @@ class ExecutionEngineCommon(object):
         for vol in list(self.opt["vol"]):
             (host_path, cont_path) = Uvolume(vol).split()
             if not (host_path and host_path.startswith('/')):
-                Msg().err("Error: invalid host volume path:", host_path)
+                LOG.error("invalid host volume path: %s", host_path)
                 return False
+
             if not (cont_path and cont_path != '/' and
                     cont_path.startswith('/')):
-                Msg().err("Error: invalid container volume path:", cont_path)
+                LOG.error("invalid container volume path: %s", cont_path)
                 return False
+
             if not os.path.exists(host_path):
                 if (host_path in Config.conf['dri_list'] or
                         host_path in Config.conf['sysdirs_list']):
                     self.opt["vol"].remove(vol)
                     continue
-                Msg().err("Error: invalid host volume path:", host_path)
+
+                LOG.error("invalid host volume path: %s", host_path)
                 return False
+
             if not self._create_mountpoint(host_path, cont_path):
-                Msg().err("Error: creating mountpoint:", host_path, cont_path)
+                LOG.error("creating mountpoint: %s:%s", host_path, cont_path)
                 return False
+
         return True
 
     def _get_bindhome(self):
[…]
