### 2021-07-21 cf32984
implement logging in execmode
