### 2021-07-21 3b4a0b4
implement logging in fakechroot
