+++ udocker/engine/fakechroot.py
@@ -6,10 +6,9 @@ import sys
 import re
 import subprocess
 
-from udocker import is_genstr
+from udocker import is_genstr, LOG
 from udocker.engine.base import ExecutionEngineCommon
 from udocker.helper.osinfo import OSInfo
-from udocker.msg import Msg
 from udocker.config import Config
 from udocker.utils.fileutil import FileUtil
 from udocker.utils.uvolume import Uvolume
@@ -73,10 +72,10 @@ class FakechrootEngine(ExecutionEngineCommon):
         f_util = FileUtil(self.localrepo.libdir)
         fakechroot_so = f_util.find_file_in_dir(image_list)
         if not os.path.exists(fakechroot_so):
-            Msg().err("Error: no libfakechroot found", image_list)
+            LOG.error("no libfakechroot found: %s", image_list)
             sys.exit(1)
 
-        Msg().out("Info: fakechroot_so:", fakechroot_so, l=Msg.DBG)
+        LOG.debug("fakechroot_so: %s", fakechroot_so)
         return fakechroot_so
 
     def _setup_container_user(self, user):
@@ -87,8 +86,7 @@ class FakechrootEngine(ExecutionEngineCommon):
         """Check the uid_map string for container run command"""
         if ("user" in self.opt and (self.opt["user"] == '0' or
                                     self.opt["user"] == "root")):
-            Msg().out("Warning: this engine does not support execution as root",
-                      l=Msg.WAR)
+            LOG.warning("this engine does not support execution as root")
 
     def _get_volume_bindings(self):
         """Get the volume bindings string for fakechroot run"""
@@ -99,6 +97,7 @@ class FakechrootEngine(ExecutionEngineCommon):
             (host_path, cont_path) = Uvolume(vol).split()
             if not (host_path and cont_path):
                 continue
+
             real_host_path = os.path.realpath(host_path)
             if (host_path == cont_path and
                     Config.conf['fakechroot_expand_symlinks'] is False):
@@ -112,8 +111,10 @@ class FakechrootEngine(ExecutionEngineCommon):
                 map_volumes_dict[cont_path] = real_host_path + '!' + cont_path
                 if host_path != real_host_path or os.path.isdir(real_host_path):
                     self._recommend_expand_symlinks = True
+
         for cont_path in sorted(map_volumes_dict, reverse=True):
             map_volumes_list.append(map_volumes_dict[cont_path])
+
         return (':'.join(host_volumes_list), ':'.join(map_volumes_list))
 
     def _get_access_filesok(self):
@@ -127,6 +128,7 @@ class FakechrootEngine(ExecutionEngineCommon):
                                                              self.opt["vol"])
             if h_file and os.path.exists(h_file):
                 file_list.append(c_path)
+
         return ":".join(file_list)
 
     def _fakechroot_env_set(self):
@@ -155,7 +157,7 @@ class FakechrootEngine(ExecutionEngineCommon):
         if map_volumes:
             self.opt["env"].append("FAKECHROOT_DIR_MAP=" + map_volumes)
 
-        if Msg.level >= Msg.DBG:
+        if Config.conf['verbose_level'] == logging.DEBUG:
             self.opt["env"].append("FAKECHROOT_DEBUG=true")
             self.opt["env"].append("LD_DEBUG=libs:files")
 
@@ -196,11 +198,11 @@ class FakechrootEngine(ExecutionEngineCommon):
     def _run_invalid_options(self):
         """check -p --publish -P --publish-all --net-coop"""
         if self.opt["portsmap"]:
-            Msg().out("Warning: this execution mode does not support "
-                      "-p --publish", l=Msg.WAR)
+            LOG.warning("this execution mode does not support -p --publish")
+
         if self.opt["netcoop"]:
-            Msg().out("Warning: this execution mode does not support "
[…]
