### 2021-07-21 ab8633d
implement logging in nvidia
