+++ udocker/engine/nvidia.py
@@ -7,7 +7,7 @@ import glob
 import re
 import shutil
 
-from udocker.msg import Msg
+from udocker import LOG
 from udocker.config import Config
 from udocker.utils.fileutil import FileUtil
 from udocker.utils.uprocess import Uprocess
@@ -37,8 +37,8 @@ class NvidiaMode(object):
 
     def _copy_files(self, host_src_dir, cont_dst_dir, files_list, force=False):
         """copy or link file to destination creating directories as needed"""
-        Msg().out("Source (host) dir ", host_src_dir, l=Msg.DBG)
-        Msg().out("Destination (container) dir ", cont_dst_dir, l=Msg.DBG)
+        LOG.debug("Source (host) dir %s", host_src_dir)
+        LOG.debug("Destination (container) dir %s", cont_dst_dir)
         for fname in files_list:
             srcname = host_src_dir + '/' + fname
             dstname = self.container_root + '/' + cont_dst_dir + '/' + fname
@@ -47,10 +47,10 @@ class NvidiaMode(object):
                     try:
                         os.remove(dstname)
                     except OSError:
-                        Msg().err("Error: deleting nvidia file", dstname)
+                        LOG.error("deleting nvidia file %s", dstname)
                 else:
-                    Msg().out("Info: nvidia file already exists", dstname,
-                              ", use --force to overwrite", l=Msg.INF)
+                    LOG.info("nvidia file already exists %s, "
+                            "use --force to overwrite", dstname)
                     return False
 
             srcdir = os.path.dirname(srcname)
@@ -61,15 +61,15 @@ class NvidiaMode(object):
                     os.chmod(dstdir, stat.S_IMODE(os.stat(srcdir).st_mode) |
                              stat.S_IRWXU)
                 except OSError:
-                    Msg().err("Error: creating nvidia dir", dstdir)
+                    LOG.error("creating nvidia dir %s", dstdir)
 
             if os.path.islink(srcname):
                 linkto = os.readlink(srcname)
                 os.symlink(linkto, dstname)
-                Msg().out("Info: is link", srcname, "to", dstname, l=Msg.DBG)
+                LOG.info("is link %s to %s", srcname, dstname)
             elif os.path.isfile(srcname):
                 shutil.copy2(srcname, dstname)
-                Msg().out("Info: is file", srcname, "to", dstname, l=Msg.DBG)
+                LOG.info("is file %s to %s", srcname, dstname)
                 try:
                     mask = stat.S_IMODE(os.stat(srcname).st_mode) | \
                                         stat.S_IWUSR | stat.S_IRUSR
@@ -77,11 +77,11 @@ class NvidiaMode(object):
                         mask = mask | stat.S_IXUSR
                     os.chmod(dstname, mask)
                 except (IOError, OSError) as error:
-                    Msg().err("Error: change mask of nvidia file", error)
+                    LOG.error("change mask of nvidia file: %s", error)
             else:
-                Msg().err("Warn: nvidia file in config not found", srcname)
+                LOG.warning("nvidia file in config not found: %s", srcname)
 
-        Msg().out("Info: nvidia copied", srcname, "to", dstname, l=Msg.DBG)
+        LOG.debug("nvidia copied %s to %s", srcname, dstname)
         return True
 
     def _get_nvidia_libs(self, host_dir):
@@ -90,7 +90,8 @@ class NvidiaMode(object):
         for lib in Config.conf['nvi_lib_list']:
             for expanded_libs in glob.glob(host_dir + '/' + lib + '*'):
                 lib_list.append(expanded_libs.replace(host_dir, ''))
[…]
