### 2021-07-21 192c9b5
implement logging in proot
