+++ udocker/engine/runc.py
@@ -8,10 +8,11 @@ import platform
 import stat
 import select
 import json
+import logging
 
+from udocker import LOG
 from udocker.engine.base import ExecutionEngineCommon
 from udocker.engine.proot import PRootEngine
-from udocker.msg import Msg
 from udocker.config import Config
 from udocker.utils.fileutil import FileUtil
 from udocker.utils.uvolume import Uvolume
@@ -68,8 +69,9 @@ class RuncEngine(ExecutionEngineCommon):
             self.executable = f_util.find_file_in_dir(image_list)
 
         if not os.path.exists(self.executable):
-            Msg().err("Error: runc/crun executable not found")
+            LOG.error("runc/crun executable not found")
             sys.exit(1)
+
         if "crun" in os.path.basename(self.executable):
             self.engine_type = "crun"
         elif "runc" in os.path.basename(self.executable):
@@ -83,7 +85,7 @@ class RuncEngine(ExecutionEngineCommon):
 
         if FileUtil(self._container_specfile).size() == -1:
             cmd_l = [self.executable, "spec", "--rootless", ]
-            status = subprocess.call(cmd_l, shell=False, stderr=Msg.chlderr,
+            status = subprocess.call(cmd_l, shell=False, stderr=sys.stderr,
                                      close_fds=True,
                                      cwd=os.path.realpath(\
                                          self._container_specdir))
@@ -114,6 +116,7 @@ class RuncEngine(ExecutionEngineCommon):
             if outfile:
                 outfile.close()
             return False
+
         outfile.close()
         return True
 
@@ -151,6 +154,7 @@ class RuncEngine(ExecutionEngineCommon):
         else:
             json_obj["linux"]["uidMappings"] = [ \
                     {"containerID": 0, "hostID": HostInfo.uid, "size":1, }, ]
+
         if "gidMappings" in json_obj["linux"]:
             for idmap in json_obj["linux"]["gidMappings"]:
                 if "hostID" in idmap:
@@ -171,13 +175,13 @@ class RuncEngine(ExecutionEngineCommon):
         """Check the uid_map string for container run command"""
         if ("user" in self.opt and self.opt["user"] != "0" and
                 self.opt["user"] != "root"):
-            Msg().out("Warning: this engine only supports execution as root",
-                      l=Msg.WAR)
+            LOG.warning("this engine only supports execution as root")
 
     def _add_capabilities_spec(self):
         """Set the spec capabilities"""
         if not Config.conf['runc_capabilities']:
             return
+
         self._container_specjson["process"]["capabilities"]["ambient"] = \
             Config.conf['runc_capabilities']
         self._container_specjson["process"]["capabilities"]["bounding"] = \
@@ -192,25 +196,31 @@ class RuncEngine(ExecutionEngineCommon):
     def _add_device_spec(self, dev_path, mode="rwm"):
         """Add device to the configuration"""
         if not (os.path.exists(dev_path) and dev_path.startswith("/dev/")):
-            Msg().err("Error: device not found", dev_path)
+            LOG.error("device not found: %s", dev_path)
             return False
+
         dev_stat = os.stat(dev_path)
         if stat.S_ISBLK(dev_stat.st_mode):
             dev_type = 'b'
         elif stat.S_ISCHR(dev_stat.st_mode):
             dev_type = 'c'
         else:
-            Msg().err("Error: not a device", dev_path)
+            LOG.error("not a device: %s", dev_path)
             return False
+
         filemode = 0
         if 'r' in mode.lower():
             filemode += 0o444
+
         if 'w' in mode.lower():
             filemode += 0o222
+
         if not filemode:
             filemode = 0o666
+
         if "devices" not in self._container_specjson["linux"]:
             self._container_specjson["linux"]["devices"] = []
+
         device = {
[…]
