### 2021-07-21 d5a78cb
implement logging in singularity
