+++ udocker/engine/singularity.py
@@ -4,9 +4,10 @@
 import sys
 import os
 import subprocess
+import logging
 
+from udocker import LOG
 from udocker.engine.base import ExecutionEngineCommon
-from udocker.msg import Msg
 from udocker.config import Config
 from udocker.utils.fileutil import FileUtil
 from udocker.utils.uvolume import Uvolume
@@ -49,7 +50,7 @@ class SingularityEngine(ExecutionEngineCommon):
             self.executable = f_util.find_file_in_dir(image_list)
 
         if not os.path.exists(self.executable):
-            Msg().err("Error: singularity executable not found")
+            LOG.error("Error: singularity executable not found")
             sys.exit(1)
 
     def _get_volume_bindings(self):
@@ -68,16 +69,21 @@ class SingularityEngine(ExecutionEngineCommon):
                     tmp_is_binded = True
                 elif host_path == "/var/tmp" and cont_path in ("", "/var/tmp"):
                     vartmp_is_binded = True
+
             vol_list.extend(["-B", "%s:%s" % (host_path, cont_path), ])
+
         if not home_is_binded:
             vol_list.extend(["--home", "%s/root:%s" %
                              (self.container_root, "/root"), ])
+
         if not tmp_is_binded:
             vol_list.extend(["-B", "%s/tmp:/tmp" %
                              (self.container_root), ])
+
         if not vartmp_is_binded:
             vol_list.extend(["-B", "%s/var/tmp:/var/tmp" %
                              (self.container_root), ])
+
         return vol_list
 
     def _singularity_env_get(self):
@@ -87,6 +93,7 @@ class SingularityEngine(ExecutionEngineCommon):
         singularityenv = dict()
         for (key, val) in self.opt["env"]:
             singularityenv['SINGULARITYENV_%s' % key] = val
+
         return singularityenv
 
     def _make_container_directories(self):
@@ -101,11 +108,12 @@ class SingularityEngine(ExecutionEngineCommon):
     def _run_invalid_options(self):
         """check -p --publish -P --publish-all --net-coop"""
         if self.opt["portsmap"]:
-            Msg().out("Warning: this execution mode does not support "
-                      "-p --publish", l=Msg.WAR)
+            LOG.warning("this execution mode does not support "
+                        "-p --publish")
+
         if self.opt["netcoop"]:
-            Msg().out("Warning: this execution mode does not support "
-                      "-P --netcoop --publish-all", l=Msg.WAR)
+            LOG.warning("Warning: this execution mode does not support "
+                        "-P --netcoop --publish-all")
 
     def _run_as_root(self):
         """Set configure running as normal user or as root via --fakeroot
@@ -116,7 +124,7 @@ class SingularityEngine(ExecutionEngineCommon):
                 return False
 
             if self.opt["user"] != "root" and self.opt["uid"] != '0':
-                Msg().out("Warning: running as another user not supported")
+                LOG.warning("running as another user not supported")
                 return False
 
             if self._has_option("--fakeroot", "exec"):
@@ -134,7 +142,6 @@ class SingularityEngine(ExecutionEngineCommon):
           * argument: container_id or name
           * options:  many via self.opt see the help
         """
-
         if os.path.isdir(
                 FileBind(self.localrepo, container_id).container_orig_dir):
             FileBind(self.localrepo, container_id).restore() # legacy 1.1.3
@@ -144,25 +151,17 @@ class SingularityEngine(ExecutionEngineCommon):
             "/etc/resolv.conf", "/etc/host.conf",
             "/lib/modules",
         )
-
-        # setup execution
-        exec_path = self._run_init(container_id)
[…]
