### 2021-07-21 3351479
implement logging in elfpatcher
