+++ udocker/helper/elfpatcher.py
@@ -6,8 +6,7 @@ import os
 import sys
 import time
 
-from udocker import is_genstr
-from udocker.msg import Msg
+from udocker import is_genstr, LOG
 from udocker.config import Config
 from udocker.utils.uprocess import Uprocess
 from udocker.utils.fileutil import FileUtil
@@ -30,6 +29,7 @@ class ElfPatcher(object):
             os.path.realpath(self.localrepo.cd_container(container_id))
         if not self._container_dir:
             raise ValueError("invalid container id")
+
         self._container_root = self._container_dir + "/ROOT"
         self._container_ld_so_path = self._container_dir + "/ld.so.path"
         self._container_ld_so_orig = self._container_dir + "/ld.so.orig"
@@ -55,7 +55,7 @@ class ElfPatcher(object):
         f_util = FileUtil(self.localrepo.bindir)
         patchelf_exec = f_util.find_file_in_dir(image_list)
         if not os.path.exists(patchelf_exec):
-            Msg().err("Error: patchelf executable not found")
+            LOG.error("patchelf executable not found")
             sys.exit(1)
 
         return patchelf_exec
@@ -67,6 +67,7 @@ class ElfPatcher(object):
             if "#f" in arg:
                 arg = arg.replace("#f", path)
             cmd_out.append(arg)
+
         return cmd_out
 
     def _walk_fs(self, cmd, root_path, action=BIN):
@@ -118,15 +119,18 @@ class ElfPatcher(object):
                                        self.ONE_OUTPUT | self.BIN)
             if elf_loader and ".so" in elf_loader:
                 return elf_loader
+
         return ""
 
     def get_original_loader(self):
         """Get the pathname of the original ld.so"""
         if os.path.exists(self._container_ld_so_path):
             return FileUtil(self._container_ld_so_path).getdata('r').strip()
+
         elf_loader = self.guess_elf_loader()
         if elf_loader:
             FileUtil(self._container_ld_so_path).putdata(elf_loader, 'w')
+
         return elf_loader
 
     def get_container_loader(self):
@@ -134,6 +138,7 @@ class ElfPatcher(object):
         elf_loader = self.get_original_loader()
         if not elf_loader:
             return ""
+
         elf_loader = self._container_root + "/" + elf_loader
         return elf_loader if os.path.exists(elf_loader) else ""
 
@@ -142,6 +147,7 @@ class ElfPatcher(object):
         last_path = FileUtil(self._container_patch_path).getdata('r')
         if last_path and is_genstr(last_path):
             return last_path.strip()
+
         return ""
 
     def check_container_path(self):
@@ -149,6 +155,7 @@ class ElfPatcher(object):
         last_path = self.get_patch_last_path()
         if last_path and last_path != self._container_dir:
             return False
+
         return True
 
     def get_patch_last_time(self):
@@ -163,6 +170,7 @@ class ElfPatcher(object):
         """Set all executables and libs to the ld.so absolute pathname"""
         if not self.check_container_path():
             self.restore_binaries()
+
         last_time = '0'
         patchelf_exec = self.select_patchelf()
         elf_loader = self.get_container_loader()
@@ -174,10 +182,12 @@ class ElfPatcher(object):
                 last_time = str(int(time.time()))
             except ValueError:
                 pass
+
             futil_time = FileUtil(self._container_patch_time)
             futil_path = FileUtil(self._container_patch_path)
             return (futil_time.putdata(last_time, 'w') and
                     futil_path.putdata(self._container_dir, 'w'))
+
         return False
 
     def restore_binaries(self):
@@ -191,11 +201,13 @@ class ElfPatcher(object):
         else:
             cmd = [patchelf_exec, "--restore-root-prefix",
                    self._container_root, "#f"]
+
[…]
