### 2021-07-21 9e54367
add some logging to hostinfo
