### 2021-07-21 87c240d
add logging to keystore
