### 2021-07-21 eecfd34
add logging to nixauth
