### 2021-07-21 8fedbaa
add logging to osinfo
