### 2021-07-21 9701499
add logging to unique
