### 2021-07-21 6a6b05f
implement logging in unshare
