### 2021-07-21 dd74f53
add logging to checksum
