+++ udocker/utils/chksum.py
@@ -3,6 +3,7 @@
 
 import re
 
+from udocker import LOG
 from udocker.utils.uprocess import Uprocess
 
 try:
@@ -20,6 +21,7 @@ class ChkSUM(object):
             self._algorithms["sha256"] = self._hashlib_sha256
         except NameError:
             self._algorithms["sha256"] = self._openssl_sha256
+
         try:
             dummy = hashlib.sha512()
             self._algorithms["sha512"] = self._hashlib_sha512
@@ -38,10 +40,12 @@ class ChkSUM(object):
 
     def _hashlib_sha256(self, filename):
         """sha256 calculation using hashlib"""
+        LOG.info("sha256 calculation using hashlib")
         return self._hashlib(hashlib.sha256(), filename)
 
     def _hashlib_sha512(self, filename):
         """sha512 calculation using hashlib"""
+        LOG.info("sha512 calculation using hashlib")
         return self._hashlib(hashlib.sha512(), filename)
 
     def _openssl(self, algorithm, filename):
@@ -50,9 +54,11 @@ class ChkSUM(object):
         output = Uprocess().get_output(cmd)
         if output is None:
             return ""
+
         match = re.match("^(\\S+) ", output)
         if match:
             return match.group(1)
+
         return ""
 
     def _openssl_sha256(self, filename):
@@ -75,4 +81,5 @@ class ChkSUM(object):
         """Compute hash algorithm for file"""
         if algorithm in self._algorithms:
             return self._algorithms[algorithm](filename)
+
         return ""
