### 2021-07-21 60aac1d
implement logging in curl
