+++ udocker/utils/curl.py
@@ -4,10 +4,10 @@
 import os
 import sys
 import json
+import logging
 
-from udocker import is_genstr
+from udocker import is_genstr, LOG
 from udocker.config import Config
-from udocker.msg import Msg
 from udocker.utils.fileutil import FileUtil
 from udocker.utils.uprocess import Uprocess
 
@@ -43,6 +43,7 @@ class CurlHeader(object):
         """Write is called by Curl()"""
         if not isinstance(buff, str):
             buff = buff.decode()
+
         pair = buff.split(":", 1)
         if len(pair) == 2:
             key = pair[0].strip().lower()
@@ -54,6 +55,7 @@ class CurlHeader(object):
               pair[0].strip() == "" and
               "location" not in self.data):
             return -1
+
         return None
 
     def setvalue_from_file(self, in_filename):
@@ -65,8 +67,10 @@ class CurlHeader(object):
             infile = open(in_filename, 'r')
         except (IOError, OSError):
             return False
+
         for line in infile:
             self.write(line)
+
         infile.close()
         return True
 
@@ -101,13 +105,12 @@ class GetURL(object):
         if GetURLpyCurl().is_available() and not self._curl_exec:
             self._geturl = GetURLpyCurl()
             self.cache_support = True
-            Msg().out("Info: using pycurl", l=Msg.DBG)
+            LOG.debug("using pycurl")
         elif GetURLexeCurl().is_available():
             self._geturl = GetURLexeCurl()
-            Msg().out("Info: using curl executable", self._geturl._curl_exec,
-                      l=Msg.DBG)
+            LOG.debug("using curl executable: %s", self._geturl._curl_exec)
         else:
-            Msg().err("Error: need curl or pycurl to perform downloads")
+            LOG.error("need curl or pycurl to perform downloads")
             raise NameError('need curl or pycurl')
 
     def get_content_length(self, hdr):
@@ -134,12 +137,14 @@ class GetURL(object):
         """
         if len(args) != 1:
             raise TypeError('wrong number of arguments')
+
         return self._geturl.get(*args, **kwargs)
 
     def post(self, *args, **kwargs):
         """POST using selected implementation"""
         if len(args) != 2:
             raise TypeError('wrong number of arguments')
+
         kwargs["post"] = args[1]
         return self._geturl.get(args[0], **kwargs)
 
@@ -169,6 +174,7 @@ class GetURLpyCurl(GetURL):
             dummy = pycurl.Curl()
         except (NameError, AttributeError):
             return False
+
         return True
 
     def _select_implementation(self):
@@ -189,10 +195,11 @@ class GetURLpyCurl(GetURL):
         pyc.setopt(pyc.CONNECTTIMEOUT, self.ctimeout)
         pyc.setopt(pyc.TIMEOUT, self.timeout)
         pyc.setopt(pyc.PROXY, self.http_proxy)
-        if Msg.level >= Msg.VER:
+        if Config.conf['verbose_level'] == logging.DEBUG:
             pyc.setopt(pyc.VERBOSE, True)
         else:
             pyc.setopt(pyc.VERBOSE, False)
+
         self._url = ""
 
     def _mkpycurl(self, pyc, hdr, buf, *args, **kwargs):
@@ -201,32 +208,44 @@ class GetURLpyCurl(GetURL):
         pyc.setopt(pycurl.URL, self._url)
         if "follow" in kwargs:
             pyc.setopt(pyc.FOLLOWLOCATION, kwargs["follow"])
+
         if "post" in kwargs:
             pyc.setopt(pyc.POST, 1)
             pyc.setopt(pyc.HTTPHEADER, ['Content-Type: application/json'])
             pyc.setopt(pyc.POSTFIELDS, json.dumps(kwargs["post"]))
+
[…]
