### 2021-07-21 4911588
implement logging in filebind
