+++ udocker/utils/filebind.py
@@ -3,7 +3,7 @@
 
 import os
 
-from udocker.msg import Msg
+from udocker import LOG
 from udocker.utils.fileutil import FileUtil
 
 
@@ -12,7 +12,6 @@ class FileBind(object):
     a container when binding of files is not possible such as when
     using rootless namespaces.
     """
-
     bind_dir = "/.bind_host_files"
     orig_dir = "/.bind_orig_files"
 
@@ -30,12 +29,14 @@ class FileBind(object):
         """Prepare container for FileBind"""
         if not os.path.isdir(self.container_orig_dir):
             if not FileUtil(self.container_orig_dir).mkdir():
-                Msg().err("Error: creating dir:", self.container_orig_dir)
+                LOG.error("creating dir: %s", self.container_orig_dir)
                 return False
+
         if not os.path.isdir(self.container_bind_dir):
             if not FileUtil(self.container_bind_dir).mkdir():
-                Msg().err("Error: creating dir:", self.container_bind_dir)
+                LOG.error("creating dir: %s", self.container_bind_dir)
                 return False
+
         return True
 
     def restore(self, force=False):
@@ -43,10 +44,12 @@ class FileBind(object):
         error = False
         if not os.path.isdir(self.container_orig_dir):
             return
+
         for f_name in os.listdir(self.container_orig_dir):
             orig_file = self.container_orig_dir + '/' + f_name
             if not os.path.isfile(orig_file):
                 continue
+
             cont_file = os.path.basename(f_name).replace('#', '/')
             cont_file = self.container_root + '/' + cont_file
             if os.path.islink(cont_file):
@@ -54,11 +57,14 @@ class FileBind(object):
                 FileUtil(cont_file).remove()
             elif os.path.exists(cont_file):
                 continue
+
             if not FileUtil(orig_file).rename(cont_file):
-                Msg().err("Error: restoring binded file:", cont_file)
+                LOG.error("restoring binded file: %s", cont_file)
                 error = True
+
         if not error or force:
             FileUtil(self.container_orig_dir).remove(recursive=True)
+
         FileUtil(self.container_bind_dir).remove(recursive=True)
 
     def start(self, files_list=None):
@@ -71,6 +77,7 @@ class FileBind(object):
         self.host_bind_dir = FileUtil("BIND_FILES").mktmpdir()
         if files_list:
             self.set_list(files_list)
+
         return (self.host_bind_dir, self.bind_dir)
 
     def set_list(self, files_list):
@@ -82,6 +89,7 @@ class FileBind(object):
         """Prepare individual file mapping"""
         if not os.path.isfile(host_file):
             return
+
         orig_file = cont_file.replace('/', '#')
         orig_file_path = self.container_orig_dir + '/' + orig_file
         cont_file_path = self.container_root + '/' + cont_file
@@ -91,7 +99,9 @@ class FileBind(object):
                 os.rename(cont_file_path, orig_file_path)
             else:
                 return
+
             os.symlink(link_path, cont_file_path)
+
         FileUtil(orig_file_path).copyto(self.host_bind_dir)
 
     def add_file(self, host_file, cont_file):
