### 2021-07-21 fd4bf41
implement logging in fileutil
