+++ udocker/utils/fileutil.py
@@ -5,9 +5,9 @@ import os
 import sys
 import stat
 import re
+import logging
 
-from udocker import is_genstr
-from udocker.msg import Msg
+from udocker import is_genstr, LOG
 from udocker.config import Config
 from udocker.helper.unique import Unique
 from udocker.helper.hostinfo import HostInfo
@@ -29,12 +29,14 @@ class FileUtil(object):
             self.filename = "-"
             self.basename = "-"
             return
+
         try:
             self.filename = os.path.abspath(filename)
             self.basename = os.path.basename(self.filename)
         except (AttributeError, TypeError):
             self.filename = filename
             self.basename = filename
+
         self._register_prefix(self._tmpdir)
 
     def _register_prefix(self, prefix):
@@ -60,6 +62,7 @@ class FileUtil(object):
                 old_umask = os.umask(new_umask)
             except (TypeError, ValueError):
                 return False
+
             if FileUtil.orig_umask is None:
                 FileUtil.orig_umask = old_umask
         else:
@@ -67,6 +70,7 @@ class FileUtil(object):
                 os.umask(FileUtil.orig_umask)
             except (TypeError, ValueError):
                 return False
+
         return True
 
     def mktmp(self):
@@ -85,6 +89,7 @@ class FileUtil(object):
             os.makedirs(self.filename)
         except (OSError, IOError, AttributeError):
             return False
+
         return True
 
     def rmdir(self):
@@ -93,6 +98,7 @@ class FileUtil(object):
             os.rmdir(self.filename)
         except (OSError, IOError, AttributeError):
             return False
+
         return True
 
     def mktmpdir(self):
@@ -100,6 +106,7 @@ class FileUtil(object):
         dirname = self.mktmp()
         if FileUtil(dirname).mkdir():
             return dirname
+
         return None
 
     def uid(self):
@@ -114,11 +121,14 @@ class FileUtil(object):
         filename = os.path.realpath(filename)
         if os.path.isdir(filename):
             filename += '/'
+
         for safe_prefix in FileUtil.safe_prefixes:
             if filename.startswith(safe_prefix):
                 return True
+
             if filename.startswith(os.path.realpath(safe_prefix)):
                 return True
+
         return False
 
     def chown(self, uid=0, gid=0, recursive=False):
@@ -132,6 +142,7 @@ class FileUtil(object):
                 os.lchown(self.filename, uid, gid)
         except OSError:
             return False
+
         return True
 
     def rchown(self, uid=0, gid=0):
@@ -154,7 +165,7 @@ class FileUtil(object):
                 mode = (stat.S_IMODE(filestat) & mask) | filemode
                 os.chmod(filename, mode)
         except OSError:
-            Msg().err("Error: changing permissions of:", filename, l=Msg.VER)
+            LOG.error("changing permissions of: %s", filename)
             raise OSError
 
     def chmod(self, filemode=0o600, dirmode=0o700, mask=0o755, recursive=False):
@@ -165,12 +176,15 @@ class FileUtil(object):
                     for f_name in files:
                         self._chmod(dir_path + '/' + f_name,
                                     filemode, None, mask)
+
                     for f_name in dirs:
                         self._chmod(dir_path + '/' + f_name,
                                     None, dirmode, mask)
+
             self._chmod(self.filename, filemode, dirmode, mask)
         except OSError:
             return False
+
         return True
 
     def rchmod(self, filemode=0o600, dirmode=0o700, mask=0o755):
@@ -187,18 +201,22 @@ class FileUtil(object):
                     if not os.path.islink(f_path):
                         os.chmod(f_path, stat.S_IWUSR | stat.S_IRUSR)
                     os.unlink(f_path)
+
                 for f_name in dirs:
[…]
