### 2021-07-21 e66f864
implement logging in mountpoint
