### 2021-07-21 9b3b694
added logging to uenv
