+++ udocker/utils/uenv.py
@@ -5,28 +5,28 @@ import os
 import sys
 import string
 
-from udocker import is_genstr
+from udocker import is_genstr, LOG
 
 def get_pair(envstr):
     """Split env=var into key and val"""
     if not (is_genstr(envstr) and envstr):
         return ("", "")
+
     if '=' in envstr:
         try:
             (key, val) = envstr.split('=', 1)
         except (ValueError, NameError, AttributeError):
             return ("", "")
+
         key = key.strip()
         val = val.strip()
-        #for quote in ("'", '"'):
-        #    if quote == val[0]:
-        #        val = val.strip(quote)
-        #        break
     else:
         key = envstr.strip()
         val = os.getenv(envstr, "")
+
     if not key or ' ' in key or key[0] in string.digits:
         return ("", "")
+
     return (key, val)
 
 
@@ -40,6 +40,7 @@ class UenvIterator(object):
             self._keys = list(uenv.env.keys())
         else:
             self._keys = uenv.env.keys()
+
         self._index = 0
 
     def __next__(self):
@@ -49,6 +50,7 @@ class UenvIterator(object):
             value = self._uenv.env[key]
             self._index += 1
             return (key, value)
+
         raise StopIteration
 
     next = __next__ # Python 2
@@ -73,9 +75,11 @@ class Uenv(object):
 
     def append(self, envstr):
         """Add string with key=val to Uenv"""
+        LOG.debug("adding key=val to Uenv: %s", envstr)
         (key, val) = get_pair(envstr)
         if key:
             self.env[key] = val
+
         return self
 
     def appendif(self, envstr):
@@ -83,6 +87,8 @@ class Uenv(object):
         (key, val) = get_pair(envstr)
         if key and key not in self.env:
             self.env[key] = val
+            LOG.debug("added key=val: %s", envstr)
+
         return self
 
     def extend(self, envlist):
@@ -93,6 +99,7 @@ class Uenv(object):
         elif isinstance(envlist, (dict)):
             for key in envlist.keys():
                 self.env[key] = envlist[key]
+
         return self
 
     def extendif(self, envlist):
@@ -104,6 +111,7 @@ class Uenv(object):
             for key in envlist.keys():
                 if key not in self.env:
                     self.env[key] = envlist[key]
+
         return self
 
     def add(self, thisenv):
@@ -112,6 +120,7 @@ class Uenv(object):
             self.append(thisenv)
         elif isinstance(thisenv, (list)):
             self.extend(thisenv)
+
         return self
 
     def getenv(self, key):
@@ -138,6 +147,7 @@ class Uenv(object):
         env_list = []
         for (key, val) in self.env.items():
             env_list.append('%s=%s' % (key, val))
+
         return env_list
 
     def dict(self):
