### 2021-07-21 2bf91bb
add logging to uprocess
