+++ udocker/utils/uprocess.py
@@ -5,8 +5,7 @@ import os
 import sys
 import subprocess
 
-from udocker import is_genstr
-from udocker.msg import Msg
+from udocker import is_genstr, LOG
 from udocker.config import Config
 
 
@@ -17,17 +16,23 @@ class Uprocess(object):
         """Find file in a path set such as PATH=/usr/bin:/bin"""
         if not (filename and path):
             return ""
+
         basename = os.path.basename(filename)
+        LOG.debug("find file in path: %s", basename)
         if is_genstr(path):
             if "=" in path:
                 path = "".join(path.split("=", 1)[1:])
+
             path = path.split(":")
+
         if isinstance(path, (list, tuple)):
             for directory in path:
                 full_path = rootdir + directory + "/" + basename
                 if os.path.lexists(full_path):
                     return directory + "/" + basename
+
             return ""
+
         return ""
 
     def _check_output(self, *popenargs, **kwargs):
@@ -39,7 +44,9 @@ class Uprocess(object):
             cmd = kwargs.get("args")
             if cmd is None:
                 cmd = popenargs[0]
+
             raise subprocess.CalledProcessError(retcode, cmd)
+
         return output
 
     def check_output(self, *popenargs, **kwargs):
@@ -55,6 +62,7 @@ class Uprocess(object):
             # if Python < 2.7
             else:
                 chk_out = self._check_output(*popenargs, **kwargs)
+
         except OSError:
             return ""
 
@@ -67,20 +75,24 @@ class Uprocess(object):
             cmd_path = self.find_inpath(cmd[0], path)
             if cmd_path:
                 cmd[0] = cmd_path
+
         content = ""
         try:
-            content = self.check_output(cmd, shell=False, stderr=Msg.chlderr,
+            content = self.check_output(cmd, shell=False, stderr=sys.stderr,
                                         close_fds=True)
         except subprocess.CalledProcessError:
             if not ignore_error:
                 return None
+
         return content.strip()
 
     def call(self, cmd, **kwargs):
         """Execute one shell command"""
+        LOG.debug("executing: %s", cmd)
         if not cmd[0].startswith("/"):
             path = Config.conf["root_path"] + ":" + os.getenv("PATH", "")
             cmd[0] = self.find_inpath(cmd[0], path)
+
         kwargs["shell"] = False
         return subprocess.call(cmd, **kwargs)
 
@@ -89,20 +101,25 @@ class Uprocess(object):
         path = Config.conf["root_path"] + ":" + os.getenv("PATH", "")
         if not cmd1[0].startswith("/"):
             cmd1[0] = self.find_inpath(cmd1[0], path)
+
         if not cmd2[0].startswith("/"):
             cmd2[0] = self.find_inpath(cmd2[0], path)
+
         try:
-            proc_1 = subprocess.Popen(cmd1, stderr=Msg.chlderr, shell=False,
+            proc_1 = subprocess.Popen(cmd1, stderr=sys.stderr, shell=False,
                                       stdout=subprocess.PIPE, **kwargs)
         except (OSError, ValueError):
             return False
+
         try:
-            proc_2 = subprocess.Popen(cmd2, stderr=Msg.chlderr, shell=False,
+            proc_2 = subprocess.Popen(cmd2, stderr=sys.stderr, shell=False,
                                       stdin=proc_1.stdout)
         except (OSError, ValueError):
             proc_1.kill()
             return False
+
         while proc_1.returncode is None or proc_2.returncode is None:
             proc_1.wait()
             proc_2.wait()
+
         return not (proc_1.returncode or proc_2.returncode)
