### 2021-07-21 2cc7bef
fix some logging levels
