+++ udocker/container/localrepo.py
@@ -719,11 +719,11 @@ class LocalRepository(object):
                     elif ':' in fname:
                         structure["repolayers"][layer_id]["layer_f"] = f_path
                     else:
-                        LOG.warning("unkwnon file in layer: %s", f_path)
+                        LOG.warning("unknown file in layer: %s", f_path)
                 elif fname in ("TAG", "v1", "v2", "PROTECT", "container.json"):
                     pass
 
-                LOG.warning("unkwnon file in image: %s", f_path)
+                LOG.warning("unknown file in image: %s", f_path)
 
         return structure
 
