+++ udocker/utils/uprocess.py
@@ -4,6 +4,7 @@
 import os
 import sys
 import subprocess
+import logging
 
 from udocker import is_genstr, LOG
 from udocker.config import Config
@@ -11,6 +12,13 @@ from udocker.config import Config
 
 class Uprocess(object):
     """Provide alternative implementations for subprocess"""
+
+    def _get_stderr(self):
+        stderror = None
+        if Config.conf['verbose_level'] == logging.DEBUG:
+            stderror = sys.stderr
+
+        return stderror
 
     def find_inpath(self, filename, path, rootdir=""):
         """Find file in a path set such as PATH=/usr/bin:/bin"""
@@ -70,6 +78,10 @@ class Uprocess(object):
 
     def get_output(self, cmd, ignore_error=False):
         """Execute a shell command and get its output"""
+        stderror = None
+        if Config.conf['verbose_level'] == logging.DEBUG:
+            stderror = sys.stderr
+
         if not cmd[0].startswith("/"):
             path = Config.conf["root_path"] + ":" + os.getenv("PATH", "")
             cmd_path = self.find_inpath(cmd[0], path)
@@ -78,7 +90,8 @@ class Uprocess(object):
 
         content = ""
         try:
-            content = self.check_output(cmd, shell=False, stderr=sys.stderr,
+            content = self.check_output(cmd, shell=False,
+                                        stderr=stderror,
                                         close_fds=True)
         except subprocess.CalledProcessError:
             if not ignore_error:
@@ -98,6 +111,10 @@ class Uprocess(object):
 
     def pipe(self, cmd1, cmd2, **kwargs):
         """Pipe two shell commands"""
+        stderror = None
+        if Config.conf['verbose_level'] == logging.DEBUG:
+            stderror = sys.stderr
+
         path = Config.conf["root_path"] + ":" + os.getenv("PATH", "")
         if not cmd1[0].startswith("/"):
             cmd1[0] = self.find_inpath(cmd1[0], path)
@@ -106,14 +123,15 @@ class Uprocess(object):
             cmd2[0] = self.find_inpath(cmd2[0], path)
 
         try:
-            proc_1 = subprocess.Popen(cmd1, stderr=sys.stderr, shell=False,
+            proc_1 = subprocess.Popen(cmd1, stderr=stderror,
+                                      shell=False,
                                       stdout=subprocess.PIPE, **kwargs)
         except (OSError, ValueError):
             return False
 
         try:
-            proc_2 = subprocess.Popen(cmd2, stderr=sys.stderr, shell=False,
-                                      stdin=proc_1.stdout)
+            proc_2 = subprocess.Popen(cmd2, stderr=stderror,
+                                      shell=False, stdin=proc_1.stdout)
         except (OSError, ValueError):
             proc_1.kill()
             return False
