### 2021-07-22 68b1c32
wip on stderr of subrocess
