### 2021-07-23 0f701aa
add MSG ContainerID as env variable
