+++ utils/udocker_test.sh
@@ -169,8 +169,9 @@ udocker inspect ${DOCKER_IMG}; return=$?
 result
 
 STRING="T020: udocker create ${DOCKER_IMG}"
-CONT_ID=`udocker create ${DOCKER_IMG}`; return=$?
-echo "ContainerID = ${CONT_ID}"
+export `udocker create ${DOCKER_IMG}`; return=$?
+CONT_ID=${ContainerID}
+echo "CONT_ID = ${CONT_ID}"
 result
 
 STRING="T021: udocker create --name=${CONT} ${DOCKER_IMG}"
