### 2021-07-23 8ccc004
remove and refactor chksum solves issue #347
