### 2021-07-23 ea14819
update doc remove openssl ref
