### 2021-07-23 0832f74
fix log, test in non default udocker_dir
