+++ utils/udocker_test.sh
@@ -32,7 +32,7 @@ THIS_SCRIPT_NAME=$( basename "$0" )
 
 # Variables for the tests
 declare -a FAILED_TESTS
-DEFAULT_UDIR=$HOME/.udocker
+DEFAULT_UDIR=$HOME/.udocker-tests
 TEST_UDIR=$HOME/.udocker-test-h45y7k9X
 TAR_IMAGE="centos7.tar"
 TAR_CONT="centos7-cont.tar"
@@ -40,6 +40,7 @@ TAR_IMAGE_URL="https://download.ncg.ingrid.pt/webdav/udocker_test/${TAR_IMAGE}"
 TAR_CONT_URL="https://download.ncg.ingrid.pt/webdav/udocker_test/${TAR_CONT}"
 DOCKER_IMG="ubuntu:18.04"
 CONT="ubuntu"
+export UDOCKER_DIR=${DEFAULT_UDIR}
 
 function print_ok
 {
