+++ tests/unit/test_cli.py
@@ -5,6 +5,7 @@
 udocker unit tests: UdockerCLI
 """
 
+import logging
 from unittest import TestCase, main
 from unittest.mock import Mock, patch
 from udocker.config import Config
@@ -23,14 +24,14 @@ class UdockerCLITestCase(TestCase):
         Config().conf['hostauth_list'] = ("/etc/passwd", "/etc/group")
         Config().conf['cmd'] = "/bin/bash"
         Config().conf['cpu_affinity_exec_tools'] = \
-            (["numactl", "-C", "%s", "--", ],
-             ["taskset", "-c", "%s", ])
+            (["numactl", "-C", "%s", "--", ], ["taskset", "-c", "%s", ])
         Config().conf['valid_host_env'] = "HOME"
         Config().conf['username'] = "user"
         Config().conf['userhome'] = "/"
         Config().conf['oskernel'] = "4.8.13"
         Config().conf['location'] = ""
         Config().conf['keystore'] = "KEYSTORE"
+        Config().conf['verbose_level'] = logging.NOTSET
 
         str_local = 'udocker.container.localrepo.LocalRepository'
         self.lrepo = patch(str_local)
@@ -89,10 +90,8 @@ class UdockerCLITestCase(TestCase):
         self.assertTrue(self.local.setup.called)
 
     @patch('udocker.cli.DockerIoAPI.is_repo_name')
-    @patch('udocker.cli.Msg')
-    def test_03__check_imagespec(self, mock_msg, mock_reponame):
+    def test_03__check_imagespec(self, mock_reponame):
         """Test03 UdockerCLI()._check_imagespec()."""
-        mock_msg.level = 0
         mock_reponame.return_value = False
         udoc = UdockerCLI(self.local)
         status = udoc._check_imagespec("")
@@ -109,10 +108,8 @@ class UdockerCLITestCase(TestCase):
         self.assertEqual(status, ("AAA", "45"))
 
     @patch('udocker.cli.DockerIoAPI.is_repo_name')
-    @patch('udocker.cli.Msg')
-    def test_04__check_imagerepo(self, mock_msg, mock_reponame):
+    def test_04__check_imagerepo(self, mock_reponame):
         """Test04 UdockerCLI()._check_imagerepo()."""
-        mock_msg.level = 0
         mock_reponame.return_value = False
         udoc = UdockerCLI(self.local)
         status = udoc._check_imagerepo("")
@@ -126,11 +123,8 @@ class UdockerCLITestCase(TestCase):
     @patch('udocker.cli.DockerIoAPI.set_index')
     @patch('udocker.cli.DockerIoAPI.set_registry')
     @patch('udocker.cli.DockerIoAPI.set_proxy')
-    @patch('udocker.cli.Msg')
-    def test_05__set_repository(self, mock_msg, mock_proxy,
-                                mock_reg, mock_idx):
+    def test_05__set_repository(self, mock_proxy, mock_reg, mock_idx):
         """Test05 UdockerCLI()._set_repository()."""
-        mock_msg.level = 0
         regist = "registry.io"
         idxurl = "dockerhub.io"
         imgrepo = "dockerhub.io/myimg:1.2"
@@ -175,11 +169,8 @@ class UdockerCLITestCase(TestCase):
         self.assertEqual(status, res)
 
     @patch('udocker.cli.os.path.exists')
-    @patch('udocker.cli.Msg')
-    def test_07_do_mkrepo(self, mock_msg, mock_exists):
+    def test_07_do_mkrepo(self, mock_exists):
         """Test07 UdockerCLI().do_mkrepo()."""
-        mock_msg.level = 0
-
[…]
