### 2021-07-24 07c9c39
setlevel of logging in unit tests
