### 2021-07-24 dda812d
LOG in config tests
