### 2021-07-24 9ceeefe
LOG in containerstructure tests
