+++ tests/unit/test_containerstructure.py
@@ -7,7 +7,7 @@ udocker unit tests: ContainerStructure
 
 from unittest import TestCase, main
 from unittest.mock import patch, Mock
-from udocker.container.structure import ContainerStructure
+from udocker.container.structure import ContainerStructure, LOG
 from udocker.config import Config
 
 
@@ -15,6 +15,7 @@ class ContainerStructureTestCase(TestCase):
     """Test ContainerStructure() class for containers structure."""
 
     def setUp(self):
+        LOG.setLevel(100)
         Config().getconf()
         Config().conf['hostauth_list'] = ("/etc/passwd", "/etc/group")
         Config().conf['cmd'] = "/bin/bash"
@@ -48,10 +49,8 @@ class ContainerStructureTestCase(TestCase):
         self.assertEqual(prex.imagerepo, "")
         self.assertEqual(prex.container_id, "123456")
 
-    @patch('udocker.container.structure.Msg')
-    def test_02_get_container_attr(self, mock_msg):
+    def test_02_get_container_attr(self):
         """Test02 ContainerStructure().get_container_attr()."""
-        mock_msg.return_value.level.return_value = 0
         Config().conf['location'] = "/"
         prex = ContainerStructure(self.local)
         (container_dir, container_json) = prex.get_container_attr()
@@ -64,7 +63,6 @@ class ContainerStructureTestCase(TestCase):
         (container_dir, container_json) = prex.get_container_attr()
         self.assertEqual(container_dir, False)
         self.assertEqual(container_json, False)
-        self.assertTrue(mock_msg.return_value.err.called)
 
         Config().conf['location'] = ""
         self.local.cd_container.return_value = "/"
@@ -73,7 +71,6 @@ class ContainerStructureTestCase(TestCase):
         (container_dir, container_json) = prex.get_container_attr()
         self.assertEqual(container_dir, False)
         self.assertEqual(container_json, False)
-        self.assertTrue(mock_msg.return_value.err.called)
 
         Config().conf['location'] = ""
         self.local.cd_container.return_value = "/"
@@ -165,10 +162,8 @@ class ContainerStructureTestCase(TestCase):
 
     @patch.object(ContainerStructure, '_untar_layers')
     @patch('udocker.container.structure.Unique.uuid')
-    @patch('udocker.container.structure.Msg')
-    def test_07_create_fromimage(self, mock_msg, mock_uuid, mock_untar):
+    def test_07_create_fromimage(self, mock_uuid, mock_untar):
         """Test07 ContainerStructure().create_fromimage()."""
-        mock_msg.return_value.level.return_value = 0
         self.local.cd_imagerepo.return_value = ""
         prex = ContainerStructure(self.local)
         status = prex.create_fromimage("imagerepo", "tag")
@@ -199,13 +194,10 @@ class ContainerStructureTestCase(TestCase):
     @patch.object(ContainerStructure, '_chk_container_root')
     @patch.object(ContainerStructure, '_untar_layers')
     @patch('udocker.container.structure.Unique.uuid')
-    @patch('udocker.container.structure.Msg')
-    def test_08_create_fromlayer(self, mock_msg, mock_uuid,
-                                 mock_untar, mock_chkcont):
+    def test_08_create_fromlayer(self, mock_uuid, mock_untar, mock_chkcont):
         """Test08 ContainerStructure().create_fromlayer()."""
         # Empty container_json
         cont_json = dict()
-        mock_msg.return_value.level.return_value = 0
         mock_uuid.return_value = "123456"
         prex = ContainerStructure(self.local)
         status = prex.create_fromlayer("imagerepo", "tag", "layer", cont_json)
[…]
