### 2021-07-24 c8834c0
LOG in curl tests, delete duplicate
