+++ tests/unit/test_curl.py
@@ -8,20 +8,18 @@ udocker unit tests: CurlHeader
 from unittest import TestCase, main
 from unittest.mock import patch
 from io import StringIO
-from udocker.utils.curl import CurlHeader
+from udocker.utils.curl import CurlHeader, LOG
 from udocker.utils.curl import GetURL
 from udocker.utils.curl import GetURLpyCurl
 from udocker.utils.curl import GetURLexeCurl
 from udocker.config import Config
-
-BUILTINS = "builtins"
 
 
 class CurlHeaderTestCase(TestCase):
     """Test CurlHeader() http header parser."""
 
     def setUp(self):
-        pass
+        LOG.setLevel(100)
 
     def tearDown(self):
         pass
@@ -62,9 +60,8 @@ class CurlHeaderTestCase(TestCase):
     def test_03_setvalue_from_file(self, mock_write):
         """Test03 CurlHeader().setvalue_from_file()."""
         fakedata = StringIO('XXXX')
-        with patch(BUILTINS + '.open') as mopen:
-            mopen.return_value.__iter__ = (
-                lambda self: iter(fakedata.readline, ''))
+        with patch('builtins.open') as mopen:
+            mopen.return_value = fakedata
             curl_header = CurlHeader()
             self.assertTrue(curl_header.setvalue_from_file("filename"))
             self.assertTrue(mock_write.called_with('XXXX'))
@@ -81,6 +78,7 @@ class GetURLTestCase(TestCase):
     """Test GetURL() perform http operations portably."""
 
     def setUp(self):
+        LOG.setLevel(100)
         Config().getconf()
         Config().conf['timeout'] = 1
         Config().conf['ctimeout'] = 1
@@ -100,11 +98,8 @@ class GetURLTestCase(TestCase):
     @patch.object(GetURLexeCurl, '_select_implementation')
     @patch.object(GetURLexeCurl, 'is_available')
     @patch.object(GetURLpyCurl, 'is_available')
-    @patch('udocker.utils.curl.Msg')
-    def test_01_init(self, mock_msg, mock_gupycurl,
-                     mock_guexecurl, mock_select):
+    def test_01_init(self, mock_gupycurl, mock_guexecurl, mock_select):
         """Test01 GetURL() constructor."""
-        mock_msg.level = 0
         mock_gupycurl.return_value = False
         mock_guexecurl.return_value = True
         geturl = GetURL()
@@ -113,13 +108,10 @@ class GetURLTestCase(TestCase):
         self.assertEqual(geturl.insecure, Config().conf['http_insecure'])
         self.assertFalse(geturl.cache_support)
 
-    @patch('udocker.utils.curl.Msg')
     @patch.object(GetURLexeCurl, 'is_available')
     @patch.object(GetURLpyCurl, 'is_available')
-    def test_02__select_implementation(self, mock_gupycurl,
-                                       mock_guexecurl, mock_msg):
+    def test_02__select_implementation(self, mock_gupycurl, mock_guexecurl):
         """Test02 GetURL()._select_implementation()."""
-        mock_msg.level = 0
         mock_gupycurl.return_value = True
         geturl = GetURL()
         geturl._select_implementation()
@@ -188,14 +180,18 @@ class GetURLTestCase(TestCase):
         status = geturl.post("http://host", {"DATA": 1, })
         self.assertEqual(status, "http://host")
 
-    # def test_08_get_status_code(self):
-    #     """Test08 GetURL().get_status_code()."""
+    def test_08_get_status_code(self):
[…]
