### 2021-07-24 9ff0966
LOG in docker tests
