+++ tests/unit/test_dockerioapi.py
@@ -6,7 +6,7 @@ udocker unit tests: DockerIoAPI
 from unittest import TestCase, main
 from unittest.mock import patch, Mock
 from io import BytesIO as strio
-from udocker.docker import DockerIoAPI
+from udocker.docker import DockerIoAPI, LOG
 from udocker.config import Config
 
 
@@ -17,6 +17,7 @@ class DockerIoAPITestCase(TestCase):
     """
 
     def setUp(self):
+        LOG.setLevel(100)
         Config().getconf()
         str_local = 'udocker.container.localrepo.LocalRepository'
         self.lrepo = patch(str_local)
@@ -257,12 +258,10 @@ class DockerIoAPITestCase(TestCase):
         out = doia._get_v1_auth(www_authenticate)
         self.assertEqual(out, "Not Empty")
 
-    @patch('udocker.docker.Msg')
     @patch('udocker.utils.curl.CurlHeader')
     @patch.object(DockerIoAPI, '_get_url')
-    def test_13_get_v1_image_tags(self, mock_dgu, mock_hdr, mock_msg):
+    def test_13_get_v1_image_tags(self, mock_dgu, mock_hdr):
         """Test13 DockerIoAPI().get_v1_image_tags"""
-        mock_msg.level = 0
         mock_dgu.return_value = (mock_hdr, [])
         endpoint = "docker.io"
         imagerepo = "REPO"
@@ -270,12 +269,10 @@ class DockerIoAPITestCase(TestCase):
         out = doia.get_v1_image_tags(endpoint, imagerepo)
         self.assertIsInstance(out, list)
 
-    @patch('udocker.docker.Msg')
     @patch('udocker.utils.curl.CurlHeader')
     @patch.object(DockerIoAPI, '_get_url')
-    def test_14_get_v1_image_tag(self, mock_dgu, mock_hdr, mock_msg):
+    def test_14_get_v1_image_tag(self, mock_dgu, mock_hdr):
         """Test14 DockerIoAPI().get_v1_image_tag"""
-        mock_msg.level = 0
         mock_dgu.return_value = (mock_hdr, [])
         endpoint = "docker.io"
         imagerepo = "REPO"
@@ -284,12 +281,10 @@ class DockerIoAPITestCase(TestCase):
         out = doia.get_v1_image_tag(endpoint, imagerepo, tag)
         self.assertIsInstance(out, tuple)
 
-    @patch('udocker.docker.Msg')
     @patch('udocker.utils.curl.CurlHeader')
     @patch.object(DockerIoAPI, '_get_url')
-    def test_15_get_v1_image_ancestry(self, mock_dgu, mock_hdr, mock_msg):
+    def test_15_get_v1_image_ancestry(self, mock_dgu, mock_hdr):
         """Test15 DockerIoAPI().get_v1_image_ancestry"""
-        mock_msg.level = 0
         mock_dgu.return_value = (mock_hdr, [])
         endpoint = "docker.io"
         image_id = "IMAGEID"
@@ -297,11 +292,9 @@ class DockerIoAPITestCase(TestCase):
         out = doia.get_v1_image_ancestry(endpoint, image_id)
         self.assertIsInstance(out, tuple)
 
-    @patch('udocker.docker.Msg')
     @patch.object(DockerIoAPI, '_get_file')
-    def test_16_get_v1_image_json(self, mock_dgf, mock_msg):
+    def test_16_get_v1_image_json(self, mock_dgf):
         """Test16 DockerIoAPI().get_v1_image_json"""
-        mock_msg.level = 0
         mock_dgf.return_value = True
         endpoint = "docker.io"
         layer_id = "LAYERID"
[…]
