### 2021-07-24 76b65eb
LOG in elfpatcher tests
