### 2021-07-24 569b0c3
implement LOG in tests
