+++ tests/unit/test_execenginecommon.py
@@ -5,7 +5,7 @@ udocker unit tests: ExecutionEngineCommon
 
 from unittest import TestCase, main
 from unittest.mock import Mock, patch
-from udocker.engine.base import ExecutionEngineCommon
+from udocker.engine.base import ExecutionEngineCommon, LOG
 from udocker.utils.uenv import Uenv
 from udocker.config import Config
 
@@ -16,6 +16,7 @@ class ExecutionEngineCommonTestCase(TestCase):
     """
 
     def setUp(self):
+        LOG.setLevel(100)
         Config().getconf()
         Config().conf['hostauth_list'] = ("/etc/passwd", "/etc/group")
         Config().conf['cmd'] = "/bin/bash"
@@ -82,12 +83,10 @@ class ExecutionEngineCommonTestCase(TestCase):
         status = ex_eng._get_portsmap()
         self.assertEqual(status, {1024: 1024, 2048: 2048})
 
-    @patch('udocker.engine.base.Msg')
     @patch('udocker.engine.base.HostInfo')
     @patch.object(ExecutionEngineCommon, '_get_portsmap')
-    def test_04__check_exposed_ports(self, mock_getports, mock_hinfo, mock_msg):
+    def test_04__check_exposed_ports(self, mock_getports, mock_hinfo):
         """Test04 ExecutionEngineCommon()._check_exposed_ports()."""
-        mock_msg.level = 0
         mock_getports.return_value = {22: 22, 2048: 2048}
         mock_hinfo.uid = 1000
         ex_eng = ExecutionEngineCommon(self.local, self.xmode)
@@ -154,14 +153,12 @@ class ExecutionEngineCommonTestCase(TestCase):
         status = ex_eng._create_mountpoint(hpath, cpath, True)
         self.assertFalse(status)
 
-    @patch('udocker.engine.base.Msg')
     @patch.object(ExecutionEngineCommon, '_create_mountpoint')
     @patch('udocker.engine.base.os.path.exists')
     @patch('udocker.engine.base.Uvolume.split')
     def test_07__check_volumes(self, mock_uvolsplit, mock_exists,
-                               mock_crmpoint, mock_msg):
+                               mock_crmpoint):
         """Test07 ExecutionEngineCommon()._check_volumes()."""
-        mock_msg.level = 0
         mock_uvolsplit.return_value = ("HOSTDIR", "CONTDIR")
         ex_eng = ExecutionEngineCommon(self.local, self.xmode)
         ex_eng.opt["vol"] = list()
@@ -292,16 +289,13 @@ class ExecutionEngineCommonTestCase(TestCase):
         status = ex_eng._set_volume_bindings()
         self.assertTrue(status)
 
-    @patch('udocker.engine.base.Msg')
     @patch('udocker.engine.base.Uenv')
     @patch('udocker.engine.base.os.path.isdir')
     @patch('udocker.engine.base.FileUtil.cont2host')
-    def test_12__check_paths(self, mock_fuc2h, mock_isdir, mock_uenv,
-                             mock_msg):
+    def test_12__check_paths(self, mock_fuc2h, mock_isdir, mock_uenv):
         """Test12 ExecutionEngineCommon()._check_paths()."""
         Config.conf['root_path'] = "/sbin"
         Config.conf['user_path'] = "/bin"
-        mock_msg.level = 0
         mock_fuc2h.return_value = "/container/bin"
         mock_isdir.return_value = False
         ex_eng = ExecutionEngineCommon(self.local, self.xmode)
@@ -329,13 +323,11 @@ class ExecutionEngineCommonTestCase(TestCase):
         self.assertTrue(mock_fuc2h.called)
         self.assertTrue(mock_isdir.called)
 
-    @patch('udocker.engine.base.Msg')
[…]

[…]
