### 2021-07-24 35aed35
implement LOG in tests
