### 2021-07-25 3b45aa5
wip code readability, line lenght 100
