+++ udocker/cli.py
@@ -47,8 +47,7 @@ class UdockerCLI(object):
         if Config.conf['keystore'].startswith("/"):
             self.keystore = KeyStore(Config.conf['keystore'])
         else:
-            self.keystore = KeyStore(self.localrepo.homedir + "/" +
-                                     Config.conf['keystore'])
+            self.keystore = KeyStore(self.localrepo.homedir + "/" + Config.conf['keystore'])
 
     def _cdrepo(self, cmdp):
         """Select the top directory of a local repository"""
@@ -74,8 +73,7 @@ class UdockerCLI(object):
             imagerepo = imagespec
             tag = "latest"
 
-        if not (imagerepo and tag and
-                self.dockerioapi.is_repo_name(imagespec)):
+        if not (imagerepo and tag and self.dockerioapi.is_repo_name(imagespec)):
             LOG.error("must specify image:tag or repository/image:tag")
             return(None, None)
 
@@ -86,15 +84,13 @@ class UdockerCLI(object):
         if (not imagerepo) and def_imagerepo:
             imagerepo = def_imagerepo
 
-        if not (imagerepo and
-                self.dockerioapi.is_repo_name(imagerepo)):
+        if not (imagerepo and self.dockerioapi.is_repo_name(imagerepo)):
             LOG.error("enter image or repository/image without tag")
             return None
 
         return imagerepo
 
-    def _set_repository(self, registry_url, index_url=None,
-                        imagerepo=None, http_proxy=None):
+    def _set_repository(self, registry_url, index_url=None, imagerepo=None, http_proxy=None):
         """Select docker respository"""
         transport = "https:"
         if http_proxy:
@@ -127,10 +123,8 @@ class UdockerCLI(object):
 
             if "." in hostname:
                 try:
-                    self.dockerioapi.set_registry( \
-                            Config.conf['docker_registries'][hostname][0])
-                    self.dockerioapi.set_index( \
-                            Config.conf['docker_registries'][hostname][1])
+                    self.dockerioapi.set_registry(Config.conf['docker_registries'][hostname][0])
+                    self.dockerioapi.set_index(Config.conf['docker_registries'][hostname][1])
                 except (KeyError, NameError, TypeError):
                     self.dockerioapi.set_registry(transport + "//" + hostname)
                     self.dockerioapi.set_index(transport + "//" + hostname)
@@ -229,8 +223,7 @@ class UdockerCLI(object):
         while True:
             lines = term_lines
             while lines > 0:
-                repo_list = \
-                        self.dockerioapi.search_get_page(expression, term_lines)
+                repo_list = self.dockerioapi.search_get_page(expression, term_lines)
                 if not repo_list:
                     return self.STATUS_OK
 
@@ -418,13 +411,10 @@ class UdockerCLI(object):
 
         if to_container or clone:
             if clone:
-                container_id = self.localfileapi.import_clone(
-                    tarfile, name)
+                container_id = self.localfileapi.import_clone(tarfile, name)
             else:
-                (imagerepo, tag) = self._check_imagespec(imagespec,
-                                                         "IMPORTED:unknown")
-                container_id = self.localfileapi.import_tocontainer(
-                    tarfile, imagerepo, tag, name)
+                (imagerepo, tag) = self._check_imagespec(imagespec, "IMPORTED:unknown")
+                container_id = self.localfileapi.import_tocontainer(tarfile, imagerepo, tag, name)
 
             if container_id:
                 LOG.info("container ID: %s", container_id)
@@ -434,8 +424,7 @@ class UdockerCLI(object):
             if not imagerepo:
                 return self.STATUS_ERROR
 
[…]
