+++ udocker/engine/base.py
@@ -128,8 +128,7 @@ class ExecutionEngineCommon(object):
 
         exec_cmd = []
         for exec_cmd in Config.conf['cpu_affinity_exec_tools']:
-            exec_name = \
-                FileUtil(exec_cmd[0]).find_exec()
+            exec_name = FileUtil(exec_cmd[0]).find_exec()
             if exec_name:
                 exec_cmd[0] = exec_name
                 for (index, arg) in enumerate(exec_cmd):
@@ -160,14 +159,14 @@ class ExecutionEngineCommon(object):
                 LOG.error("invalid host volume path: %s", host_path)
                 return False
 
-            if not (cont_path and cont_path != '/' and
-                    cont_path.startswith('/')):
+            if not (cont_path and cont_path != '/' and cont_path.startswith('/')):
                 LOG.error("invalid container volume path: %s", cont_path)
                 return False
 
             if not os.path.exists(host_path):
-                if (host_path in Config.conf['dri_list'] or
-                        host_path in Config.conf['sysdirs_list']):
+                dril = Config.conf['dri_list']
+                sysl = Config.conf['sysdirs_list']
+                if (host_path in dril or host_path in sysl):
                     self.opt["vol"].remove(vol)
                     continue
 
@@ -248,8 +247,7 @@ class ExecutionEngineCommon(object):
         if not self.opt["cwd"]:
             self.opt["cwd"] = self.opt["home"]
 
-        cwd_path = FileUtil(self.container_root).cont2host(self.opt["cwd"],
-                                                           self.opt["vol"])
+        cwd_path = FileUtil(self.container_root).cont2host(self.opt["cwd"], self.opt["vol"])
         if os.path.isdir(cwd_path):
             return True
 
@@ -279,8 +277,7 @@ class ExecutionEngineCommon(object):
 
         path = self.opt["env"].getenv("PATH")
         exec_name = FileUtil(exec_name).find_exec(path, self.container_root,
-                                                  self.opt["vol"],
-                                                  self.opt["cwd"])
+                                                  self.opt["vol"], self.opt["cwd"])
         if exec_name:
             return self.container_root + '/' + exec_name
 
@@ -292,64 +289,44 @@ class ExecutionEngineCommon(object):
         # get container metadata unless we are dealing with a simple directory
         # tree in which case we don't have metadata
         if Config.conf['location']:
-            container_dir = ""
-            container_json = []
+            cont_dir = ""
+            cont_json = []
         else:
-            container_structure = \
-                ContainerStructure(self.localrepo, container_id)
-            (container_dir, container_json) = \
-                container_structure.get_container_attr()
-            if not container_dir:
+            cstruc = ContainerStructure(self.localrepo, container_id)
+            (cont_dir, cont_json) = cstruc.get_container_attr()
+            if not cont_dir:
                 return(None, None)
 
             # load metadata from container
             if not self.opt["nometa"]:
                 if not self.opt["user"]:
-                    self.opt["user"] = \
-                        container_structure.get_container_meta(
-                            "User", "", container_json)
+                    self.opt["user"] = cstruc.get_container_meta("User", "", cont_json)
 
                 if not self.opt["cwd"]:
-                    self.opt["cwd"] = \
-                        container_structure.get_container_meta(
-                            "WorkingDir", "", container_json)
+                    self.opt["cwd"] = cstruc.get_container_meta("WorkingDir", "", cont_json)
 
                 if not self.opt["hostname"]:
-                    self.opt["hostname"] = \
-                        container_structure.get_container_meta(
-                            "Hostname", "", container_json)
[…]
