### 2021-07-26 31b64f2
code readability line lenght 100 solves issue #351
