+++ udocker/utils/chksum.py
@@ -8,8 +8,7 @@ class ChkSUM(object):
     """Checksumming for files"""
 
     def __init__(self):
-        self._algorithms = {"sha256": self.sha256,
-                            "sha512": self.sha512}
+        self._algorithms = {"sha256": self.sha256, "sha512": self.sha512}
 
     def _hashlib(self, algorithm, filename):
         """hash calculation using hashlib"""
@@ -17,6 +16,7 @@ class ChkSUM(object):
             with open(filename, "rb") as filep:
                 for chunk in iter(lambda: filep.read(4096), b""):
                     algorithm.update(chunk)
+
             return algorithm.hexdigest()
         except (IOError, OSError):
             return ""
@@ -37,46 +37,3 @@ class ChkSUM(object):
             return self._algorithms[algorithm](filename)
 
         return ""
-
-# class ChkSUM(object):
-#     """Checksumming for files"""
-
-#     def __init__(self):
-#         self._algorithms = dict()
-#         self._algorithms["sha256"] = self._hashlib_sha256
-#         self._algorithms["sha512"] = self._hashlib_sha512
-
-#     def _hashlib(self, algorithm, filename):
-#         """hash calculation using hashlib"""
-#         try:
-#             with open(filename, "rb") as filep:
-#                 for chunk in iter(lambda: filep.read(4096), b""):
-#                     algorithm.update(chunk)
-#             return algorithm.hexdigest()
-#         except (IOError, OSError):
-#             return ""
-
-#     def _hashlib_sha256(self, filename):
-#         """sha256 calculation using hashlib"""
-#         LOG.info("sha256 calculation using hashlib")
-#         return self._hashlib(hashlib.sha256(), filename)
-
-#     def _hashlib_sha512(self, filename):
-#         """sha512 calculation using hashlib"""
-#         LOG.info("sha512 calculation using hashlib")
-#         return self._hashlib(hashlib.sha512(), filename)
-
-#     def sha256(self, filename):
-#         """Call the actual implementation selected in __init__"""
-#         return self._algorithms["sha256"](filename)
-
-#     def sha512(self, filename):
-#         """Call the actual implementation selected in __init__"""
-#         return self._algorithms["sha512"](filename)
-
-#     def hash(self, filename, algorithm):
-#         """Compute hash algorithm for file"""
-#         if algorithm in self._algorithms:
-#             return self._algorithms[algorithm](filename)
-
-#         return ""
