+++ tests/unit/test_cli.py
@@ -4,6 +4,9 @@
 """
 udocker unit tests: UdockerCLI
 """
+import sys
+sys.path.append('.')
+sys.path.append('../')
 
 import logging
 from unittest import TestCase, main
@@ -24,8 +27,8 @@ class UdockerCLITestCase(TestCase):
         Config().getconf()
         Config().conf['hostauth_list'] = ("/etc/passwd", "/etc/group")
         Config().conf['cmd'] = "/bin/bash"
-        Config().conf['cpu_affinity_exec_tools'] = \
-            (["numactl", "-C", "%s", "--", ], ["taskset", "-c", "%s", ])
+        Config().conf['cpu_affinity_exec_tools'] = (["numactl", "-C", "%s", "--", ],
+                                                    ["taskset", "-c", "%s", ])
         Config().conf['valid_host_env'] = "HOME"
         Config().conf['username'] = "user"
         Config().conf['userhome'] = "/"
@@ -207,26 +210,26 @@ class UdockerCLITestCase(TestCase):
     # def test_08__search_print_lines(self):
     #     """Test08 UdockerCLI()._search_print_lines()."""
 
-    # @patch('udocker.cli.DockerIoAPI.search_get_page')
-    # @patch('udocker.cli.HostInfo.termsize')
-    # def test_09__search_repositories(self, mock_termsz, mock_doiasearch):
-    #     """Test09 UdockerCLI()._search_repositories()."""
-    #     repo_list = [{"count": 1, "next": "", "previous": "",
-    #                   "results": [
-    #                       {
-    #                           "repo_name": "lipcomputing/ipyrad",
-    #                           "short_description": "Docker to run ipyrad",
-    #                           "star_count": 0,
-    #                           "pull_count": 188,
-    #                           "repo_owner": "",
-    #                           "is_automated": True,
-    #                           "is_official": False
-    #                       }]}]
-    #     mock_termsz.return_value = (40, "")
-    #     mock_doiasearch.return_value = repo_list
-    #     udoc = UdockerCLI(self.local)
-    #     status = udoc._search_repositories("ipyrad")
-    #     self.assertEqual(status, 0)
+    @patch('udocker.cli.DockerIoAPI.search_get_page')
+    @patch('udocker.cli.HostInfo.termsize')
+    def test_09__search_repositories(self, mock_termsz, mock_doiasearch):
+        """Test09 UdockerCLI()._search_repositories()."""
+        repo_list = {"count": 1, "next": "", "previous": "",
+                     "results": [
+                         {
+                             "repo_name": "lipcomputing/ipyrad",
+                             "short_description": "Docker to run ipyrad",
+                             "star_count": 0,
+                             "pull_count": 188,
+                             "repo_owner": "",
+                             "is_automated": True,
+                             "is_official": False
+                          }]}
+        mock_termsz.return_value = (3, "")
+        mock_doiasearch.side_effect = [repo_list, []]
+        udoc = UdockerCLI(self.local)
+        status = udoc._search_repositories("ipyrad", False)
+        self.assertEqual(status, 0)
 
     @patch('udocker.cli.DockerIoAPI.get_tags')
     def test_10__list_tags(self, mock_gettags):
@@ -248,9 +251,8 @@ class UdockerCLITestCase(TestCase):
     @patch.object(UdockerCLI, '_list_tags')
     @patch.object(UdockerCLI, '_split_imagespec')
     @patch.object(UdockerCLI, '_set_repository')
[…]
