### 2021-07-26 b304aa6
fix missing creation of list, increase unit cov
