+++ tests/unit/test_localrepository.py
@@ -35,8 +35,7 @@ class LocalRepositoryTestCase(TestCase):
     @patch('udocker.container.localrepo.FileUtil')
     def test_01_init(self, mock_fu):
         """Test01 LocalRepository() constructor."""
-        mock_fu.return_value.register_prefix.side_effect = \
-            [None, None, None]
+        mock_fu.return_value.register_prefix.side_effect = [None, None, None]
         lrepo = LocalRepository(UDOCKER_TOPDIR)
         self.assertTrue(lrepo.topdir)
         self.assertTrue(lrepo.reposdir)
@@ -54,8 +53,7 @@ class LocalRepositoryTestCase(TestCase):
     @patch('udocker.container.localrepo.FileUtil')
     def test_02_setup(self, mock_fu):
         """Test02 LocalRepository().setup()."""
-        mock_fu.return_value.register_prefix.side_effect = \
-            [None, None, None, None, None, None]
+        mock_fu.return_value.register_prefix.side_effect = [None, None, None, None, None, None]
         newdir = "/home/u2/.udocker"
         lrepo = LocalRepository(UDOCKER_TOPDIR)
         lrepo.setup(newdir)
@@ -67,12 +65,9 @@ class LocalRepositoryTestCase(TestCase):
     def test_03_create_repo(self, mock_fu, mock_mkdir, mock_exists):
         """Test03 LocalRepository().create_repo()."""
         Config.conf['keystore'] = "tmp"
-        mock_fu.return_value.register_prefix.side_effect = \
-            [None, None, None]
-        mock_exists.side_effect = [False, False, False, False,
-                                   False, False, False, False]
-        mock_mkdir.side_effect = [None, None, None, None,
-                                  None, None, None, None]
+        mock_fu.return_value.register_prefix.side_effect = [None, None, None]
+        mock_exists.side_effect = [False, False, False, False, False, False, False, False]
+        mock_mkdir.side_effect = [None, None, None, None, None, None, None, None]
         lrepo = LocalRepository(UDOCKER_TOPDIR)
         status = lrepo.create_repo()
         self.assertTrue(status)
@@ -80,8 +75,7 @@ class LocalRepositoryTestCase(TestCase):
         self.assertTrue(mock_mkdir.call_count, 8)
 
         Config.conf['keystore'] = "tmp"
-        mock_fu.return_value.register_prefix.side_effect = \
-            [None, None, None]
+        mock_fu.return_value.register_prefix.side_effect = [None, None, None]
         mock_exists.side_effect = OSError("fail")
         lrepo = LocalRepository(UDOCKER_TOPDIR)
         status = lrepo.create_repo()
@@ -91,19 +85,15 @@ class LocalRepositoryTestCase(TestCase):
     @patch('udocker.container.localrepo.FileUtil')
     def test_04_is_repo(self, mock_fu, mock_exists):
         """Test04 LocalRepository().is_repo()."""
-        mock_fu.return_value.register_prefix.side_effect = \
-            [None, None, None]
-        mock_exists.side_effect = [False, True, False, False,
-                                   True]
+        mock_fu.return_value.register_prefix.side_effect = [None, None, None]
+        mock_exists.side_effect = [False, True, False, False, True]
         lrepo = LocalRepository(UDOCKER_TOPDIR)
         status = lrepo.is_repo()
         self.assertTrue(mock_exists.call_count, 5)
         self.assertFalse(status)
 
-        mock_fu.return_value.register_prefix.side_effect = \
-            [None, None, None]
-        mock_exists.side_effect = [True, True, True, True,
-                                   True]
[…]
