+++ tests/unit/test_fileutil.py
@@ -4,6 +4,9 @@ udocker unit tests: FileUtil
 """
 
 import sys
+# sys.path.append('.')
+# sys.path.append('../')
+
 import os
 from unittest import TestCase, main
 from unittest.mock import patch, mock_open
@@ -65,31 +68,40 @@ class FileUtilTestCase(TestCase):
         self.assertEqual(futil.filename, '-')
         self.assertEqual(futil.basename, '-')
 
-    # @patch('udocker.utils.fileutil.os.path.isdir')
-    # @patch('udocker.utils.fileutil.os.path.realpath')
-    # def test_02__register_prefix(self, mock_rpath,
-    #                              mock_isdir):
-    #     """Test02 FileUtil._register_prefix()."""
-    #     prefix = "/dir"
-    #     res = ["/dir/", "/dir/", "/dir/", "/dir/"]
-    #     mock_rpath.return_value = "/dir"
-    #     mock_isdir.return_value = True
-    #     futil = FileUtil('filename.txt')
-    #     futil._register_prefix(prefix)
-    #     self.assertEqual(futil.safe_prefixes, res)
+    @patch('udocker.utils.fileutil.os.path.isdir')
+    @patch('udocker.utils.fileutil.os.path.realpath')
+    @patch('udocker.utils.fileutil.os.path.abspath')
+    @patch('udocker.utils.fileutil.os.path.basename')
+    def test_02__register_prefix(self, mock_base, mock_absp, mock_rpath, mock_isdir):
+        """Test02 FileUtil._register_prefix()."""
+        prefix = "/dir"
+        res = ['/tmp', '/dir/', '/dir/', '/dir/', '/dir/']
+        mock_base.return_value = 'filename.txt'
+        mock_absp.return_value = '/tmp/filename.txt'
+        mock_rpath.side_effect = ["/dir", "/dir", "/dir", "/dir"]
+        mock_isdir.return_value = True
+        futil = FileUtil('filename.txt')
+        futil._register_prefix(prefix)
+        self.assertEqual(futil.safe_prefixes, res)
 
-    #     prefix = "/dir/"
-    #     res = ["/dir/", "/dir/", "/dir/", "/dir/"]
-    #     mock_rpath.return_value = "/dir/"
-    #     mock_isdir.return_value = False
-    #     futil = FileUtil('filename.txt')
-    #     futil._register_prefix(prefix)
-    #     self.assertEqual(futil.safe_prefixes, res)
+        prefix = "/file"
+        res = ['/tmp', '/dir/', '/dir/', '/dir/', '/dir/', '/file', '/file']
+        mock_base.return_value = 'filename.txt'
+        mock_absp.return_value = '/tmp/filename.txt'
+        mock_rpath.side_effect = ["/file", "/file", "/file", "/file"]
+        mock_isdir.return_value = False
+        futil = FileUtil('filename2.txt')
+        futil._register_prefix(prefix)
+        self.assertEqual(futil.safe_prefixes, res)
 
+    @patch('udocker.utils.fileutil.os.path.abspath')
+    @patch('udocker.utils.fileutil.os.path.basename')
     @patch.object(FileUtil, '_register_prefix')
-    def test_03_register_prefix(self, mock_regpre):
+    def test_03_register_prefix(self, mock_regpre, mock_base, mock_absp):
         """Test03 FileUtil.register_prefix()."""
         mock_regpre.return_value = None
+        mock_base.return_value = 'filename.txt'
+        mock_absp.return_value = '/tmp/filename.txt'
         futil = FileUtil('filename.txt')
         futil.register_prefix()
         self.assertTrue(mock_regpre.called)
@@ -99,8 +111,7 @@ class FileUtilTestCase(TestCase):
[…]
