+++ tests/unit/test_dockerlocalfileapi.py
@@ -256,8 +256,7 @@ class DockerLocalFileAPITestCase(TestCase):
         tmp_imgdir = "/tmp/img1"
         manifest = dict()
         self.local.cd_imagerepo.return_value = None
-        self.local.get_image_attributes.return_value = \
-            (manifest, ['/cont/lname1.layer'])
+        self.local.get_image_attributes.return_value = (manifest, ['/cont/lname1.layer'])
         self.local.save_json.side_effect = [False, True]
         dlocapi = DockerLocalFileAPI(self.local)
         status = dlocapi._save_image(imgrepo, tag, struct, tmp_imgdir)
@@ -274,16 +273,14 @@ class DockerLocalFileAPITestCase(TestCase):
             "fsLayers": ({"blobSum": "foolayername"},),
             "history": ({"v1Compatibility": '["foojsonstring"]'},)
         }
-        mock_sha256.return_value = \
-            "8a29a15cefaeccf6545f7ecf11298f9672d2f0cdaf9e357a95133ac3ad3e1f07"
+        mock_sha256.return_value = "8a29a15cefaeccf6545f7ecf11298f9672d2f0cdaf9e357a95133ac3ad3e1f07"
         mock_rename.return_value = None
         mock_osbase.return_value = "lname1.layer"
         mock_exists.return_value = False
         mock_mkdir.return_value = None
         mock_copyto.return_value = False
         self.local.cd_imagerepo.return_value = None
-        self.local.get_image_attributes.return_value = \
-            (manifest, ['/cont/lname1.layer'])
+        self.local.get_image_attributes.return_value = (manifest, ['/cont/lname1.layer'])
         self.local.save_json.side_effect = [True, True]
         dlocapi = DockerLocalFileAPI(self.local)
         status = dlocapi._save_image(imgrepo, tag, struc, tmp_imgdir)
@@ -301,8 +298,7 @@ class DockerLocalFileAPITestCase(TestCase):
                  "repositories": {"repo1": {"TAG": "tag1", }, },
                  "manifest": [manifest]}
         tmp_imgdir = "/tmp/img1"
-        mock_sha256.return_value = \
-            "8a29a15cefaeccf6545f7ecf11298f9672d2f0cdaf9e357a95133ac3ad3e1f07"
+        mock_sha256.return_value = "8a29a15cefaeccf6545f7ecf11298f9672d2f0cdaf9e357a95133ac3ad3e1f07"
         mock_rename.return_value = None
         mock_osbase.return_value = "lname1.layer"
         mock_exists.return_value = False
@@ -312,8 +308,7 @@ class DockerLocalFileAPITestCase(TestCase):
         self.local.save_json.return_value = True
         mock_put.return_value = True
         self.local.cd_imagerepo.return_value = None
-        self.local.get_image_attributes.return_value = \
-            (manifest, ['/cont/lname1.layer'])
+        self.local.get_image_attributes.return_value = (manifest, ['/cont/lname1.layer'])
         self.local.save_json.side_effect = [True, True]
         dlocapi = DockerLocalFileAPI(self.local)
         status = dlocapi._save_image(imgrepo, tag, struc, tmp_imgdir)
@@ -324,8 +319,7 @@ class DockerLocalFileAPITestCase(TestCase):
     @patch('udocker.docker.FileUtil.tar')
     @patch('udocker.docker.os.makedirs')
     @patch('udocker.docker.FileUtil.mktmp')
-    def test_10_save(self, mock_mktmp, mock_mkdir, mock_tar,
-                     mock_rm, mock_svimg):
+    def test_10_save(self, mock_mktmp, mock_mkdir, mock_tar, mock_rm, mock_svimg):
[…]
