+++ tests/unit/test_fakechroot.py
@@ -21,8 +21,8 @@ class FakechrootEngineTestCase(TestCase):
         Config().getconf()
         Config().conf['hostauth_list'] = ("/etc/passwd", "/etc/group")
         Config().conf['cmd'] = "/bin/bash"
-        Config().conf['cpu_affinity_exec_tools'] = \
-            (["numactl", "-C", "%s", "--", ], ["taskset", "-c", "%s", ])
+        Config().conf['cpu_affinity_exec_tools'] = (["numactl", "-C", "%s", "--", ],
+                                                    ["taskset", "-c", "%s", ])
         Config().conf['valid_host_env'] = "HOME"
         Config().conf['username'] = "user"
         Config().conf['userhome'] = "/"
@@ -75,8 +75,7 @@ class FakechrootEngineTestCase(TestCase):
         Config().conf['fakechroot_so'] = ""
         mock_exists.return_value = False
         mock_osinfo.return_value.arch.return_value = "amd64"
-        mock_osinfo.return_value.osdistribution.return_value = \
-            ("linux", "4.8.1")
+        mock_osinfo.return_value.osdistribution.return_value = ("linux", "4.8.1")
         mock_futil.return_value.find_file_in_dir.return_value = ""
         mock_sysex.return_value = 1
         ufake = FakechrootEngine(self.local, self.xmode)
@@ -87,10 +86,8 @@ class FakechrootEngineTestCase(TestCase):
         Config().conf['fakechroot_so'] = ""
         mock_exists.return_value = False
         mock_osinfo.return_value.arch.return_value = "amd64"
-        mock_osinfo.return_value.osdistribution.return_value = \
-            ("linux", "4.8.1")
-        mock_futil.return_value.find_file_in_dir.return_value = \
-            "/libfakechroot.so"
+        mock_osinfo.return_value.osdistribution.return_value = ("linux", "4.8.1")
+        mock_futil.return_value.find_file_in_dir.return_value = "/libfakechroot.so"
         ufake = FakechrootEngine(self.local, self.xmode)
         out = ufake.select_fakechroot_so()
         self.assertTrue(mock_futil.called)
@@ -99,10 +96,8 @@ class FakechrootEngineTestCase(TestCase):
         Config().conf['fakechroot_so'] = ""
         mock_exists.return_value = False
         mock_osinfo.return_value.arch.return_value = "i386"
-        mock_osinfo.return_value.osdistribution.return_value = \
-            ("linux", "4.8.1")
-        mock_futil.return_value.find_file_in_dir.return_value = \
-            "/libfakechroot.so"
+        mock_osinfo.return_value.osdistribution.return_value = ("linux", "4.8.1")
+        mock_futil.return_value.find_file_in_dir.return_value = "/libfakechroot.so"
         ufake = FakechrootEngine(self.local, self.xmode)
         out = ufake.select_fakechroot_so()
         self.assertTrue(mock_futil.called)
@@ -111,10 +106,8 @@ class FakechrootEngineTestCase(TestCase):
         Config().conf['fakechroot_so'] = ""
         mock_exists.return_value = False
         mock_osinfo.return_value.arch.return_value = "arm64"
-        mock_osinfo.return_value.osdistribution.return_value = \
-            ("linux", "4.8.1")
-        mock_futil.return_value.find_file_in_dir.return_value = \
-            "/libfakechroot.so"
+        mock_osinfo.return_value.osdistribution.return_value = ("linux", "4.8.1")
[…]
