+++ tests/unit/test_execenginecommon.py
@@ -20,8 +20,8 @@ class ExecutionEngineCommonTestCase(TestCase):
         Config().getconf()
         Config().conf['hostauth_list'] = ("/etc/passwd", "/etc/group")
         Config().conf['cmd'] = "/bin/bash"
-        Config().conf['cpu_affinity_exec_tools'] = \
-            (["numactl", "-C", "%s", "--", ], ["taskset", "-c", "%s", ])
+        Config().conf['cpu_affinity_exec_tools'] = (["numactl", "-C", "%s", "--", ],
+                                                    ["taskset", "-c", "%s", ])
         Config().conf['location'] = ""
         Config().conf['uid'] = 1000
         Config().conf['sysdirs_list'] = ["/", ]
@@ -156,8 +156,7 @@ class ExecutionEngineCommonTestCase(TestCase):
     @patch.object(ExecutionEngineCommon, '_create_mountpoint')
     @patch('udocker.engine.base.os.path.exists')
     @patch('udocker.engine.base.Uvolume.split')
-    def test_07__check_volumes(self, mock_uvolsplit, mock_exists,
-                               mock_crmpoint):
+    def test_07__check_volumes(self, mock_uvolsplit, mock_exists, mock_crmpoint):
         """Test07 ExecutionEngineCommon()._check_volumes()."""
         mock_uvolsplit.return_value = ("HOSTDIR", "CONTDIR")
         ex_eng = ExecutionEngineCommon(self.local, self.xmode)
@@ -396,8 +395,7 @@ class ExecutionEngineCommonTestCase(TestCase):
     @patch('udocker.engine.base.os.path.islink')
     @patch('udocker.engine.base.FileBind')
     @patch.object(ExecutionEngineCommon, '_is_mountpoint')
-    def test_15__select_auth_files(self, mock_ismpoint, mock_fbind, mock_islink,
-                                   mock_isfile):
+    def test_15__select_auth_files(self, mock_ismpoint, mock_fbind, mock_islink, mock_isfile):
         """Test15 ExecutionEngineCommon()._select_auth_files()."""
         resp = "/fbdir/#etc#passwd"
         resg = "/fbdir/#etc#group"
@@ -469,8 +467,7 @@ class ExecutionEngineCommonTestCase(TestCase):
         auth.return_value = mock_auth
         ex_eng = ExecutionEngineCommon(self.local, self.xmode)
         ex_eng.opt["hostauth"] = True
-        auth.get_user.return_value = ("user1", "1000", "1000", "usr",
-                                      "/home/user", "/bin/bash")
+        auth.get_user.return_value = ("user1", "1000", "1000", "usr", "/home/user", "/bin/bash")
         status = ex_eng._user_from_str(userstr, host_auth=auth)
         self.assertEqual(status, (True, res))
         auth = nixauth.stop()
@@ -484,8 +481,7 @@ class ExecutionEngineCommonTestCase(TestCase):
         auth.return_value = mock_auth
         ex_eng = ExecutionEngineCommon(self.local, self.xmode)
         ex_eng.opt["hostauth"] = False
-        auth.get_user.return_value = ("user1", "1000", "1000", "usr",
-                                      "/home/user", "/bin/bash")
+        auth.get_user.return_value = ("user1", "1000", "1000", "usr", "/home/user", "/bin/bash")
         status = ex_eng._user_from_str(userstr, container_auth=auth)
         self.assertEqual(status, (True, res))
         auth = nixauth.stop()
@@ -499,8 +495,7 @@ class ExecutionEngineCommonTestCase(TestCase):
         auth.return_value = mock_auth
         ex_eng = ExecutionEngineCommon(self.local, self.xmode)
[…]
