+++ tests/unit/test_runc.py
@@ -20,8 +20,8 @@ class RuncEngineTestCase(TestCase):
         Config().getconf()
         Config().conf['hostauth_list'] = ("/etc/passwd", "/etc/group")
         Config().conf['cmd'] = "/bin/bash"
-        Config().conf['cpu_affinity_exec_tools'] = \
-            (["numactl", "-C", "%s", "--", ], ["taskset", "-c", "%s", ])
+        Config().conf['cpu_affinity_exec_tools'] = (["numactl", "-C", "%s", "--", ],
+                                                    ["taskset", "-c", "%s", ])
         Config().conf['runc_capabilities'] = [
             "CAP_KILL", "CAP_NET_BIND_SERVICE", "CAP_CHOWN", "CAP_DAC_OVERRIDE",
             "CAP_FOWNER", "CAP_FSETID", "CAP_KILL", "CAP_SETGID", "CAP_SETUID",
@@ -59,8 +59,7 @@ class RuncEngineTestCase(TestCase):
     @patch('udocker.engine.runc.HostInfo.arch')
     @patch('udocker.engine.runc.FileUtil.find_file_in_dir')
     @patch('udocker.engine.runc.FileUtil.find_exec')
-    def test_02_select_runc(self, mock_findexe, mock_find,
-                            mock_arch, mock_exists):
+    def test_02_select_runc(self, mock_findexe, mock_find, mock_arch, mock_exists):
         """Test02 RuncEngine().select_runc()."""
         Config.conf['use_runc_executable'] = ""
         mock_findexe.return_value = "/bin/runc-arm"
@@ -93,11 +92,10 @@ class RuncEngineTestCase(TestCase):
     @patch('udocker.engine.runc.FileUtil.size')
     @patch('udocker.engine.runc.subprocess.call')
     @patch('udocker.engine.runc.os.path.realpath')
-    def test_03__load_spec(self, mock_realpath, mock_call, mock_size,
-                           mock_rm, mock_reg, mock_jload):
+    def test_03__load_spec(self, mock_rpath, mock_call, mock_size, mock_rm, mock_reg, mock_jload):
         """Test03 RuncEngine()._load_spec()."""
         mock_size.side_effect = [-1, -1]
-        mock_realpath.return_value = "/container/ROOT"
+        mock_rpath.return_value = "/container/ROOT"
         mock_call.return_value = 1
         mock_rm.return_value = None
         mock_reg.return_value = None
@@ -106,7 +104,7 @@ class RuncEngineTestCase(TestCase):
         self.assertFalse(mock_rm.called)
         self.assertFalse(mock_reg.called)
         self.assertTrue(mock_call.called)
-        self.assertTrue(mock_realpath.called)
+        self.assertTrue(mock_rpath.called)
         self.assertFalse(status)
 
         jload = {"container": "cxxx", "parent": "dyyy",
@@ -116,7 +114,7 @@ class RuncEngineTestCase(TestCase):
                  "Image": "sha256:aa"
                 }
         mock_size.side_effect = [100, 100]
-        mock_realpath.return_value = "/container/ROOT"
+        mock_rpath.return_value = "/container/ROOT"
         mock_call.return_value = 0
         mock_rm.return_value = None
         mock_reg.return_value = None
@@ -145,12 +143,16 @@ class RuncEngineTestCase(TestCase):
             self.assertTrue(mopen.called)
             self.assertTrue(status)
 
+        mock_jdump.side_effect = OSError("fail")
+        rcex = RuncEngine(self.local, self.xmode)
+        status = rcex._save_spec()
+        self.assertFalse(status)
+
     @patch('udocker.engine.runc.os.getgid')
     @patch('udocker.engine.runc.os.getuid')
[…]
