+++ tests/unit/test_elfpatcher.py
@@ -39,8 +39,7 @@ class ElfPatcherTestCase(TestCase):
     @patch('udocker.helper.elfpatcher.HostInfo.arch')
     @patch('udocker.helper.elfpatcher.FileUtil.find_file_in_dir')
     @patch('udocker.helper.elfpatcher.os.path.realpath')
-    def test_02_select_patchelf(self, mock_path, mock_find,
-                                mock_arch, mock_exists):
+    def test_02_select_patchelf(self, mock_path, mock_find, mock_arch, mock_exists):
         """Test02 ElfPatcher().select_patchelf()."""
         mock_path.return_value = "/some_contdir"
         mock_arch.return_value = "arm"
@@ -49,6 +48,22 @@ class ElfPatcherTestCase(TestCase):
         elfp = ElfPatcher(self.local, self.contid)
         output = elfp.select_patchelf()
         self.assertEqual(output, "runc-arm")
+
+        mock_path.return_value = "/some_contdir"
+        mock_arch.return_value = "amd64"
+        mock_find.return_value = "runc-x86_64"
+        mock_exists.return_value = True
+        elfp = ElfPatcher(self.local, self.contid)
+        output = elfp.select_patchelf()
+        self.assertEqual(output, "runc-x86_64")
+
+        mock_path.return_value = "/some_contdir"
+        mock_arch.return_value = "i386"
+        mock_find.return_value = "runc-i386"
+        mock_exists.return_value = True
+        elfp = ElfPatcher(self.local, self.contid)
+        output = elfp.select_patchelf()
+        self.assertEqual(output, "runc-i386")
 
         mock_path.return_value = "/some_contdir"
         mock_arch.return_value = "arm"
@@ -104,8 +119,7 @@ class ElfPatcherTestCase(TestCase):
     @patch('udocker.helper.elfpatcher.os.path.realpath')
     @patch.object(ElfPatcher, '_walk_fs')
     @patch.object(ElfPatcher, 'select_patchelf')
-    def test_05_guess_elf_loader(self, mock_spelf, mock_walk,
-                                 mock_path):
+    def test_05_guess_elf_loader(self, mock_spelf, mock_walk, mock_path):
         """Test05 ElfPatcher().guess_elf_loader()."""
         mock_spelf.return_value = "ld.so"
         mock_walk.return_value = ""
@@ -140,7 +154,6 @@ class ElfPatcherTestCase(TestCase):
     @patch.object(ElfPatcher, 'get_original_loader')
     def test_07_get_container_loader(self, mock_gol, mock_exists, mock_path):
         """Test07 ElfPatcher().get_container_loader().
-
         Get an absolute pathname to the container ld.so"""
         mock_gol.return_value = ""
         mock_path.return_value = "/some_contdir"
@@ -151,13 +164,11 @@ class ElfPatcherTestCase(TestCase):
         mock_gol.return_value = "ld.so"
         mock_path.return_value = "/some_contdir"
         elfp = ElfPatcher(self.local, self.contid)
-        self.assertEqual(elfp.get_container_loader(),
-                         elfp._container_root + "/" + "ld.so")
+        self.assertEqual(elfp.get_container_loader(), elfp._container_root + "/" + "ld.so")
 
     @patch('udocker.helper.elfpatcher.os.path.realpath')
     @patch('udocker.helper.elfpatcher.FileUtil.getdata')
-    def test_08_get_patch_last_path(self, mock_getdata,
-                                    mock_path):
+    def test_08_get_patch_last_path(self, mock_getdata, mock_path):
         """Test08 ElfPatcher().get_patch_last_path()."""
         mock_getdata.return_value = ""
[…]
