+++ tests/unit/test_nixauthentication.py
@@ -41,56 +41,54 @@ class NixAuthenticationTestCase(TestCase):
     @patch.object(NixAuthentication, '_get_user_from_file')
     def test_03_user_in_subuid(self, mock_file, mock_host):
         """Test03 NixAuthentication().user_in_subuid."""
-        mock_host.return_value = ("user", "1000", "1000", "usr",
-                                  "/home/user", "/bin/bash")
+        mock_host.return_value = ("user", "1000", "1000", "usr", "/home/user", "/bin/bash")
         auth = NixAuthentication()
         auth.subuid_file = "/etc/subuid"
         subuid_line = StringIO('user:100000:65536')
         with patch(BUILTINS + '.open') as mopen:
-            mopen.return_value.__iter__ = (
-                lambda self: iter(subuid_line.readline, ''))
+            mopen.return_value.__iter__ = (lambda self: iter(subuid_line.readline, ''))
             listuser = auth.user_in_subuid("user")
             self.assertEqual(listuser, [("100000", "65536")])
             self.assertTrue(mock_host.called)
 
-        mock_file.return_value = ("user", "1000", "1000", "usr",
-                                  "/home/user", "/bin/bash")
+        mock_file.return_value = ("user", "1000", "1000", "usr", "/home/user", "/bin/bash")
         auth = NixAuthentication()
         auth.passwd_file = "/etc/passwd"
         auth.subuid_file = "/etc/subuid"
         subuid_line = StringIO('user:100000:65536')
         with patch(BUILTINS + '.open') as mopen:
-            mopen.return_value.__iter__ = (
-                lambda self: iter(subuid_line.readline, ''))
+            mopen.return_value.__iter__ = (lambda self: iter(subuid_line.readline, ''))
             listuser = auth.user_in_subuid("user")
             self.assertEqual(listuser, [("100000", "65536")])
             self.assertTrue(mock_file.called)
+
+        mock_open.side_effect = OSError("fail")
+        auth = NixAuthentication()
+        auth.passwd_file = "/etc/passwd"
+        listuser = auth.user_in_subuid("user")
+        self.assertEqual(listuser, list())
 
     @patch.object(NixAuthentication, '_get_user_from_host')
     @patch.object(NixAuthentication, '_get_user_from_file')
     def test_04_user_in_subgid(self, mock_file, mock_host):
         """Test04 NixAuthentication().user_in_subgid."""
-        mock_host.return_value = ("user", "1000", "1000", "usr",
-                                  "/home/user", "/bin/bash")
+        mock_host.return_value = ("user", "1000", "1000", "usr", "/home/user", "/bin/bash")
         auth = NixAuthentication()
         auth.subuid_file = "/etc/subgid"
         subgid_line = StringIO('user:100000:65536')
         with patch(BUILTINS + '.open') as mopen:
-            mopen.return_value.__iter__ = (
-                lambda self: iter(subgid_line.readline, ''))
+            mopen.return_value.__iter__ = (lambda self: iter(subgid_line.readline, ''))
             listuser = auth.user_in_subgid("user")
             self.assertEqual(listuser, [("100000", "65536")])
             self.assertTrue(mock_host.called)
 
-        mock_file.return_value = ("user", "1000", "1000", "usr",
-                                  "/home/user", "/bin/bash")
+        mock_file.return_value = ("user", "1000", "1000", "usr", "/home/user", "/bin/bash")
         auth = NixAuthentication()
         auth.passwd_file = "/etc/passwd"
         auth.subgid_file = "/etc/subgid"
[…]
