### 2021-08-02 666a683
wip tests oci structure
