+++ tests/unit/test_tools.py
@@ -39,10 +39,12 @@ class UdockerToolsTestCase(TestCase):
         self.assertTrue(mock_geturl.called)
         self.assertEqual(utools.localrepo, self.local)
 
-    # def test_02__instructions(self):
-    #     """Test02 UdockerTools()._instructions()."""
-    #     utools = UdockerTools(self.local)
-    #     utools._instructions()
+    @patch('udocker.tools.MSG.info')
+    def test_02__instructions(self, mock_msg):
+        """Test02 UdockerTools()._instructions()."""
+        utools = UdockerTools(self.local)
+        utools._instructions()
+        self.assertTrue(mock_msg.called_count, 2)
 
     def test_03__version2int(self):
         """Test03 UdockerTools()._version2int()."""
@@ -76,9 +78,7 @@ class UdockerToolsTestCase(TestCase):
     @patch('udocker.tools.os.listdir')
     def test_06_purge(self, mock_lsdir, mock_fureg, mock_furm):
         """Test06 UdockerTools().purge()."""
-        mock_lsdir.side_effect = [["f1", "f2"],
-                                  ["f3", "f4"],
-                                  ["f5", "f6"]]
+        mock_lsdir.side_effect = [["f1", "f2"], ["f3", "f4"], ["f5", "f6"]]
         mock_fureg.side_effect = [None, None, None, None, None, None]
         mock_furm.side_effect = [None, None, None, None, None, None]
         utools = UdockerTools(self.local)
@@ -118,8 +118,7 @@ class UdockerToolsTestCase(TestCase):
     @patch('udocker.tools.os.path.realpath')
     @patch('udocker.tools.os.path.exists')
     @patch.object(UdockerTools, '_download')
-    def test_08__get_file(self, mock_downl, mock_exists, mock_rpath,
-                          mock_isfile):
+    def test_08__get_file(self, mock_downl, mock_exists, mock_rpath, mock_isfile):
         """Test08 UdockerTools()._get_file()."""
         url = ""
         mock_downl.return_value = ""
@@ -149,8 +148,7 @@ class UdockerToolsTestCase(TestCase):
     @patch('udocker.tools.os.path.basename')
     @patch('udocker.tools.FileUtil.mktmpdir')
     @patch('udocker.tools.os.path.isfile')
-    def test_09__verify_version(self, mock_isfile, mock_fumktmp,
-                                mock_osbase, mock_fugetdata,
+    def test_09__verify_version(self, mock_isfile, mock_fumktmp, mock_osbase, mock_fugetdata,
                                 mock_furm, mock_versionok):
         """Test09 UdockerTools()._verify_version()."""
         tball = "/home/udocker.tar"
@@ -189,8 +187,7 @@ class UdockerToolsTestCase(TestCase):
     @patch('udocker.tools.os.path.basename')
     @patch('udocker.tools.FileUtil')
     @patch('udocker.tools.os.path.isfile')
-    def test_10__install(self, mock_isfile, mock_futil,
-                         mock_osbase):
+    def test_10__install(self, mock_isfile, mock_futil, mock_osbase):
         """Test10 UdockerTools()._install()."""
         tfile = ""
         mock_isfile.return_value = False
@@ -203,8 +200,7 @@ class UdockerToolsTestCase(TestCase):
         tfile = "udocker.tar"
         mock_isfile.return_value = True
         mock_futil.return_value.chmod.return_value = None
-        mock_futil.return_value.rchmod.side_effect = [None, None, None,
-                                                      None, None, None]
+        mock_futil.return_value.rchmod.side_effect = [None, None, None, None, None, None]
         mock_osbase.side_effect = ["ls", "ls", "lib1", "lib1", "doc", "doc1"]
         self.local.create_repo.return_value = None
[…]
