### 2021-08-06 indigo-dc/udocker#348 @jorge-lip
Yes not all the authors are contributing with code. All the persons listed as authors in the paper have contributed to udocker in different ways including the concept, design, requirements, reviews, documentation, SQA implementation and validation of the releases against use cases. Don't know if there is a requirement in JOSS that only code contributors can be authors so please advise.
