### 2021-08-08 #3295 @mviereck
Finally, about 4 weeks later, I got a short response by @jorge-lip  about the list of authors.
The list might be appropriate or not, I don't know. It contains beta testers and others, too.

It is very hard to get any response since the very begin of the review.
I don't want to ask any questions anymore and than to wait and beg for answers. There are more reminders then responses.
I've lost my good will to get the paper through the review.

I vote to reject the paper due to missing cooperation by the submitting author.
