### 2021-08-09 7687060
improved summary of developments
