+++ paper/paper.md
@@ -167,265 +167,31 @@ cannot provide features that require privileges.
 
 # Developments since 1.1.1
 
-udocker was initially developed in the context of the INDIGO-DataCloud
-[@INDIGO2018] research project between 2015 and 2017 as a proof of concept 
-aimed to show that scientific applications could be encapsulated in Linux 
-containers to ease execution across the growing ecosystem of computing 
-resources available to researchers including Linux batch systems and 
-interactive clusters. In particular it aimed to show that containers could 
-be executed by the end-users without requiring changes to the computing 
-systems and without system administrator intervention, thus empowering users 
-and promoting the adoption of containers in these environments.
-
-Being a proof of concept the initial versions were not designed for
-production use. Later in the project it become evident that udocker
-had gain adoption beyond its original purpose and scope and that is was 
-already being actively used in production environments. After the 
-first udocker publication [@GOMES2018] produced using versions
-1.1.0 and 1.1.1, the development effort was directed
-to enhance udocker for production use by improving the design, robustness 
-and functionality. Two code branches became supported in parallel.
-The *devel* branch for the production versions 1.1.x retained the original 
-proof of concept code for Python 2, while the *devel3* branch supported 
-the development of the new modular design with support for Python 3 that 
-later gave origin to the 1.2.x pre-releases. The version 1.3.0 released in 
-June of 2021 is the first production release having the new design and 
-support for Python 2 and 3.
-
-Since version 1.1.1 the udocker code was reorganized, largely rewritten 
-and improved. Starting with the 1.2.0 pre-release, udocker was completely 
-restructured moving from being a single large 
-monolithic Python script to become a modular Python application, making 
-maintenance and contributions easier. The new code structure has 40 Python
-modules and supports both Python 2.6 or higher and Python 3. Since container 
-technologies are in constant evolution, this new code structure was 
-essential to accommodate any future improvements such as new container 
-formats, APIs and execution engines as they become mainstream. 
-
-The improvements added after 1.1.1 include a more robust command line 
-interface addressing several of the problems that affected the initial 
-versions in terms of command line parsing and validation of arguments.
-The parsing of the configuration files was reimplemented to simplify
-the design and prevent injection of code via the configuration files.
-Configuration is now possible at three levels, system configuration via 
-`/etc/udocker.conf`, user configuration via `$HOME/.udocker/udocker.conf` 
-and udocker repository level via `$UDOCKER_DIR/udocker.conf`. 
-The most relevant configuration options can now be overridden through
-new environment variables.
-
-* `UDOCKER_DEFAULT_EXECUTION_MODE`: to change the default execution 
-engine mode, which is currently **P1** using *PRoot*.
-* `UDOCKER_FAKECHROOT_SO`: to enforce the use of a specific *Fakechroot*
-shareable library for use in **F** execution modes.
-* `UDOCKER_USE_CURL_EXECUTABLE`: to select a *curl* executable as
-alternative to `pycurl` for downloads and interaction with REST APIs.
-* `UDOCKER_USE_PROOT_EXECUTABLE`: to enforce the use of a given *PRoot*
-executable for use in **P** execution modes.
-* `UDOCKER_USE_RUNC_EXECUTABLE`: to enforce the use of a given *runc*
-or *crun* executable for use in **R** execution modes.
-* `UDOCKER_USE_SINGULARITY_EXECUTABLE`: to enforce the use of a given 
-*Singularity* executable for use in **S** execution modes.
-* `UDOCKER_FAKECHROOT_EXPAND_SYMLINKS`: to control the expansion of
-symbolic links in paths pointing to volumes when using the **F** modes. 
-This variable allows to disable the new path translation algorithm that 
-is now accurate but slower. 
[…]
