### 2021-08-09 #3295 @reidpr
Given that I lead Charliecloud which is or at least could be perceived to be a competing product, I probably have a conflict of interest.

Unrelated, authors might find our pre-print interesting. It covers the workings unprivileged containers in some detail and offers a new classification scheme for container privilege models. It will be published at SC21 this fall. https://arxiv.org/abs/2104.07508
