### 2021-08-09 #3295 @vsoch
@danielskatz both reviewers have expressed concerns about publishing this paper - and it seems like they aren't being heard -  @mviereck put a lot of time/energy into the review and wasn't satisfied at the end, and instead of hearing his concerns (which are somewhat overlapping with mine) he's simply removed from the process. What is the point of having reviewers if you aren't listening to them? :thinking:
