+++ paper/paper.md
@@ -167,31 +167,265 @@ cannot provide features that require privileges.
 
 # Developments since 1.1.1
 
-Since 1.1.1, udocker moved from a proof of concept tool suitable for production. 
-The focus has been put on sustainability and interoperability. The code structure 
-was redesigned moving from the Python 2 monolithic code to a new Python 3 modular 
-implementation. New features and improvements were added to support a wider range 
-of Linux distributions and environments. To this end the execution engines now 
-support *crun* for namespace based execution with cgroups v2 and a port of *Fakechroot*
-to support *MUSL* libc. The existing execution engines were also enhanced to address 
-the shortcomings present in the proof of concept and improve interoperability for 
-execution across a wider range of distributions, kernels and better supporting the 
-functionalities required for containers execution. Support for multiple architectures 
-was added and binaries for ARM and x86 are now provided for *PRoot* based execution, 
-further binaries can be easily added as needed. Support for accelerated computing 
-was implemented by introducing the capability to setup containers to use NVIDIA GPUs 
-according to the host environment and across the several execution engines. The 
-handling of container images was improved both at the level of interaction with 
-container registries and handling of formats. The interaction with registries is 
-now more interoperable and robust and support for the OCI image specification v1 
-was introduced. Furthermore the search capabilities support docker registry v1 and 
-v2 specifications and listing of tags. The command line interface and configuration 
-was improved with new commands, flags and better parsing of both the command line 
-and config files. Finally the installation of the tools and libraries was decoupled 
-from udocker itself, both can now be released separately. The shared installation 
-of udocker in distributed filesystems was addressed with changes introduced to 
-handle the cases where udocker is executed from readonly filesystems. Finally, udocker 
-is now also available from PyPI.
+udocker was initially developed in the context of the INDIGO-DataCloud
+[@INDIGO2018] research project between 2015 and 2017 as a proof of concept 
+aimed to show that scientific applications could be encapsulated in Linux 
+containers to ease execution across the growing ecosystem of computing 
+resources available to researchers including Linux batch systems and 
+interactive clusters. In particular it aimed to show that containers could 
+be executed by the end-users without requiring changes to the computing 
+systems and without system administrator intervention, thus empowering users 
+and promoting the adoption of containers in these environments.
+
+Being a proof of concept the initial versions were not designed for
+production use. Later in the project it become evident that udocker
+had gain adoption beyond its original purpose and scope and that is was 
+already being actively used in production environments. After the 
+first udocker publication [@GOMES2018] produced using versions
+1.1.0 and 1.1.1, the development effort was directed
+to enhance udocker for production use by improving the design, robustness 
+and functionality. Two code branches became supported in parallel.
+The *devel* branch for the production versions 1.1.x retained the original 
+proof of concept code for Python 2, while the *devel3* branch supported 
+the development of the new modular design with support for Python 3 that 
+later gave origin to the 1.2.x pre-releases. The version 1.3.0 released in 
+June of 2021 is the first production release having the new design and 
+support for Python 2 and 3.
+
+Since version 1.1.1 the udocker code was reorganized, largely rewritten 
+and improved. Starting with the 1.2.0 pre-release, udocker was completely 
+restructured moving from being a single large 
+monolithic Python script to become a modular Python application, making 
+maintenance and contributions easier. The new code structure has 40 Python
+modules and supports both Python 2.6 or higher and Python 3. Since container 
+technologies are in constant evolution, this new code structure was 
+essential to accommodate any future improvements such as new container 
+formats, APIs and execution engines as they become mainstream. 
+
[…]
