### 2021-08-09 #3295 @danielskatz
@vsoch - Did you feel that your concerns weren't addressed?  I know you had some, but I thought from your agreeing that this could be published after discussions with me and the author, you were ok with this, given your comments in https://github.com/openjournals/joss-reviews/issues/3295#issuecomment-863345897 in particular.

The issue with @mviereck and the author is something different to me, where the relationship broke down - note that the problems in https://github.com/openjournals/joss-reviews/issues/3295#issuecomment-894837288 are generally not technical but are based on this problematic relationship.  Given this, it seemed to me that a change was needed.
