### 2021-08-09 #3295 @vsoch
That was my response after I expressed concerns to you privately and you emailed me back that you "don't have any issue with accepting this" and I shrugged my shoulders and decided that the other reviewer would be the deciding factor. But now that the other reviewer has come to the same conclusions as me (and a few more) I'm thinking that my views weren't so unfair or far off. To reiterate my concerns, I believe that a refactor to Python 3 and adding a few new features does not constitute enough for a new publication, especially since the first was only in 2018. There are so many pieces of software out there that have updated versions, to Python 3, and they don't pursue publication because the core of the software just isn't different. This entire review feels really strange to me, almost like what we have to say doesn't matter.
