### 2021-08-09 #3295 @danielskatz
I'm now asking other JOSS AEiCs for advice on how to handle this - perhaps my judgement that this was in scope was incorrect, and we should have had all the JOSS editors vote.
