### 2021-08-09 #3295 @mariojmdavid
dear reviewers and editor,
it is unfortunate that things have come to this point of rupture.
the first question is/was of course if this new version was appropriate to be published in the JOSS, this has been questioned in the very beginning of the process, and it seemed the answer was yes, and thus the effort from absolutely everybody to improve, the paper itself, but as well the SW and the documentation, and from my point of view, indeed the comments and scrutiny from the reviewers where precious for such improvements.

the release of the 1.3.0 version was somewhat disruptive but necessary, since it was being improved and tested in real production environments by several users for many months,  and that is why it was deemed mature for production release, while it made all the sense that the paper would be about this release.

I have been into both sides of the process of review of papers, as well as of review of projects, in either case I have seen my papers and my projects being rejected, as well as I have rejected both papers and projects.
when I reject a paper or a project I say outright from the beginning, and of course state all the reasons to do that
it does not mean that I may not change my mind or vote later with discussion with other reviewers that have other or different viewpoints that I may come to agree with.

Also happens that I accept a paper but with comments that should be implemented in order to be published, I think this has to be evaluated from the beginning, if a paper has the potential to be publish but needs improvement, or if the paper is not appropriate for publishing, you say it outright in the beginning.

As I stated earlier, whatever the decision is to publish or not the paper, the SW and the documentation have had a large improvement, and of course this was due to the work of the reviewers, and ours on the implementation.

with kind regards
Mario David
