### 2021-08-09 #3295 @vsoch
Thank you @mariojmdavid. Just to be transparent, my concerns were there from the beginning, and I emailed the editor here and was sort of pushed in the direction of "I think this is okay" (subtext you should too) and I proceeded with the review. I didn't feel that I had the power or right to be assertive about my opinion, so I stopped there. I think if I were truly in charge I would have concluded (before any review) that I didn't consider this second paper appropriate for a piece of software that already exists and has been published - in my mind updates to or new features, refactor to different structure or language, does not constitute enough for a paper. And I proceeded rather submissively thinking I didn't think I really had a choice to outright reject reviewing it.

But regardless of my critique - I realize that you are a person with priorities, feelings, etc., and I'm really sorry for not being more up front about this. This says nothing about my feelings about udocker (I'm a huge fan of container technologies, and I champion your work) but rather if I think it's appropriate for a publication. I also doubt myself in this entire process that maybe my standards are too high and I should just give the a-ok, but I've had unpublished software rejected for (what I deem to be) much sillier reasons, so I don't think my intuition is that unreasonable. Part of this practice / exercise is convincing the reviewers that there is some paramount change that warrants a second publication, and I was not convinced.
