### 2021-08-09 #3295 @mariojmdavid
@vsoch thanks for your directness and I totally understand, of course I as one of the authors and contributors would like to see this published, but of course it's up to the editors and reviewers to evaluate that, and I will accept whatever decision is made, and I say sincerely on my part that even if it is not accepted for publication, it was a productive exercise on improvements,
I would like to say that I am expressing my point of view and that does not reflect necessarily the same views from the other authors
