### 2021-08-09 #3295 @danielskatz
@vsoch - I don't want to make this a major point, but I've generally tried to keep all conversations as public in the issue, so that there are no questions about who said what. I wish we had done this in this case, rather than our discussion happening in another channel.

While I accept that you are saying what you believe happened, this was not my perception or intent at the time, which was much more that there were questions and I was providing an opinion by talking about JOSS guidelines, though in hindsight, I recognize that my JOSS position gave my opinion more weight than I intended.  I took your acceptance to mean you were ok with the decision that I thought we collectively agreed on, which I now realize was not correct.

I apologize for not being more aware of all your and @mviereck's opinions, and for doing something that you felt pressured you to act in a way you didn't want to act, and am glad you brought this up today. I will be much more careful about things like this in the future.
