### 2021-08-09 #3295 @danielskatz
The JOSS AEiCs have now looked at this review thread and the software, and I'm going to let @arfon as the overall JOSS AEiC take over to state the decision as a person not directly involved in the process.
