### 2021-08-09 #3295 @reidpr
As someone (possibly) invited to review, I'm not familiar with this journal nor its review process.

But, seeing R1 (@mviereck) removed from the paper immediately after their reject recommendation, and R2 (@vsoch) feeling her concerns aren't being heard, makes it clear to me I do not want to be involved. Thanks for thinking of me and good luck.
