### 2021-08-09 #3295 @danielskatz
@reidpr - this has been a very unusual review, so please don't take this as typical of the journal or most submissions or reviews - see https://github.com/openjournals/joss-reviews/issues/3457 as one that was much more straightforward and typical, with a happy author, and https://github.com/openjournals/joss-reviews/issues/3311 as another, from @vsoch.  Unfortunately, sometimes (and I think rarely) things don't go as well, but we can learn from these cases and do better.
