### 2021-08-10 #3295 @arfon
This has been a challenging review for many reasons and I would like to thank @vsoch and @mviereck for their efforts and time invested here and the authors for their work to improve their submission based on the feedback of the reviewers.

@jorge-lip @vsoch @mviereck – as @danielskatz mentioned, the broader EiC team at JOSS has reviewed the status of this submission and arrived at a decision about how to proceed.

After reviewing the feedback from @mviereck & @vsoch and conducting our own internal review, we agree with their assessment that the work done between the [earlier paper](https://www.sciencedirect.com/science/article/pii/S0010465518302042) and this submission is insufficient to warrant a second publication. As such, this submission doesn't meet JOSS' [substantial scholarly effort](https://joss.readthedocs.io/en/latest/submitting.html#substantial-scholarly-effort) criterion.
